### 2024-02-07 #626 @simpar1471
gigs: Assess Fetal, Newborn, and Child Growth with International Standards
Submitting Author Name: Simon Parker
Submitting Author Github Handle: @simpar1471
Repository:  https://github.com/lshtm-gigs/gigs/
Version submitted: 0.4.1
Submission type: Stats
Badge grade: gold
Version accepted: TBD
---

```
Package: gigs
Title: Assess Fetal, Newborn, and Child Growth with International Standards
Version: 0.4.1
    person(given = "Simon", family = "Parker",
           email = "simon.parker@lshtm.ac.uk",
           role = c("aut", "cre"), comment = c(ORCID = "0009-0003-8214-4496")),
    person(given = "Linda", family = "Vesel", role = c("dtc"),
    person(given = "Eric", family = "Ohuma", role = c("dtc"),
    person(given = "Bill & Melinda Gates Foundation", role = c("fnd", "cph"))
    )
Description: Convert between anthropometric measures and z-scores/centiles in
    multiple growth standards, and classify fetal, newborn, and child growth
    accordingly. With a simple interface to growth standards from the World
    Health Organisation and International Fetal and Newborn Growth Consortium
    for the 21st Century, gigs makes growth assessment easy and reproducible for
    clinicians, researchers and policy-makers.
    https://lshtm-gigs.github.io/gigs/
    gamlss.dist,
    checkmate,
    vctrs
    units,
    withr
Config/testthat/parallel: true
Config/testthat/start-first: who_gs, ig_nbs, ig_fet, ig_png
Language: en-GB
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [X] A pre-submission inquiry has been approved in [issue#616](#616)

The target audience for this package is researchers, clinicians, and policy-makers interested in nutrition and fetal/newborn/child health. The package itself is used to convert measurements, for example a baby's weight at a given age, into summary statistics such as z-scores (i.e. number of standard deviations away from mean) and percentiles (i.e. number of percentage points above/below the median, expressed as a decimal). It can also be used to classify these z-scores and percentiles into specific categories for specific facets of growth. Scientific applications of this package would therefore include using our functions to generate statistical measures of growth in individuals or populations across time or pre/post a healthcare intervention. These could then be used in further downstream analyses.

Several R packages which implement child growth charts already exist, but they differ in the range of growth charts offered, their flexibility in conversion, and in what data they make available to users.
1. [`anthro`](https://github.com/worldhealthorganization/anthro) converts measurements into z-scores in the WHO Child Growth Standards, but lacks any INTERGROWTH-21<sup>st</sup> standards and outputs tabular data. We provide an interface which takes vectors in and gives vectors out, so is more flexible (e.g. with `dplyr` pipelines). This package is available on CRAN.
2. [`childsds`](https://github.com/mvogel78/childsds) can convert measurements to z-scores or percentiles, but cannot convert z-scores/percentiles to expected measurements. Additionally, [`childsds`](https://github.com/mvogel78/childsds) does not contain the newborn/postnatal INTERGROWTH-21<sup>st</sup> growth standards, which we implement. Though [`childsds`](https://github.com/mvogel78/childsds) does include more growth references, growth references are not within the scope of the GIGS project. We can discuss the reasons for this if necessary. This package is available on CRAN.
3. [`growthstandards`](https://github.com/ki-tools/growthstandards/) contains functions for converting between values and z-scores/percentiles, as in `gigs`. It includes the INTERGROWTH-21<sup>st</sup> fetal standards, but not the newborn or postnatal standards we implement. This package makes coefficients available to end-users, but not reference growth curves. This package is not available on CRAN. It was last updated in January 2024.
4. [`intergrowth`](https://github.com/nutriverse/intergrowth/) provides more fetal growth standards than `gigs`, but cannot convert between z-scores/percentiles in the lacks the INTERGROWTH-21<sup>st</sup> newborn or postnatal growth standards, which we have implemented in `gigs`. This package also does not make coefficients for the growth standards available to end-users, though it it does provide growth curve data to end-users. This package is not available on CRAN, and was last updated in January 2023.
5. `gigs` provides a simple interface for working with the growth standards it implements, which can be easily included in `dplyr`-like data wrangling pipelines. It implements all available WHO 2006/2007 and INTERGROWTH-21<sup>st</sup> growth standards, and makes published growth curves and model coefficients are available wherever possible. In our own testing, we found that it also warns end users about bad input data more comprehensively.

The [Other packages](https://lshtm-gigs.github.io/gigs/index.html#other-packages) section of the repository README also contains links to each of these packages, and a table describing features present in gigs compared to other growth standards packages.

We have a [benchmarking vignette](https://lshtm-gigs.github.io/gigs/articles/benchmarking.html) which shows the scaling of these  packages when analysing datasets containing 1 to 100,000 cases. In summary: `anthro` scales the worst (29 ms to 2.24 seconds). The other packages are then clustered together: `growthstandards` is the fastest for 100,000 cases (4.62 ms to 125 ms); `childsds` is next (2.40 ms to 126 ms); then `gigs` (1.32 ms to 128 ms).

Silver

* Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories. I feel this package has a comprehensive test suite and good error/warning behaviours in response to undesired input.
