### 2024-02-19 5dd733e
fix: Backwards-compatible formatting change in gigs_option_* function msgs. Also removed pipes in `checkmate` checks for gigs_option fns.
