### 2024-02-19 6fc45fd
docs: GIGS package-level options documentation improvement
