+++ R/gigs_options.R
@@ -1,40 +1,57 @@
-#' Get and set gigs package options
+#' Get and set gigs package-level options
 #'
-#' @param option A single character variable, one of:
-#'   * `"handle_missing"` - How should gigs handle missing (`NA`) data?
-#'   * `"handle_undefined"` - How should gigs handle undefined (`NaN`, `Inf`,
-#'       `-Inf`) data?
-#'   * `"handle_invalid_xvar"` - How should gigs handle `x` variables that
-#'     are out of bounds for the standard in use?
-#'   * `"handle_invalid_sex"` - How should gigs handle elements of `sex` which
-#'     are not one of `"M"` or `"F"`?
-#'   * `"handle_invalid_acronym"` - In conversion functions, how should gigs
-#'     handle `acronym` elements which do not map onto any implemented growth
-#'     standard?
-#'   * `"handle_invalid_centiles"` - In `centile2value` functions, how should
-#'     gigs handle elements of `p` that are not between `0` and `1`?
 #' @description By default, gigs will handle invalid input data, e.g. missing
 #'   (`NA`) or undefined (`NaN`, `Inf` and `-Inf`) values, by warning you of
 #'   their presence and replacing the invalid values with `NA`. You can change
 #'   this behaviour with `gigs_option_set()` so that gigs does this silently, or
 #'   make gigs throw errors when confronted with bad data.
-#' @returns For [gigs_option_get()], noisiliy returns the current value of
-#'   `option`. For [gigs_option_set()], invisibly returns the new value of
-#'   `option`.
+#' @param option A single case-sensitive character variable, one of:
+#'   * `"handle_missing_data"` - How should gigs handle missing (`NA`) elements?
+#'   * `"handle_undefined_data"` - How should gigs handle undefined (`NaN`, `Inf`,
+#'       `-Inf`) elements?
+#'   * `"handle_oob_xvar"` - How should gigs handle elements of `x`
+#'       variables which are out of bounds for the standard in use?
+#'   * `"handle_invalid_sex"` - How should gigs handle elements of `sex` which
+#'     are not one of `"M"` or `"F"`?
+#'   * `"handle_invalid_acronym"` - In conversion functions, how should gigs
+#'     handle `acronym` elements which do not map onto any implemented growth
+#'     standard?
+#'   * `"handle_oob_centiles"` - In `centile2value` functions, how should
+#'     gigs handle elements of `p` that are not between `0` and `1`?
+#' @param silent A single logical value controlling whether a message() will be
+#'   printed to the console describing either:
+#'   * The current value of `option` in `gigs_option_get()`.
+#'   * The `new_value` of `option` in `gigs_option_set()`.
+#'
+#'   This argument will not accept an integer value in place of a logical value.
+#' @param new_value A single case-sensitive character variable, one of:
+#'   * `"quiet"` - gigs will replace invalid vector elements to `NA` quietly.
+#'   * `"warn"` - gigs will replace invalid vector elements with `NA` loudly.
+#'   * `"error"` - gigs will throw informative errors if invalid inputs are
+#'     encountered.
+#' @returns Invisibly returns either a single character variable with the
+#'   current or new value of `option`, depending on which of [gigs_option_get()]
+#'   or [gigs_option_set()] is called. If `silent = FALSE` (default), then an
+#'   informative message will be printed to the console regarding the
+#'   current/new value of `option`.
 #' @examples
-#' \dontrun{
-#'   library(gigs)
+#' # Show the names of all available options
+#' names(.gigs_options)
 #'
-#'   # Get the names of all available options
-#'   names(.gigs_options)
[…]
