+++ _pkgdown.yml
@@ -62,9 +62,10 @@ reference:
       - who_gs
       - who_gs_coeffs
       - life6mo
-  - title: GIGS package options
+  - title: Package-level options
     desc: >
-      Options for how gigs should handle missing or undefined data, and some
-      out-of bounds values.
+      Options to control how gigs should handle missing or undefined data, and 
+      some out-of-bounds values.
     contents:
+      - .gigs_options
       - gigs_options
