### 2024-02-19 dfe2401
refactor: changed order of arguments in gigs_zscoring.R
