+++ R/classification.R
@@ -116,12 +116,11 @@ classify_svn <- function(weight_kg, gest_days, sex) {
 #' Classify stunting using WHO or INTERGROWTH-21<sup>st</sup>
 #' length/height-for-age standards
 #'
-#' Classify stunting (low length/height-for-age) using WHO or
-#' INTERGROWTH-21<sup>st</sup> length/height-for-age standards depending on the
-#' gestational age at birth of the child. Severe stunting is below <-3 SD
-#' relative to mean length/height at a given age, whereas moderate stunting is
-#' -2SD from the mean.
-#'
+#' @description Classify stunting (low length/height-for-age) using WHO or
+#'   INTERGROWTH-21<sup>st</sup> length/height-for-age standards depending on
+#'   the gestational age at birth of the child. Severe stunting is below <-3 SD
+#'   relative to mean length/height at a given age, whereas moderate stunting is
+#'   -2 SD from the mean.
 #' @inheritParams classify_sfga
 #' @param lenht_cm Numeric vector of length one or more with length/height
 #'   measurement(s) in cm.
@@ -134,15 +133,24 @@ classify_svn <- function(weight_kg, gest_days, sex) {
 #'   implausible z-score value thresholds should be applied. Default = `FALSE`.
 #' @returns An object of class factor with the same length as the longest input
 #'   vector, containing stunting classifications. Its levels are
-#'   `c("stunting_severe", "stunting", "not_stunting")` if `outliers = FALSE`
-#'   (the default), else `c("stunting_severe", "stunting", "not_stunting",
-#'   "outlier")`.
+#'   `c("stunting_severe", "stunting", "not_stunting")` if `outliers =
+#'   FALSE` (the default), else `c("stunting_severe", "stunting",
+#'   "not_stunting", "outlier")`.
+#'
+#'   \tabular{lll}{
+#'     \strong{Category} \tab \strong{Factor level} \tab
+#'       \strong{Z-score bounds} \cr
+#'     Severe stunting \tab `"stunting_severe"`     \tab `z` ≤ -3         \cr
+#'     Stunting        \tab `"stunting"`            \tab -3 < `z` ≤ -2    \cr
+#'     No stunting     \tab `"not_stunting"`        \tab `z` > -2         \cr
+#'     Outlier         \tab `"outlier"`             \tab `abs(z)` > 5
+#'   }
 #' @note Input vectors are recycled by [vctrs::vec_recycle_common()], and must
 #'   adhere to the [vctrs] recycling rules. This function assumes that your
 #'   measurements were taken according to the WHO guidelines, which stipulate
-#'   that recumbent length should not be measured after 730 days. Implausible
-#'   z-score bounds are sourced from the referenced WHO report, and
-#'   classification cut-offs from the DHS manual.
+#'   that recumbent length should not be measured after 730 days. Instead,
+#'   standing height should be used. Implausible z-score bounds are sourced from
+#'   the referenced WHO report, and classification cut-offs from the DHS manual.
 #' @references
 #' **'Implausible z-score values'** *in* World Health Organization (ed.)
 #' *Recommendations for data collection, analysis and reporting on
@@ -173,7 +181,10 @@ classify_svn <- function(weight_kg, gest_days, sex) {
 #'   outliers = TRUE
 #' )
 #' @export
-classify_stunting <- function(lenht_cm, age_days, gest_days, sex,
+classify_stunting <- function(lenht_cm,
+                              age_days,
+                              gest_days,
+                              sex,
[…]
