### 2024-02-19 ace720d
docs: Very minor improvements to gigs.Rmd
