### 2024-02-19 70652ff
chore: Very minor improvement to LICENSE.md
