### 2024-02-19 1417bc0
fix: roxygen docs for .gigs_options
