### 2024-02-19 dad5a6b
docs: Improvements to who_gs_value2zscore roxygen docs
