### 2024-02-19 18c4b4b
fix: Added .gigs_options to NAMESPACE exports
