+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(.gigs_options)
 export(classify_sfga)
 export(classify_stunting)
 export(classify_svn)
