### 2024-03-18 19a9798
feat: Standard-agnostic conversion functions now only take a single-string for `acronym`. Also speed improvements (to be doc'd later)
