+++ R/ig_nbs.R
@@ -13,11 +13,11 @@
 #'   * Between 154 and 280 days for `"fmfga"`, `"bfpfga"`, and `"ffmfga"`.
 #'
 #'   By default, gigs will replace elements in `gest_days` that are out of
-#'   bounds for the growth standard in use with `NA` and warn you. This
-#'   behaviour can be customised using the functions in [gigs_options].
-#' @param acronym Character vector of length one or more denoting the
-#'   INTERGROWTH-21<sup>st</sup> Newborn Size standard to use for each
-#'   observation. Each element should be one of:
+#'   bounds for the growth standard in use with `NA` and warn you. You can
+#'   customise this behaviour using the [GIGS package-level
+#'   options][gigs_options].
+#' @param acronym A single string denoting the INTERGROWTH-21<sup>st</sup>
+#'   Newborn Size standard to use. Should be one of:
 #'   * `"wfga"` (weight-for-GA)
 #'   * `"lfga"` (length-for-GA)
 #'   * `"hcfga"` (head circumference-for-GA)
@@ -26,11 +26,8 @@
 #'   * `"bfpfga"` (body fat %-for-GA)
 #'   * `"ffmfga"` (fat-free mass-for-GA)
 #'
-#'   This argument is case-sensitive. By default, gigs will replace elements in
-#'   `acronym` which are not one of the above values with `NA` and warn you.
-#'   This behaviour can be customised using the functions in [gigs_options]. If
-#'   all elements in `acronym` are not one of the above values, gigs will throw
-#'   an error.
+#'   This argument is case-sensitive. If `acronym` is not one of the values
+#'   listed above, the function will throw an error.
 #' @srrstats {G2.3b} Explicit reference to `acronym` case-sensitivity.
 #' @inherit shared_roxygen_params params note
 #' @inherit shared_zscore2value_returns return
@@ -480,20 +477,20 @@ ig_nbs_ffmfga_value2zscore <- function(fatfree_mass_g, gest_days, sex) {
 #' @note This function will fail if given inputs of different lengths.
 #' @noRd
 ig_nbs_c2v_internal <- function(p, gest_days, sex, acronym) {
-  vpns_lim <- 231
-  ifelse(
-    test = acronym == "wlrfga",
-    yes = ig_nbs_wlrfga_p2v(p, gest_days, sex),
-    no = ifelse(
-      test = acronym %in% c("fmfga", "bfpfga", "ffmfga"),
-      yes =  ig_nbs_bodycomp_p2v(p, gest_days, sex, acronym),
-      no = ifelse(
-        test = gest_days >= vpns_lim,
-        yes = ig_nbs_msnt_p2v(p, gest_days, sex, acronym),
-        no = ig_vpns_zscore2value(z = qnorm(p), gest_days, sex, acronym)
-      )
+  if (acronym == "wlrfga") {
+    y <- ig_nbs_wlrfga_p2v(p, gest_days, sex)
+  } else if (acronym == "fmfga" || acronym == "bfpfga" || acronym == "ffmfga") {
+    y <- ig_nbs_bodycomp_p2v(p, gest_days, sex, acronym)
+  } else {
+    vpns_lim <- 231
+    # y <- rep_len(x = NA_real_, length = length(p))
+    y <- ifelse(
+      test = gest_days >= vpns_lim,
+      yes = ig_nbs_msnt_p2v(p, gest_days, sex, acronym),
+      no = ig_vpns_zscore2value(z = qnorm(p), gest_days, sex, acronym)
[…]
