### 2024-03-18 722850c
refactor: Moved GIGS package-level doxs to gigs-package.R from data.R
