+++ R/data.R
@@ -1,20 +1,3 @@
-#' gigs: Assess Fetal, Newborn, and Child Growth with International Standards
-#'
-#' Convert between anthropometric measures and z-scores/centiles using WHO
-#' and INTERGROWTH-21<sup>st</sup> growth standards. This includes functions for
-#' classification of newborn and infant growth, including size for gestational
-#' age, stunting, wasting, and weight-for-age.
-#'
-#' @name gigs-package
-#' @aliases gigs
-#' @examples
-#' \dontrun{
-#' help(package = "gigs")
-#' browseURL("https://www.github.com/lshtm-gigs/gigs")
-#' }
-#' @keywords internal
-"_PACKAGE"
-
 #' INTERGROWTH-21<sup>st</sup> Newborn Size Standards (including very preterm)
 #' growth curve data
 #'
