+++ man/gigs-package.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
+% Please edit documentation in R/gigs-package.R
 \docType{package}
 \name{gigs-package}
 \alias{gigs-package}
