### 2024-03-18 6b7d033
docs: Improved readability in some docs re GIGS package-level options
