### 2024-03-27 eb53cc8
feat: Added `gigs_options_set()` for changing all gigs options at once
