+++ R/gigs_options.R
@@ -1,10 +1,11 @@
-#' Get and set gigs package-level options
+#' Functions to get and set gigs package-level options
 #'
 #' @description By default, gigs will handle invalid input data, e.g. missing
 #'   (`NA`) or undefined (`NaN`, `Inf` and `-Inf`) values, by warning you of
 #'   their presence and replacing the invalid values with `NA`. You can change
 #'   this behaviour with `gigs_option_set()` so that gigs does this silently, or
-#'   make gigs throw errors when confronted with bad data.
+#'   make gigs throw errors when confronted with bad data. You can also change
+#'   all GIGS options at once with `gigs_options_set()`.
 #' @param option A single case-sensitive character variable, one of:
 #'   * `"handle_missing_data"` - How should gigs handle missing (`NA`) elements?
 #'   * `"handle_undefined_data"` - How should gigs handle undefined (`NaN`,
@@ -15,27 +16,27 @@
 #'     are not one of `"M"` or `"F"`?
 #'   * `"handle_oob_centiles"` - In `centile2value` functions, how should
 #'     gigs handle elements of `p` that are not between `0` and `1`?
-#' @param silent A single logical value controlling whether a message() will be
-#'   printed to the console describing either:
+#' @param silent A single logical value controlling whether a `message()` will
+#'   be printed to the console describing either:
 #'   * The current value of `option` in `gigs_option_get()`.
 #'   * The `new_value` of `option` in `gigs_option_set()`.
 #'
-#'   This argument will not accept an integer value in place of a logical value.
+#'   This argument will not accept an integer in place of a logical value.
 #' @param new_value A single case-sensitive character variable, one of:
-#'   * `"quiet"` - gigs will replace invalid vector elements to `NA` quietly.
+#'   * `"quiet"` - gigs will replace invalid vector elements with `NA` quietly.
 #'   * `"warn"` - gigs will replace invalid vector elements with `NA` loudly.
-#'   * `"error"` - gigs will throw informative errors if invalid inputs are
+#'   * `"error"` - gigs will throw informative errors if any invalid inputs are
 #'     encountered.
-#' @returns Invisibly returns either a single character variable with the
-#'   current or new value of `option`, depending on which of [gigs_option_get()]
-#'   or [gigs_option_set()] is called. If `silent = FALSE` (default), then an
-#'   informative message will be printed to the console regarding the
-#'   current/new value of `option`.
+#' @returns A single-length character vector. For `gigs_option_get()`, the
+#'   current value of `option`; for `gigs_option_get()`, the new value of
+#'   `option`; for `gigs_options_set()`, `new_value`. If `silent = FALSE`
+#'   (default), then informative messages will be printed to the console
+#'   regarding the current/new values of `option`.
 #' @examples
 #' # Show the names of all available options
 #' names(.gigs_options)
 #'
-#' # Retrieve the value of a gigs option (with `silent =  FALSE`, would print a
+#' # Retrieve the value of a gigs option (if `silent = FALSE`, would print a
 #' # message)
 #' option_value <- gigs_option_get("handle_missing_data", silent = TRUE)
 #' print(option_value)
@@ -48,7 +49,10 @@
 #' print(option_value)
 #'
 #' # Suppress printed output with `silent = TRUE`
-#' gigs_option_set("handle_undefined_data", "warn", silent = TRUE)
+#' gigs_option_set("handle_undefined_data", "quiet", silent = TRUE)
+#'
+#' # Set all GIGS options
+#' gigs_options_set("warn", silent = TRUE)
 #' @name gigs_options
 #' @export
[…]
