+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(.gigs_options)
 export(classify_sfga)
 export(classify_stunting)
 export(classify_svn)
@@ -7,6 +8,7 @@ export(classify_wasting)
 export(classify_wfa)
 export(gigs_option_get)
 export(gigs_option_set)
+export(gigs_options_set)
 export(ig_fet_acfga_centile2value)
 export(ig_fet_acfga_value2centile)
 export(ig_fet_acfga_value2zscore)
