### 2024-03-27 a826faf
feat: Moved to nicer rlang-based `abort()`/`warn()`/`inform()` structure for input checking
