+++ R/check_params.R
@@ -23,9 +23,11 @@ handle_var <- function(vec, varname, option, test_lgl, msg_fn) {
   }
   out <- replace(vec, lgl_is_invalid, values = NA)
   str_msg <- msg_fn(lgl_is_invalid, varname = varname)
-  switch(gigs_option_get(option, silent = TRUE),
-         warn = warning(str_msg, call. = FALSE),
-         error = stop(str_msg, call. = FALSE))
+  option_value <- gigs_option_get(option, silent = TRUE)
+  warn_stop_fn <- switch(option_value,
+                         warn = rlang::warn,
+                         error = rlang::abort)
+  warn_stop_fn(str_msg, call = NULL, class = sprintf("gigs_%s", option))
   out
 }
 
@@ -103,7 +105,7 @@ handle_oob_xvar <- function(vec, varname, is_oob) {
 #' @rdname msg_oob_centiles
 #' @noRd
 msg_oob_centiles <- function(lgl_oob_centiles, varname) {
-  paste0("Variable '", varname, "': ", sum(lgl_oob_centiles), " in ",
+  paste0("Argument `", varname, "`: ", sum(lgl_oob_centiles), " in ",
          length(lgl_oob_centiles), " elements were not between 0 and 1.")
 }
 
@@ -113,9 +115,9 @@ msg_oob_centiles <- function(lgl_oob_centiles, varname) {
 #' @noRd
 msg_missing_data <- function(lgl_missing_data, varname) {
   if (all(lgl_missing_data) & varname == "acronym") {
-    stop("Variable '", varname, "': All elements were missing (`NA`).")
+    stop("Argument `", varname, "`: All elements were missing (`NA`).")
   }
-  paste0("Variable '", varname, "': ", sum(lgl_missing_data), " in ",
+  paste0("Argument `", varname, "`: ", sum(lgl_missing_data), " in ",
          length(lgl_missing_data), " elements were missing (`NA`).")
 }
 
@@ -124,7 +126,7 @@ msg_missing_data <- function(lgl_missing_data, varname) {
 #'   whether each element of a vector was either `NaN`, `-Inf`, or `Inf`.
 #' @noRd
 msg_undefined_data <- function(lgl_undefined_data, varname) {
-  paste0("Variable '", varname, "': ", sum(lgl_undefined_data),
+  paste0("Argument `", varname, "`: ", sum(lgl_undefined_data),
          " in ", length(lgl_undefined_data), " elements were undefined (`NaN`,",
          " `Inf`, or `-Inf`).")
 }
@@ -136,7 +138,7 @@ msg_undefined_data <- function(lgl_undefined_data, varname) {
 #' @noRd
 msg_oob_xvar <- function(lgl_oob_xvar, varname) {
   standard <- get(x = "standard", envir = parent.frame(n = 3))
-  paste0("Variable '", varname, "': ", sum(lgl_oob_xvar), " in ",
+  paste0("Argument `", varname, "`: ", sum(lgl_oob_xvar), " in ",
          length(lgl_oob_xvar), " elements were out-of-bounds (see the '",
          standard, "' conversion functions documentation).")
 }
@@ -155,16 +157,22 @@ msg_invalid_sex_acronym <- function(lgl_invalid_sex_acronym, varname) {
   chr_x_in_y <- if (all_invalid) "All" else paste0(count, " in ", len)
 
   if (varname == "sex") {
-    paste0("Variable 'sex': ", chr_x_in_y, " elements were neither ",
-           "\"M\" nor \"F\".")
+    c("i" = paste0("Argument `sex`: ", chr_x_in_y, " elements were neither ",
+                   "\"M\" nor \"F\"."))
   } else {
-    # This get() use smells --> refactor later if possible
[…]
