### 2024-03-27 d106b56
feat: Moved to nicer rlang-based `abort()`/`warn()`/`inform()` structure for input checking (added `rlang` to Imports)
