+++ DESCRIPTION
@@ -1,7 +1,6 @@
 Type: Package
 Package: gigs
-Title: Assess Fetal, Newborn, and Child Growth with International
-    Standards
+Title: Assess Fetal, Newborn, and Child Growth with International Standards
 Version: 0.4.1.9000
 Authors@R: c(
     person("Simon", "Parker", , "simon.parker@lshtm.ac.uk", role = c("aut", "cre"),
@@ -28,6 +27,7 @@ Imports:
     checkmate,
     gamlss.dist,
     stats,
+    rlang,
     vctrs
 Suggests:
     autotest,
