### 2024-03-27 f2837d0
feat: Altered structure of classification functions to have `classify_*()`/`compute_*()`/`categorise_*()`
 docs: Update README.Rmd to reflect changes to package

 BREAKING CHANGE: `classify_*()` functions are data-masking now
