+++ R/classification.R
@@ -1,322 +1,0 @@
-#' Classify size for gestational age using the INTERGROWTH-21<sup>st</sup>
-#' weight-for-gestational age standard
-#'
-#' Size for gestational age categories are split by centile: small-for-GA
-#' (SGA; 10<sup>th</sup> centile), appropriate-for-GA (AGA; 10<sup>th</sup>
-#' to 90<sup>th</sup> centile) and large-for-GA (LGA; >90<sup>th</sup>
-#' centile). This function also supports classification of severe SGA
-#' (<3<sup>rd</sup> centile) using the `severe` parameter.
-#'
-#' @inheritParams shared_roxygen_params
-#' @param weight_kg Numeric vector of length one or more with weight
-#'   value(s) in kg.
-#' @param gest_days Numeric vector of length one or more with gestational
-#'   age(s) at birth in days between `168` and `300`. By default, gigs will
-#'   replace out-of-bounds elements in `gest_days` with `NA` and warn you. This
-#'   behaviour can be customised using the functions in [gigs_options].
-#' @param severe A single logical value specifying whether to categorise SGA
-#'   values are below the third centile as `"SGA(<3)"`. Default = `FALSE`.
-#' @returns An object of class factor with the same length as the longest input
-#'   vector, containing size-for-GA classifications. If `severe = FALSE`, levels
-#'   are `c("SGA", "AGA", "LGA")`. If `severe = TRUE`, levels are `c("SGA(<3)",
-#'   "SGA", "AGA", "LGA")`.
-#' @note Input vectors are recycled by [vctrs::vec_recycle_common()], and must
-#'   adhere to the [vctrs] recycling rules.
-#' @seealso [ig_nbs_wfga_value2centile()], which this function calls to get
-#'   centiles for each observation.
-#' @examples
-#' # By default, does not differentiate between p < 0.03 and p < 0.10
-#' classify_sfga(
-#'   weight_kg = c(2.2, 3.3, 4.2),
-#'   gest_days = 267:269,
-#'   sex = c("M", "F", "M")
-#' )
-#'
-#' # With severe = TRUE, highlights p < 0.03
-#' classify_sfga(
-#'   weight_kg = c(2.2, 3.3, 4.2),
-#'   gest_days = 267:269,
-#'   sex = c("M", "F", "M"),
-#'   severe = TRUE
-#' )
-#' @references
-#' WHO. **Physical status: the use and interpretation of anthropometry. Report
-#' of a WHO Expert Committee.** *World Health Organisation Technical Report
-#' Series 1995,* **854: 1–452**
-#'
-#' Royal College of Obstetricians and Gynaecologists. **The Investigation and
-#' Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No.
-#' 31.** *Technical report, Royal College of Obstetricians and Gynaecologists,
-#' London, 2013.*
-#' @export
-classify_sfga <- function(weight_kg, gest_days, sex, severe = FALSE) {
-  checkmate::qassert(severe, rules = "B1")
-  # ig_nbs_wfga_value2centile will do checks on weight_kg/gest_days/sex
-  centiles <- ig_nbs_wfga_value2centile(weight_kg = weight_kg,
-                                        gest_days = gest_days,
-                                        sex = sex)
-  sfga <- rep(NA_character_, length(centiles))
-  sfga[centiles < 0.1] <- "SGA"
-  sfga[centiles >= 0.1 & centiles <= 0.9] <- "AGA"
-  sfga[centiles > 0.9] <- "LGA"
-  sfga_lvls <- c("SGA", "AGA", "LGA")
-  if (severe) {
-    sfga[centiles < 0.03] <- "SGA(<3)"
-    sfga_lvls <- c("SGA(<3)", sfga_lvls)
[…]
