+++ R/growth_classify.R
@@ -1,0 +1,870 @@
+# GROWTH CLASSIFICATION --------------------------------------------------------
+#   Take a `data.frame` or similar tabular object, assigns new columns with
+#   growth categories, then returns the extended table to the user
+
+#' Classify size-for-gestational age using the INTERGROWTH-21<sup>st</sup>
+#' weight-for-gestational age standard
+#'
+#' @param .data A `data.frame`-like tabular object with one or more rows. Must
+#'   be permissible by [checkmate::assert_data_frame()], so you can also supply
+#'   a `tibble`, `data.table`, or similar.
+#' @param weight_kg <[`data-masking`][rlang::args_data_masking]> The name of a
+#'   column in `.data` which is a numeric vector of weight values in kg.
+#' @param gest_days <[`data-masking`][rlang::args_data_masking]> The name of a
+#'   column in `.data` which is a numeric vector of gestational age(s) at birth
+#'   in days between `168` and `300`. By default, gigs will warn you about
+#'   elements of this vector which are outside these bounds, are `NA`, or `Inf`.
+#'   You can customise this behaviour using the [GIGS package-level
+#'   options][gigs_options].
+#' @param sex <[`data-masking`][rlang::args_data_masking]> The name of a
+#'   column in `.data` which is a case-sensitive character vector of sexes,
+#'   either `"M"` (male) or `"F"` (female). By default, gigs will warn you if
+#'   any elements of `sex` are not `"M"` or `"F"`, or are missing (`NA`). You
+#'   can customise this behaviour using the [GIGS package-level
+#'   options][gigs_options].
+#' @param .new A three-length character vector with names for the output
+#'   columns. These inputs will be repaired if necessary using
+#'   [vctrs::vec_as_names()], which will print any changes to the console. If
+#'   any elements in `.new` are the same as elements in `colnames(.data)`, the
+#'   function will throw an error. Default = `c("birthweight_centile", "sfga",
+#'   "sfga_severe")`.
+#' @returns A tabular object of the same class that was provided as `.data`,
+#'   with three new columns named according to `.new`. These columns will be
+#'   (from left to right):
+#'
+#'   * `birthweight_centile` - Numeric vector of birthweight centiles from the
+#'     INTERGROWTH-21<sup>st</sup> Newborn Size standard for
+#'     weight-for-GA
+#'   * `sfga` - Factor of size-for-GA categories without severe small-for-GA
+#'     classification
+#'   * `sfga_severe` - Factor of size-for-GA categories with severe small-for-GA
+#'     classification
+#' @note This function assumes that all weights were collected within 12 hours
+#'   of birth.
+#' @seealso [ig_nbs_wfga_value2centile()], which this function uses to get
+#'   a centile for each observation. Check out the [categorise_sfga()]
+#'   documentation to see the size-for-GA categorisation thresholds.
+#' @examples
+#' data <- data.frame(
+#'  wtkg = c(2.2, 2.5, 3.3, 4.0),
+#'  gestage = 266:269,
+#'  sex = c("F", "M", "F", "M")
+#' )
+#' data |>
+#'   classify_sfga(weight_kg = wtkg,
+#'                 gest_days = gestage,
+#'                 sex = sex)
+#' @inherit categorise_sfga references
+#' @export
+classify_sfga <- function(
+    .data,
+    weight_kg,
+    gest_days,
+    sex,
+    .new = c("birthweight_centile", "sfga", "sfga_severe")) {
+  checkmate::assert_data_frame(.data, min.rows = 1)
[…]
