### 2024-03-27 93d641f
docs: Minor update to `ig_png` conversion fn documentation
