### 2024-03-27 e3be6d2
refactor: Changed `gigs_zscoring` structure to utilise 'public' and 'internal' functions, like the conversion functions
