+++ R/gigs_zscoring.R
@@ -1,8 +1,8 @@
 #' Calculate z-scores for anthropometric measures according to GIGS guidance
 #'
-#' @description These functions calculate z-scores for either weight-for-age
-#'   (WAZs), length-for-age (LAZs), head circumference-for-age (HCAZs), or
-#'   weight-for-length (WLZs).
+#' @description These functions calculate z-scores for weight-for-age (WAZs),
+#'   length/height-for-age (LHAZs), head circumference-for-age (HCAZs), or
+#'   weight-for-length/height (WLZs).
 #'
 #'   The z-scoring procedure used differs for each observation based on the
 #'   gestational and post-menstrual age of the infants being analysed. The
@@ -24,34 +24,81 @@
 #'   * `gest_days` ≥ 43 weeks:
 #'     - Birth: No standard available
 #'     - Postnatal: Uncorrected WHO Child Growth standards
+#'
+#'   For `gigs_wlz()`, two WHO standards are available:
+#'   * The weight-for-length standard, which is applied when `age_days < 731`.
+#'   * The weight-for-height standard, which is applied when `age_days >= 731`.
+#' @param weight_kg A numeric vector of length one or more with weight in kg.
+#' @param age_days A numeric vector of length one or more with ages in days.
+#' @param gest_days A numeric vector of length one or more with weight in kg.
+#' @param sex A character vector of sexes, either `"M"` or `"F"`.
 #' @note These functions expect vectors which have lengths such that they can
 #'   be recycled with [vctrs::vec_recycle_common()].
 #' @rdname gigs_waz
 #' @noRd
-gigs_waz <- function(weight_kg, gest_days, age_days, sex) {
-  validated <- validate_waz_params(weight_kg = weight_kg,
-                                   age_days = age_days,
-                                   gest_days = gest_days,
-                                   sex = sex)
-  weight_kg <- validated[[1]]
-  age_days <- validated[[2]]
-  gest_days <- validated[[3]]
-  sex <- validated[[4]]
+gigs_waz <- function(weight_kg, age_days, gest_days, sex) {
+  validate_waz_params(weight_kg = weight_kg,
+                      age_days = age_days,
+                      gest_days = gest_days,
+                      sex = sex) |>
+    do.call(what = gigs_waz_internal)
+}
 
+#' @rdname gigs_waz
+#' @param lenht_cm A numeric vector of length one or more with length or height
+#'   in cm. Should be recumbent length when `age_days < 731`, and standing
+#'   height when `age_days >= 731`.
+#' @noRd
+gigs_lhaz <- function(lenht_cm, age_days, gest_days, sex) {
+  validate_lhaz_params(lenht_cm = lenht_cm,
+                       age_days = age_days,
+                       gest_days = gest_days,
+                       sex = sex) |>
+    do.call(what = gigs_lhaz_internal)
+}
+
+#' @rdname gigs_waz
+#' @noRd
+gigs_wlz <- function(weight_kg, lenht_cm, age_days, gest_days, sex) {
+  validate_wlz_params(weight_kg = weight_kg,
+                      lenht_cm = lenht_cm,
+                      age_days = age_days,
[…]
