### 2024-03-27 9014f51
refactor: No-effect change to `gigs-package.R`
