+++ R/gigs-package.R
@@ -14,7 +14,3 @@
 #' }
 #' @keywords internal
 "_PACKAGE"
-
-## usethis namespace: start
-## usethis namespace: end
-NULL
