### 2024-03-27 237beac
test: Added test file for `growth_categorisation.R`
