### 2024-03-27 270cae4
test: Commented out error/warning testing in test suite due to earlier commmits. Needs rewriting.
