+++ tests/testthat/test-ig_fet.R
@@ -88,8 +88,8 @@ test_that(
       for (chr_z_or_p in c("zscore", "centile")) {
         xvar <- ig_fet[[acronym]][[1]][[1]]
         for (seed in seq(300, 400, 30)) {
-          dbl_z_or_p <- withr::with_seed(seed, code = {
-            rnorm(n = length(xvar))
+          withr::with_seed(seed, code = {
+            dbl_z_or_p <- rnorm(n = length(xvar))
           })
           if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
@@ -115,780 +115,727 @@ test_that(
     for (acronym in names(gigs::ig_fet)) {
       for (chr_z_or_p in c("zscore", "centile")) {
         xvar <- ig_fet[[acronym]][[1]][[1]]
-        withr::with_seed(
-          seed = 50,
-          code = {
-            dbl_z_or_p <- rnorm(n = length(xvar))
-            if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
+        xrange <- range(xvar)
+        withr::with_seed(seed = 50, code = {
+          xvar <- jitter(xvar, 1)
+          xvar[xvar < xrange[1]] <- xrange[1]
+          xvar[xvar > xrange[2]] <- xrange[2]
+          dbl_z_or_p <- rnorm(n = length(xvar))
+        })
+        if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
-            fn_stem <- paste0("ig_fet_", acronym)
-            fn_zp2val <- get(paste0(fn_stem, "_", chr_z_or_p, "2value"))
-            y_gigs <- fn_zp2val(dbl_z_or_p, xvar)
+        fn_stem <- paste0("ig_fet_", acronym)
+        fn_zp2val <- get(paste0(fn_stem, "_", chr_z_or_p, "2value"))
+        y_gigs <- fn_zp2val(dbl_z_or_p, xvar)
 
-            fn_val2zp <- get(paste0(fn_stem, "_value2", chr_z_or_p))
-            gigs_z_or_p <- fn_val2zp(y_gigs + .Machine$double.eps, xvar)
+        fn_val2zp <- get(paste0(fn_stem, "_value2", chr_z_or_p))
+        gigs_z_or_p <- fn_val2zp(y_gigs + .Machine$double.eps, xvar)
 
-            expect_equal(gigs_z_or_p, expected = dbl_z_or_p,
-                         tolerance = sqrt(.Machine$double.eps))
-          }
-        )
+        expect_equal(gigs_z_or_p, expected = dbl_z_or_p,
+                     tolerance = sqrt(.Machine$double.eps))
       }
     }
   }
 )
 
 # Appropriate errors for zero-length data + data of wrong type -----------------
-
-#' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-#' @srrstats {G5.8, G5.8a, G5.8b} Including errors for zero-length inputs and
-#'   incorrect data types.
-test_that(
-  desc = "Correct errors occur with zero-length/incorrectly typed data:",
-  code = {
-    # Error message testing: Iterate through ig_fet functions, demonstrating
-    #   with passed `expect_error()` tests that all error messages are unique
-    #   to the data passed in.
-    ig_fet_acronyms <- names(ig_fet)
-    acronyms_to_test <- withr::with_seed(150, sample(ig_fet_acronyms, 5))
-    fn_suffixes <- c("value2zscore", "value2centile", "centile2value",
-                     "zscore2value")
[…]
