+++ tests/testthat/test-ig_png.R
@@ -51,11 +51,9 @@ test_that(desc = "Conversion of values to z-scores works", {
     for (chr_z_or_p in c("zscore", "centile")) {
       xvar <- gigs::ig_png[[acronym]][[1]][[1]][[1]]
       for (seed in seq(450, 550, 30)) {
-        dbl_z_or_p <- withr::with_seed(seed = seed, {
-          rnorm(n = length(xvar))
-        })
-        sexvar <- withr::with_seed(seed = seed, {
-           sample(c("M", "F"), size = length(xvar), replace = TRUE)
+        withr::with_seed(seed, {
+          dbl_z_or_p <- rnorm(n = length(xvar))
+          sexvar <- sample(c("M", "F"), size = length(xvar), replace = TRUE)
         })
         if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
@@ -79,11 +77,13 @@ test_that(
     for (acronym in names(gigs::ig_png)) {
       for (chr_z_or_p in c("zscore", "centile")) {
         xvar <- ig_png[[acronym]][[1]][[1]][[1]]
-        dbl_z_or_p <- withr::with_seed(seed = 800, {
-          rnorm(n = length(xvar))
-        })
-        sexvar <- withr::with_seed(seed = 800, {
-           sample(c("M", "F"), size = length(xvar), replace = TRUE)
+        xrange <- range(xvar)
+        withr::with_seed(50, code = {
+          xvar <- jitter(xvar, 1)
+          xvar[xvar < xrange[1]] <- xrange[1]
+          xvar[xvar > xrange[2]] <- xrange[2]
+          dbl_z_or_p <- rnorm(n = length(xvar))
+          sexvar <- sample(c("M", "F"), size = length(xvar), replace = TRUE)
         })
         if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
@@ -103,915 +103,915 @@ test_that(
 
 # Appropriate errors for zero-length data + data of wrong type -----------------
 
-#' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-#' @srrstats {G5.8, G5.8a, G5.8b} Including errors for zero-length inputs and
-#'   incorrect data types.
-test_that(
-  desc = "Correct errors occur when .gigs_options are set to 'quiet':",
-  code = {
-    # Error message testing: Iterate through ig_png functions, demonstrating
-    #   with passed `expect_error()` tests that all error messages are unique
-    #   to the data passed in.
-    acronyms_to_test <- names(ig_png)
-    fn_suffixes <- c("value2zscore", "value2centile", "centile2value",
-                     "zscore2value")
-    z <- 0
-    p <- 0.5
-    y <- 4
-    for (fn_suffix in fn_suffixes) {
-      for (str_acronym in acronyms_to_test) {
-        fn_name <- paste("ig_png", str_acronym, fn_suffix, sep = "_")
-        gigs_fn <- get(fn_name)
-        arg_x_name <- gigs::ig_png[[str_acronym]][["x"]]
-        y_name <- gigs::ig_png[[str_acronym]][["y"]]
-
-        arg_zyp <- switch(fn_suffix, centile2value = p, zscore2value = z, y)
-        arg_zyp_name <- switch(fn_suffix, centile2value = "p",
-                               zscore2value = "z", y_name)
-        arg_x <- gigs::ig_png[[str_acronym]][[1]][[1]][[1]]
-        arg_sex <- withr::with_seed(25, {
-          sample(c("M", "F"), replace = TRUE, size = length(arg_x))
-        })
-
[…]

[…]
