### 2024-03-27 cdffdd4
tests: Dropped test-gigs_zscoring.R as gigs_hcaz() will be covered by `test-growth_(compute/classify).R`. Added test files for `growth_compute.R` and `growth_classify.R`
