+++ tests/testthat/test-growth_compute.R
@@ -1,0 +1,3 @@
+test_that("test growth_compute", {
+  expect_equal(2 * 2, 4)
+})
