+++ tests/testthat/test-growth_classify.R
@@ -1,0 +1,3 @@
+test_that("test growth_classify", {
+  expect_equal(2 * 2, 4)
+})
