### 2024-03-27 e4c4ebb
docs: pkgdown.yml changes due to earlier commits
