### 2024-04-03 8bef645
fix: handle_var() now works when GIGS options are 'quiet'
