### 2024-04-03 036300a
fix: classify_growth now outputs wasting classifications, and gets the right size-for-GA classes in the right columns
