### 2024-04-03 c42698a
fix: Updated benchmarks to include all sets of standards and use `bench::press()`
