+++ vignettes/articles/benchmarking.Rmd
@@ -12,6 +12,7 @@ knitr::opts_chunk$set(
 
 # Set ggplot theming for consistent plots
 ggplot2::theme_set(ggplot2::theme_bw())
+intergrowth21st <- knitr::asis_output(x = "INTERGROWTH-21<sup>st</sup>")
 ```
 
 ```{r srr-tags, eval = FALSE, echo = FALSE}
@@ -40,32 +41,92 @@ standards.
 Thus far, there is no comprehensive benchmark comparing these different
 packages. This short article will compare the speed of each package from 1 to
 100000 inputs, checking how fast each package can convert weight values to
-z-scores in the WHO Child Growth standards.
+z-scores in different growth standards.
 
-We have performed these benchmarks on a Windows 10 system running a Ryzen 7
-3700X processor and 16GB of DDR4 RAM, using R version 4.3.2. The Stata
-benchmarks have been run on the same system in Stata 18.0 (revision 15 May
-2023), using the [benchmark](https://github.com/mcaceresb/stata-benchmark) package for Stata.
+We performed these benchmarks on R version
+`r session_info[[1]][["major"]]`.`r session_info[[1]][["minor"]]`, using a
+Windows 10 system with a Ryzen 7 3700X processor and 16GB of DDR4 RAM. The Stata
+benchmarks were run in Stata 18.0 (revision 15 May 2023) on the same system,
+using the [benchmark](https://github.com/mcaceresb/stata-benchmark) package for Stata.
 
-## Set up benchmark dataset
-The benchmarking and package comparisons will all use the same 100,000-row
-dataset comprised of data which can be used to convert between z-scores
-and centiles in the WHO weight-for-age standard. The z-scores, `x` variable
-and sexes for each row are generated randomly with a pre-specified seed. As gigs
-is validated against the published growth curve data from the WHO, we use it to
-generate weight values in kg for each observation. This dataset can then be used
-to compare speed and accuracy in conversion between values and z-scores/centiles
-for the packages mentioned above.
+Other R packages can be used to analyse growth data with international
+standards, but have limitations which are not present in gigs. There are also
+software packages external to R which implement these standards. The table
+below describes these packages, and to what extent they have implemented
+functions that let users convert anthropometric measurements to
+z-scores/centiles in each set of standards implemented in gigs - the WHO Child
+Growth standards, `r intergrowth21st` Newborn Size standards (including Very
+Preterm), and the `r intergrowth21st` Postnatal Growth standards for preterm
+infants. A green tick (✅) indicates that a package contains functions for all
+growth standards within a set, a red cross (❌) indicates that a package has no
+functions for a set of growth standards, and an exclamation mark (❕) indicates
+that some standards in a set are implemented, but not others.
 
-```{r who_gs_dataset, echo = TRUE}
-who_gs_dataset <- function(n, acronym = "wfa", seed = 154237890) {
-  withr::with_seed(
-    seed,
-    code = {
+```{r arrows, echo = FALSE}
+ig21st <- knitr::asis_output(x = "IG-21<sup>st</sup>")
+
+larr <- knitr::asis_output(x = "&larr;")
+rarr <- knitr::asis_output(x = "&rarr;")
+harr <- knitr::asis_output(x = "&harr;")
+yes <- knitr::asis_output(x = "✅")
[…]
