### 2024-04-03 9287850
fix: `quiet` for GIGS options *really* does work now
