+++ R/check_params.R
@@ -22,15 +22,21 @@ handle_var <- function(vec, varname, option, test_lgl, msg_fn) {
     return(vec)
   }
   out <- replace(vec, lgl_is_invalid, values = NA)
-  if (option != "quiet") {
-    str_msg <- msg_fn(lgl_is_invalid, varname = varname)
-    option_value <- gigs_option_get(option, silent = TRUE)
-    warn_stop_fn <- switch(option_value,
-                           warn = rlang::warn,
-                           error = rlang::abort)
-    warn_stop_fn(str_msg, call = NULL, class = sprintf("gigs_%s", option))
-  }
+  str_msg <- msg_fn(lgl_is_invalid, varname = varname)
+  option_value <- gigs_option_get(option, silent = TRUE)
+  warn_stop_fn <- switch(option_value,
+                         warn = rlang::warn,
+                         error = rlang::abort,
+                         quiet = return(out))
+  warn_stop_fn(str_msg, call = NULL, class = sprintf("gigs_%s", option))
   out
+}
+
+test <- function (option, out = rnorm(5), msg = "nipeuwfjr") {
+  warn_stop_fn <- switch(option,
+                           warn = rlang::warn,
+                           error = rlang::abort,
+                           quiet = return(out))
 }
 
 #' @rdname handle_var
