### 2024-04-03 634802f
docs: Updated gigs.Rmd with a bibliography and extended example using `life6mo`, which matches the example given in the SJ manuscript
