+++ vignettes/gigs_bib.bib
@@ -1,0 +1,28 @@
+@article{Vesel2022LowSettings,
+  title = {Low Birthweight Infant Feeding Practices and Growth Patterns in the First Six Months of Life in Resource-Limited Settings},
+  volume = {6},
+  ISSN = {2475-2991},
+  url = {http://dx.doi.org/10.1093/cdn/nzac060.069},
+  DOI = {10.1093/cdn/nzac060.069},
+  journal = {Current Developments in Nutrition},
+  publisher = {Elsevier BV},
+  author = {Vesel,  Linda and Kisenge,  Rodrick and Koujalagi,  Murugaraj and Kusagur,  Varun and Manji,  Karim and Mvalo,  Tisungane and Nayak,  Rashmita and Phiri,  Melda and Saidi,  Friday and Semrau,  Katherine and Sudfeld,  Christopher},
+  year = {2022},
+  month = jun,
+  pages = {611}
+}
+
+@article{Vesel2023FeedingStudy,
+  title = {Feeding practices and growth patterns of moderately low birthweight infants in resource-limited settings: results from a multisite,  longitudinal observational study},
+  volume = {13},
+  ISSN = {2044-6055},
+  url = {http://dx.doi.org/10.1136/bmjopen-2022-067316},
+  DOI = {10.1136/bmjopen-2022-067316},
+  number = {2},
+  journal = {BMJ Open},
+  publisher = {BMJ},
+  author = {Vesel,  Linda and Bellad,  Roopa M and Manji,  Karim and Saidi,  Friday and Velasquez,  Esther and Sudfeld,  Christopher R and Miller,  Katharine and Bakari,  Mohamed and Lugangira,  Kristina and Kisenge,  Rodrick and Salim,  Nahya and Somji,  Sarah and Hoffman,  Irving and Msimuko,  Kingsly and Mvalo,  Tisungane and Nyirenda,  Fadire and Phiri,  Melda and Das,  Leena and Dhaded,  Sangappa and Goudar,  Shivaprasad S and Herekar,  Veena and Kumar,  Yogesh and Koujalagi,  M B and Guruprasad,  Gowdar and Panda,  Sanghamitra and Shamanur,  Latha G and Somannavar,  Manjunath and Vernekar,  Sunil S and Misra,  Sujata and Adair,  Linda and Bell,  Griffith and Caruso,  Bethany A and Duggan,  Christopher and Fleming,  Katelyn and Israel-Ballard,  Kiersten and Fishman,  Eliza and Lee,  Anne C C and Lipsitz,  Stuart and Mansen,  Kimberly L and Martin,  Stephanie L and Mokhtar,  Rana R and North,  Krysten and Pote,  Arthur and Spigel,  Lauren and Tuller,  Danielle E and Young,  Melissa and Semrau,  Katherine E A},
+  year = {2023},
+  month = feb,
+  pages = {e067316}
+}
