+++ vignettes/gigs.Rmd
@@ -3,6 +3,7 @@ title: "Introduction to gigs"
 description: >
   Learn how to get started with the basics of gigs.
 output: rmarkdown::html_vignette
+bibliography: gigs_bib.bib
 vignette: >
   %\VignetteIndexEntry{Introduction to gigs}
   %\VignetteEngine{knitr::rmarkdown}
@@ -13,6 +14,8 @@ vignette: >
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 ```
 
+# Introduction
+
 gigs is designed to make working with outputs from either the
 INTERGROWTH-21<sup>st</sup> project and World Health Organisation Child Growth
 Standards as easy as possible. It also provides functions for easy and
@@ -20,162 +23,196 @@ reproducible classification of fetal, newborn, and infant growth.
 
 Which functions you want to use depend on whether you're **converting between
 values and z-scores/centiles**, or **classifying observations**. Let's start
-by loading the package.
+by loading the package. We're also turning off the warnings which GIGS prints
+if inputs are `NA`, `NaN`, or similarly invalid, as to avoid cluttering the
+console output.
 
 ```{r setup}
 library(gigs)
+gigs_options_set(new_value = "quiet")
 ```
 
-## Converting between z-scores/centiles and values
+We will show off two ways of working with growth data in gigs. The first
+'step-by-step' approach involves you picking and choosing your growth standards
+throughout. The second 'single-step' approach lets you hand over that work to
+gigs, letting its growth classification functions do the work for you.
 
-### From values to z-scores/centiles
-Much of the time, you'll want to know where a measurement for a given
-observation lies relative to 'expected' values for that observation. This is
-where the `value2zscore` and `value2centile` functions are useful:
+Before we get started, we need some data. This tutorial uses a sample of data
+from 300 infants enrolled in the Low birthweight Infant Feeding Exploration
+(LIFE) study [@Vesel2022LowSettings; @Vesel2023FeedingStudy]. Thse infants were
+assessed at birth, and followed through after birth, with data available up to
+six months of age in this data extract.
 
-#### Example 1a: Converting newborn weights
-```{r example_1a}
-df_newborn_weights <- data.frame(
-        weight_kg = c(0.91, 3.33, 2.75, 3.35, 3.47, 1.90),
-        gest_days = c(183, 280, 293, 259, 277, 231),
-        sex = c(rep("M", 3), rep("F", 3))
+Though `gigs::life6mo` has data on length in cm (`len_cm`), head circumference
+(`headcirc_cm`), and mid-upper arm circumference (`muac_cm`), we are going to
+subset `gigs::life6mo` to focus on weight in grams (`weight_g`). The dataset
+also has columns with an ID number per infant in the study (`id`), the
+gestational age at birth for each infant (`gestage`), the sex of each infant
+(`sex`; `"M"` = male, `"F"` = female), the visit week at which weight was
+recorded (`visitweek`), and post-menstrual age in days (`pma`).
+
+After loading the dataset, we add a variable `preterm`, which is `TRUE` wherever
+an infant was born before 37 weeks' gestational age. We also convert `sex` from
+a `factor` to a `character` variable so that gigs will accept it as an input.
+```{r load_data}
+life6mo <- gigs::life6mo[, 1:7]
+life6mo$preterm <- life6mo$gestage < 37 * 7
+life6mo$sex <- as.character(life6mo$sex)
+head(life6mo, n = 5)
+```
+
+# Single-step (easier) approach
+
+# Step-by-step (harder) approach
+Instead of letting GIGS do the heavy lifting, you may want to clearly lay out
[…]
