### 2024-04-03 6900162
docs: Updated life6mo to have more cases with birth data; corresponding changes to gigs.Rmd
