### 2024-04-03 6c6674f
docs: Updated README to include timing data from benchmarking vignette
