### 2024-04-09 b0cad92
docs: Better GIGS shared example
