+++ vignettes/gigs_bib.bib
@@ -26,3 +26,128 @@
   month = feb,
   pages = {e067316}
 }
+
+@article{Villar2014InternationalProject,
+    title = {{International standards for newborn weight, length, and head circumference by gestational age and sex: The Newborn Cross-Sectional Study of the INTERGROWTH-21st Project}},
+    year = {2014},
+    journal = {The Lancet},
+    author = {Villar, José and Ismail, Leila Cheikh and Victora, Cesar G. and Ohuma, Eric O. and Bertino, Enrico and Altman, Doug G. and Lambert, Ann and Papageorghiou, Aris T. and Carvalho, Maria and Jaffer, Yasmin A. and Gravett, Michael G. and Purwar, Manorama and Frederick, Ihunnaya O. and Noble, Alison J. and Pang, Ruyan and Barros, Fernando C. and Chumlea, Cameron and Bhutta, Zulfiqar A. and Kennedy, Stephen H. and Katz, M. and Bhan, M. K. and Garza, C. and Zaidi, S. and Langer, A. and Rothwell, P. M. and Weatherall, D. and Burton, F. and Er, Y. A.Jaff and Lumbiganon, P. and Rivera, J. and Shorten, M. and Hoch, L. and Knight, H. E. and Cosgrove, C. and Blakey, I. and Roseman, F. and Kunnawar, N. and Gu, S. H. and Wang, J. H. and Wu, M. H. and Domingues, M. and Gilli, P. and Juodvirsiene, L. and Musee, N. and Al-Jabri, H. and Waller, S. and Muninzwa, D. and Yellappan, D. and Carter, A. and Reade, D. and Miller, R. and Salomon, L. and Leston, A. and Mitidieri, A. and Al-Aamri, F. and Paulsene, W. and Sande, J. and Al-Zadjali, W. K.S. and Batiuk, C. and Bornemeier, S. and Dighe, M. and Gaglioti, P. and Jacinta, N. and Jaiswal, S. and Oas, K. and Oberto, M. and Olearo, E. and Owende, M. G. and Shah, J. and Sohoni, S. and Todros, T. and Venkataraman, M. and Vinayak, S. and Wang, L. and Wilson, D. and Wu, Q. Q. and Zhang, Y. and Chamberlain, P. and Danelon, D. and Sarris, I. and Dhami, J. and Ioannou, C. and Knight, C. L. and Napolitano, R. and Wanyonyi, S. and Pace, C. and Mkrtychyan, V. and Al-Habsi, F. and Alija, M. and Jimenez-Bustos, J. M. and Kizidio, J. and Puglia, F. and Liu, H. and Lloyd, S. and Mota, D. and Ochieng, R. and Rossi, C. and Luna, M. Sanchez and Shen, Y. J. and Rocco, D. A. and Albernaz, E. and Batra, M. and Bhat, B. A. and Di Nicola, P. and Giuliani, F. and Rovelli, I. and McCormick, K. and Paul, V. and Rajan, V. and Wilkinson, A. and Varalda, A. and Eskenazi, B. and Corra, L. A. and Dolk, H. and Golding, J. and Matijasevich, A. and De Wet, T. and Zhang, J. J. and Bradman, A. and Finkton, D. and Burnham, O. and Farhi, F. and Fonseca, S. and Sclowitz, I. K. and Da Silveira, M. and He, Y. P. and Pan, Y. and Yuan, Y. and Choudhary, A. and Choudhary, S. and Deshmukh, S. and Dongaonkar, D. and Ketkar, M. and Khedikar, V. and Mahorkar, C. and Mulik, I. and Saboo, K. and Shembekar, C. and Singh, A. and Taori, V. and Tayade, K. and Somani, A. and Frigerio, M. and Gilli, G. and Giolito, M. and Occhi, L. 
[…]
