+++ vignettes/gigs.Rmd
@@ -4,6 +4,7 @@ description: >
   Learn how to get started with the basics of gigs.
 output: rmarkdown::html_vignette
 bibliography: gigs_bib.bib
+csl: nature.csl
 vignette: >
   %\VignetteIndexEntry{Introduction to gigs}
   %\VignetteEngine{knitr::rmarkdown}
@@ -12,12 +13,17 @@ vignette: >
 
 ```{r knitr_opts, include = FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
+intergrowth21st <- if (knitr::is_html_output()) {
+  "INTERGROWTH-21<sup>st</sup>"
+} else {
+  "INTERGROWTH-21\textsuperscript{st}"
+}
 ```
 
 # Introduction
 
 gigs is designed to make working with outputs from either the
-INTERGROWTH-21<sup>st</sup> project and World Health Organisation Child Growth
+`r intergrowth21st` project and World Health Organisation Child Growth
 Standards as easy as possible. It also provides functions for easy and
 reproducible classification of fetal, newborn, and infant growth.
 
@@ -32,20 +38,28 @@ library(gigs)
 gigs_options_set(new_value = "quiet")
 ```
 
-We will show off two ways of working with growth data in gigs. The first
-'step-by-step' approach involves you picking and choosing your growth standards
-throughout. The second 'single-step' approach lets you hand over that work to
-gigs, letting its growth classification functions do the work for you.
+We will show off two use cases for gigs. The first is using gigs to generate
+continuous indices of growth (i.e. centiles or z-scores). This can be done in a
+single step by letting gigs do the work in the background, or you can
 
+[//]: # (TODO: MAKE PURDY)
+
+involves simpler
+'single-step' approach involves passing the hard work to gigs, and letting it
+pick the right growth standards to apply for each measurement in your data. The
+more complex 'step-by-step' approach involves you picking and choosing your
+growth standards throughout.
+
+## Example data
 Before we get started, we need some data. This tutorial uses a sample of data
 from 300 infants enrolled in the Low birthweight Infant Feeding Exploration
 (LIFE) study [@Vesel2022LowSettings; @Vesel2023FeedingStudy]. Thse infants were
-assessed at birth, and followed through after birth, with data available up to
-six months of age in this data extract.
+assessed at birth then followed up at intervals after birth, with data available
+up to six months of age in this data extract.
 
 Though `gigs::life6mo` has data on length in cm (`len_cm`), head circumference
 (`headcirc_cm`), and mid-upper arm circumference (`muac_cm`), we are going to
-subset `gigs::life6mo` to focus on weight in grams (`weight_g`). The dataset
+subset this dataset to focus on weight in grams (`weight_g`). The dataset
 also has columns with an ID number per infant in the study (`id`), the
 gestational age at birth for each infant (`gestage`), the sex of each infant
 (`sex`; `"M"` = male, `"F"` = female), the visit week at which weight was
@@ -61,59 +75,53 @@ life6mo$sex <- as.character(life6mo$sex)
 head(life6mo, n = 5)
 ```
 
-# Single-step (easier) approach
+# Obtaining z-scores from measurements
 
-# Step-by-step (harder) approach
-Instead of letting GIGS do the heavy lifting, you may want to clearly lay out
-the steps taken in your analysis. Luckily, gigs provides all the functionality
-you need for this.
+## Single-step (easier) approach
+The easiest way to use gigs is to let it handle work for you. Using its growth
[…]
