### 2024-04-09 2e00e81
docs: Fixed issues in growth categorisation/computation/classification functions
