+++ man/categorise_wfa.Rd
@@ -17,6 +17,12 @@ implausible z-score thresholds should be applied. Default = \code{FALSE}.}
 An object of class factor with the same length as \code{waz}, containing
 weight-for-age classifications. Its levels are \code{c("underweight_severe", "underweight", "normal", "overweight")} if \code{outliers = FALSE} (the
 default), else \code{c("underweight_severe", "underweight", "normal", "overweight", "outlier")}.
+}
+\description{
+Classify weight-for-age z-scores into weight-for-age strata
+}
+\details{
+Cut-offs for weight-for-age categories are:
 
 \tabular{lll}{
 \strong{Category}    \tab \strong{Factor level}  \tab
@@ -27,9 +33,6 @@ Normal weight        \tab \code{"normal_weight"}      \tab \code{abs(waz)} < 2  
 Overweight           \tab \code{"overweight"}         \tab \code{waz} ≥ 2       \cr
 Outlier              \tab \code{"outlier"}        \tab \code{waz} < -6 or \code{waz} > 5
 }
-}
-\description{
-Classify weight-for-age z-scores into weight-for-age strata
 }
 \examples{
 waz <- c(-6.5, -3.5, -2.5, 0, 2.5, 3.5)
