### 2024-04-10 b2683a8
refactor: Added internal error to `gigs_zscoring_lgls()` and made logical statements more exclusive
