+++ R/gigs_zscoring.R
@@ -1,27 +1,30 @@
 #' Calculate z-scores for anthropometric measures according to GIGS guidance
 #'
 #' @description These functions calculate z-scores for weight-for-age (WAZs),
-#'   length/height-for-age (LHAZs), head circumference-for-age (HCAZs), or
-#'   weight-for-length/height (WLZs).
+#'   length/height-for-age (LHAZs), weight-for-length/height (WLZs), or
+#'   head circumference-for-age (HCAZs).
 #'
-#'   The z-scoring procedure used differs for each observation based on the
-#'   gestational and post-menstrual age of the infants being analysed. The
-#'   different procedures are based on the advised use of each growth standard:
-#'   * `gest_days` < 24 weeks:
+#'   The growth standard used for each observation differs based on the
+#'   gestational and post-menstrual age of the infant being analysed. The
+#'   different procedures are based on the advised use of each growth standard,
+#'   where a measurement is considered a 'birth measurement' if taken <1 day
+#'   after birth:
+#'   * `gest_days` < 24 weeks (168 days):
 #'     - Birth: No standard available
 #'     - Postnatal: IG-21<sup>st</sup> Postnatal Growth standards from 27 to 64
 #'       weeks' PMA
 #'     - Postnatal: Uncorrected WHO standard after 64 weeks' PMA
-#'   * `gest_days` >= 24 weeks:
-#'     - Birth: IG-21<sup>st</sup> Newborn Size standards at birth (including
-#'       Very Preterm Newborn Size standards if born before 33 weeks).
+#'   * `gest_days` >= 24 weeks (168 days):
+#'     - Birth: IG-21<sup>st</sup> Newborn Size standards at birth
+#'       (including Very Preterm Newborn Size standards if born before 33
+#'       weeks).
 #'     - Postnatal: IG-21<sup>st</sup> Postnatal Growth standards from 27 to 64
 #'       weeks' PMA
 #'     - Postnatal: Uncorrected WHO Child Growth standards after 64 weeks' PMA
-#'   * `gest_days` ≥ 37 weeks:
+#'   * `gest_days` ≥ 37 weeks (259 days):
 #'     - Birth: IG-21<sup>st</sup> Newborn Size standards
 #'     - Postnatal: Uncorrected WHO Child Growth standards
-#'   * `gest_days` ≥ 43 weeks:
+#'   * `gest_days` ≥ 43 weeks (301 days):
 #'     - Birth: No standard available
 #'     - Postnatal: Uncorrected WHO Child Growth standards
 #'
@@ -30,11 +33,12 @@
 #'   * The weight-for-height standard, which is applied when `age_days >= 731`.
 #' @param weight_kg A numeric vector of length one or more with weight in kg.
 #' @param age_days A numeric vector of length one or more with ages in days.
-#' @param gest_days A numeric vector of length one or more with weight in kg.
-#' @param sex A character vector of sexes, either `"M"` or `"F"`.
+#' @param gest_days A numeric vector of length one or more with gestational ages
+#'   in days.
+#' @inheritParams shared_roxygen_params
 #' @note These functions expect vectors which have lengths such that they can
 #'   be recycled with [vctrs::vec_recycle_common()].
-#' @rdname gigs_waz
+#' @rdname gigs_zscoring
 #' @noRd
 gigs_waz <- function(weight_kg, age_days, gest_days, sex) {
   validate_waz_params(weight_kg = weight_kg,
@@ -44,7 +48,7 @@ gigs_waz <- function(weight_kg, age_days, gest_days, sex) {
     do.call(what = gigs_waz_internal)
 }
 
[…]
