### 2024-04-10 3f89c62
docs: Nicer formatting in to _pkgdown.yml and improved NEWS.md
