### 2024-04-10 44900f8
docs: Minor change to growth_compute.R
