### 2024-04-16 e050699
docs: Minor changes to docs (life6mo update)
