+++ data-raw/r2sas.R
@@ -1,0 +1,151 @@
+# PREPARE DATA FOR GIGS (SAS) PACKAGE ------------------------------------------
+
+library(gigs)
+
+#' Write a data frame to SAS-readable .txt format with `write.table()`
+#' @param x A data frame to write to disk
+#' @param file Name for saved file.
+#' @param sep Separator for .txt file. Default = `","`.
+write.table_sas <- function(x, file, sep = ",") {
+  x[is.na(x)] <- " "
+  write.table(x,
+              file = file,
+              sep = sep,
+              row.names = FALSE,
+              col.names = FALSE,
+              quote = FALSE)
+  x
+}
+
+ig_fet_zscores <- list()
+for (standard in names(gigs::ig_fet)) {
+  tbl_zscores <- gigs::ig_fet[[standard]][["zscores"]]
+  if (is.null(tbl_zscores)) {
+    next
+  } else {
+    ig_fet_zscores[[standard]] <- gigs::ig_fet[[standard]][["zscores"]] |>
+      dplyr::mutate(acronym = standard,
+                    x_unit = gigs::ig_fet[[standard]][["x"]],
+                    y_unit = gigs::ig_fet[[standard]][["y"]])
+  }
+}
+ig_fet_zscores <- ig_fet_zscores |>
+  dplyr::bind_rows() |>
+  dplyr::relocate(acronym, x_unit, y_unit, .before = tidyselect::everything()) |>
+  write.table_sas(file = "exclude/r2sas/ig_fet_zscores.txt", sep = " ")
+
+ig_fet_centiles <- list()
+for (standard in names(gigs::ig_fet)) {
+  tbl_zscores <- gigs::ig_fet[[standard]][["centiles"]]
+  if (is.null(tbl_zscores)) {
+    next
+  } else {
+    ig_fet_centiles[[standard]] <- gigs::ig_fet[[standard]][["centiles"]] |>
+      dplyr::mutate(acronym = standard,
+                    x_unit = gigs::ig_fet[[standard]][["x"]],
+                    y_unit = gigs::ig_fet[[standard]][["y"]])
+    names(ig_fet_centiles[[standard]])[1] <- "x"
+  }
+}
+ig_fet_centiles <- ig_fet_centiles |>
+  dplyr::bind_rows() |>
+  dplyr::relocate(P25, .before = P50) |>
+  dplyr::relocate(P75, .after = P50) |>
+  dplyr::relocate(acronym, x_unit, y_unit, .before = tidyselect::everything()) |>
+  write.table_sas(file = "exclude/r2sas/ig_fet_centiles.txt", sep = " ")
+
+
+ig_nbs_zscores <- list()
+for (standard in names(gigs::ig_nbs)) {
+  li_zscores <- list()
+  for (sex in c("male", "female")) {
+    li_zscores[[sex]] <- gigs::ig_nbs[[standard]][[sex]][["zscores"]]
+  }
+  df_zscores <- dplyr::bind_rows(li_zscores, .id = "sex") |>
+    dplyr::mutate(sex = ifelse(sex == "male", yes = "M", no = "F"))
+  ig_nbs_zscores[[standard]] <- df_zscores |>
+    dplyr::mutate(acronym = standard,
+                  y_unit = gigs::ig_nbs[[standard]][["y"]])
+}
+ig_nbs_zscores <- ig_nbs_zscores |>
+  dplyr::bind_rows() |>
+  dplyr::relocate(acronym, y_unit, sex, .before = tidyselect::everything()) |>
+  write.table_sas(file = "exclude/r2sas/ig_nbs_zscores.txt", sep = " ")
+
+ig_nbs_centiles <- list()
+for (standard in names(gigs::ig_nbs)) {
[…]
