### 2024-04-16 fca3d08
feat: Added `err_if_.new_in_.data()` checks to all `classify_*()` functionms
