+++ R/growth_classify.R
@@ -63,9 +63,7 @@ classify_sfga <- function(
     sex,
     .new = c("birthweight_centile", "sfga", "sfga_severe")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
-  err_if_.new_in_.data(new = .new,
-                       .data_names = colnames(.data),
-                       .data_varname = rlang::expr_text(substitute(.data)))
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
   checkmate::qassert(.new, rules = "S3")
 
   p <- ig_nbs_wfga_value2centile(
@@ -121,7 +119,7 @@ classify_svn <- function(.data,
                           .new = c("birthweight_centile", "svn")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S2")
-
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
   p <- ig_nbs_wfga_value2centile(
     weight_kg = eval_tidy(enquo(weight_kg), .data),
     gest_days = eval_tidy(enquo(gest_days), .data),
@@ -189,6 +187,7 @@ classify_stunting <- function(
     .new = c("lhaz", "stunting", "stunting_outliers")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S3")
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
 
   lhaz <- gigs_lhaz(
     lenht_cm = eval_tidy(enquo(lenht_cm), .data),
@@ -249,6 +248,7 @@ classify_wasting <- function(.data,
                              .new = c("wlz", "wasting", "wasting_outliers")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S3")
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
 
   wlz <- gigs_wlz(
     weight_kg = eval_tidy(enquo(weight_kg), .data),
@@ -306,6 +306,7 @@ classify_wfa <- function(.data,
                          .new = c("waz", "wfa", "wfa_outliers")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S3")
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
 
   waz <- gigs_waz(
     weight_kg = eval_tidy(enquo(weight_kg), data = .data),
@@ -367,6 +368,7 @@ classify_headsize <- function(.data,
                               .new = c("hcaz", "headsize")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S2")
+  err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
 
   hcaz <- gigs_hcaz(
     headcirc_cm = eval_tidy(enquo(headcirc_cm), .data),
@@ -535,8 +537,7 @@ classify_growth <- function(
 
   # Ascertain integrity of `.data`, `.analyses`, and `.new`
   err_if_.new_in_.data(new = unique(unlist(.new)),
-                       .data_names = colnames(.data),
-                       .data_varname = rlang::expr_text(substitute(.data)))
+                       .data_colnames = colnames(.data))
   check_all_.new_names_valid(.new = .new, all = all_analyses)
   check_all_.analyses_in_.new(.new = .new, .analyses = .analyses)
   .new <- .new[.analyses]
@@ -545,15 +546,17 @@ classify_growth <- function(
   check_if_.new_elements_unique(.new)
   .new <- repair_.new_names(.new)
 
[…]
