### 2024-05-01 da470d4
feat: Added INTERGROWTH-21st standard for Estimated Fetal Weight from Hadlock's formula
