+++ R/ig_fet.R
@@ -16,7 +16,9 @@
 #'     INTERGROWTH-21<sup>st</sup> Fetal Doppler standards).
 #'   * Between 105 and 252 days for `"poffga"`, `"sffga"`, `"avfga"`, and
 #'     `"pvfga"` (the INTERGROWTH-21<sup>st</sup> Fetal Brain Development
-#'     standards).
+#'     standards).#
+#'   * Between 126 and 287 days for `"hefwfga"` (the INTERGROWTH-21<sup>st</sup>
+#'     standard for Hadlock-based estimated fetal weight).
 #'
 #'   By default, gigs will replace out-of-bounds elements in `x` with `NA` and
 #'   warn you. You can customise this behaviour using the [GIGS package-level
@@ -42,6 +44,7 @@
 #'   * `"avfga"` (anterior horn of lateral ventricle-for-GA)
 #'   * `"pvfga"` (atrium of posterior horn of lateral ventricle-for-GA)
 #'   * `"cmfga"` (cisterna magna-for-GA)
+#'   * `"hefwfga"` (Hadlock esimated fetal weight-for-GA)
 #'
 #'   This argument is case-sensitive. If `acronym` is not one of the values
 #'   list above, gigs will throw an error.
@@ -61,6 +64,10 @@
 #' **International Estimated Fetal Weight Standards of the INTERGROWTH-21st
 #' Project.** *Ultrasound Obstet Gynecol* 2016, **49:478-486**
 #' \doi{10.1002/uog.17347}
+#'
+#' Stirnemann J, Salomon LJ, Papageorghiou AT. **INTERGROWTH-21st standards for
+#' Hadlock's estimation of fetal weight.** *Ultrasound Obstet Gynecol* 2020,
+#' **56(6):946-948** \doi{10.1002/uog.22000}
 #'
 #' Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA, Pang R et al.
 #' **International standards for symphysis-fundal height based on serial
@@ -357,6 +364,17 @@ ig_fet_cmfga_zscore2value <- function(z, gest_days) {
 }
 
 #' @rdname ig_fet_zscore2value
+#' @export
+ig_fet_hefwfga_zscore2value <- function(z, gest_days) {
+  acronym <- "hefwfga"
+  validate_ig_fet(z = z,
+                  x = gest_days,
+                  acronym = acronym,
+                  x_name = gigs::ig_fet[[acronym]][["x"]]) |>
+    do.call(what = ig_fet_z2v_internal)
+}
+
+#' @rdname ig_fet_zscore2value
 #' @importFrom stats qnorm
 #' @export
 ig_fet_centile2value <- function(p, x, acronym) {
@@ -577,6 +595,17 @@ ig_fet_pvfga_centile2value <- function(p, gest_days) {
 #' @export
 ig_fet_cmfga_centile2value <- function(p, gest_days) {
   acronym <- "cmfga"
+  validated <- validate_ig_fet(p = p,
+                               x = gest_days,
+                               acronym = acronym,
+                               x_name = gigs::ig_fet[[acronym]][["x"]])
+  with(validated, ig_fet_z2v_internal(qnorm(p), x, acronym))
+}
+
+#' @rdname ig_fet_zscore2value
+#' @export
+ig_fet_hefwfga_centile2value <- function(p, gest_days) {
+  acronym <- "hefwfga"
   validated <- validate_ig_fet(p = p,
                                x = gest_days,
[…]
