+++ NAMESPACE
@@ -70,6 +70,10 @@ export(ig_fet_hcfga_centile2value)
 export(ig_fet_hcfga_value2centile)
 export(ig_fet_hcfga_value2zscore)
 export(ig_fet_hcfga_zscore2value)
+export(ig_fet_hefwfga_centile2value)
+export(ig_fet_hefwfga_value2centile)
+export(ig_fet_hefwfga_value2zscore)
+export(ig_fet_hefwfga_zscore2value)
 export(ig_fet_ofdfga_centile2value)
 export(ig_fet_ofdfga_value2centile)
 export(ig_fet_ofdfga_value2zscore)

[…]
