### 2024-05-01 1e1fb1a
docs: Fixed bad units in `x` arg for WHO GS functions
