### 2024-05-01 52075b2
tests: Updated test suite for new "hefwfga" standard in ig_fet
