### 2024-05-02 d4dacfe
refactor: Changed name of `.analyses` arg to `.outcomes` in `classify_growth()`; fixed bug where non-claculated outcomes led to 'success' being printed to console
