+++ R/growth_classify.R
@@ -45,13 +45,13 @@
 #'   a centile for each observation.
 #' @examples
 #' data <- data.frame(
-#'  wtkg = c(2.2, 2.5, 3.3, 4.0),
+#'  wt_kg = c(2.2, 2.5, 3.3, 4.0),
 #'  gestage = 266:269,
 #'  sex = c("F", "M", "F", "M")
 #' )
 #'
 #' data |>
-#'   classify_sfga(weight_kg = wtkg,
+#'   classify_sfga(weight_kg = wt_kg,
 #'                 gest_days = gestage,
 #'                 sex = sex)
 #' @inherit categorise_sfga details references
@@ -99,13 +99,13 @@ classify_sfga <- function(
 #' @inherit categorise_svn details
 #' @examples
 #' data <- data.frame(
-#'   wtkg = c(1.5, 2.6, 2.6, 3.5),
+#'   wt_kg = c(1.5, 2.6, 2.6, 3.5),
 #'   gestage = c(235, 257, 275, 295),
 #'   sex = c("F", "M", "F", "M")
 #' )
 #'
 #' data |>
-#'   classify_svn(weight_kg = wtkg,
+#'   classify_svn(weight_kg = wt_kg,
 #'                gest_days = gestage,
 #'                sex = sex)
 #' @seealso [ig_nbs_wfga_value2centile()], which this function uses to get
@@ -226,14 +226,14 @@ classify_stunting <- function(
 #' @examples
 #' # Returns factor with stunting classifications
 #' data <- data.frame(
-#'   wtkg = c(5.75, 2.18, 3.00, 6.75),
+#'   wt_kg = c(5.75, 2.18, 3.00, 6.75),
 #'   length_height = c(67.7, 46.6, 50.0, 80.1),
 #'   gestage = c(251, 197, 225, 243),
 #'   age = c(251, 197, 225, 243),
 #'   sex =  c("F", "M", "F", "M")
 #' )
 #' data |>
-#'   classify_wasting(weight_kg = wtkg,
+#'   classify_wasting(weight_kg = wt_kg,
 #'                    lenht_cm = length_height,
 #'                    gest_days = gestage,
 #'                    age_days = age,
@@ -287,13 +287,13 @@ classify_wasting <- function(.data,
 #' @inherit categorise_stunting references
 #' @examples
 #' data <- data.frame(
-#'   wtkg = c(7.2, 4.5, 9.1, 24),
+#'   wt_kg = c(7.2, 4.5, 9.1, 24),
 #'   age = c(401, 185, 101, 607),
 #'   gestage = 7 * c(27, 36, 40, 41),
 #'   sex = c("F", "M", "F", "M")
 #' )
 #' data |>
-#'   classify_wfa(weight_kg = wtkg,
+#'   classify_wfa(weight_kg = wt_kg,
 #'                age_days = age,
 #'                gest_days = gestage,
 #'                sex = sex)
@@ -396,7 +396,7 @@ classify_headsize <- function(.data,
 #'   taken <12hrs after birth, i.e. `age_days < 0.5`.
 #' @inheritParams classify_wasting
 #' @inheritParams classify_headsize
-#' @param .analyses A character vector of up to six elements in length
+#' @param .outcomes A character vector of up to six elements in length
 #'   describing which growth analyses you want to run. Use this when you supply
 #'   data that can be used to generate multiple growth indicators, but only want
 #'   to run a specific set. This argument is case-sensitive, and an error will
@@ -431,40 +431,49 @@ classify_headsize <- function(.data,
 #' data <- data.frame(
 #'   agedays = c(0, 100, 100),
 #'   gestage = c(270, 270, 270),
-#'   wtkg = c(2.5, 7.5, 7.5),
[…]
