### 2024-05-02 9d5e745
docs: Changed wtkg to wt_kg for consistency
