### 2024-05-02 7471d2b
chore: Changed RoxygenNote to reflect update to `roxygen2` pkg
