+++ CITATION.cff
@@ -28,7 +28,7 @@ preferred-citation:
     email: simon.parker@lshtm.ac.uk
     orcid: https://orcid.org/0009-0003-8214-4496
   year: '2023'
-  notes: 0.4.1
+  notes: 0.4.1.9000
   url: https://github.com/lshtm-gigs/gigs/
 repository-code: https://github.com/lshtm-gigs/gigs
 url: https://github.com/lshtm-gigs/gigs/
