+++ DESCRIPTION
@@ -46,4 +46,4 @@ Language: en-GB
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
     "srr::srr_stats_roclet"))
-RoxygenNote: 7.3.0
+RoxygenNote: 7.3.1
