### 2024-05-02 5a27178
docs: Minor updates to ig_fet.R
