### 2024-05-02 95bead0
refactor: Messages in gigs_options.R now use `rlang::inform()`
