### 2024-05-02 a67c3be
refactor: Messages in gigs_options.R now use `rlang::inform()`
