### 2024-05-15 3a0c7c9
chore: Updated life6mo to have weight in kgs in `wt_kg` col
