### 2024-05-15 bd8f0ff
refactor: Changed `.analyses` arg in `gigs::classify_growth()` to `.outcomes`
