+++ man/life6mo.Rd
@@ -8,15 +8,15 @@
 A data frame with 2,191 rows and 10 variables:
 \itemize{
 \item{\code{id}:}{  Unique ID for each infant in the dataset (1--300).}
-\item{\code{gestage}:}{  Best obstetric estimate of gestational age in days
-(181--291).}
-\item{\code{sex}:}{  Sex of the infant as a factor (\code{"M"} = Male; \code{"F"} =
-Female).}
 \item{\code{visitweek}:}{  Chronological age in weeks (±1) when study visit
 occurred (0--26).}
-\item{\code{pma}:}{  Post-menstrual age in days (182--528).}
+\item{\code{sex}:}{  Sex of the infant as a factor (\code{"M"} = Male; \code{"F"} =
+Female).}
+\item{\code{gestage}:}{  Best obstetric estimate of gestational age in days
+(181--291).}
 \item{\code{age_days}:}{  Chronological age in days at each visit; equal to
 \code{pma - gestage} (0--242).}
+\item{\code{pma}:}{  Post-menstrual age in days (182--528).}
 \item{\code{wt_kg}:}{  Mean weight in kg (1.24--9.40667).}
 \item{\code{len_cm}:}{  Mean length in cm (37.37--72.93).}
 \item{\code{headcirc_cm}:}{  Mean head circumference in cm (23.20--44.87).}
