+++ R/data.R
@@ -51,12 +51,7 @@ NULL
 #' gestational ages for either sex, for the INTERGROWTH-21<sup>st</sup>
 #' weight/length/head circumference-for-gestational age standards. The lists are
 #' ordered by acronym, then sex.
-#' @references
-#' Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
-#' **International standards for newborn weight, length, and head circumference
-#' by gestational age and sex: the Newborn Cross-Sectional Study of the
-#' INTERGROWTH-21st Project.** *Lancet* 2014, **384(9946):857-68.**
-#' \doi{10.1016/S0140-6736(14)60932-6}
+#' @inherit ig_nbs_centile2value references
 #' @srrstats {G1.0} Primary literature referenced here.
 #' @source
 #' Mu/sigma/nu/tau values were provided by Dr Eric Ohuma.
@@ -74,11 +69,7 @@ NULL
 #' A set of nested lists containing tables with reference values at different
 #' z-scores/centiles for valid post-menstrual ages. The list is ordered by
 #' acronym first, then by sex, and finally by z-score/centile.
-#' @references
-#' Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
-#' **Postnatal growth standards for preterm infants: the Preterm Postnatal
-#' Follow-up Study of the INTERGROWTH-21st Project.** *Lancet Glob Health* 2015,
-#' *3(11):e681-e691.* \doi{10.1016/S2214-109X(15)00163-1}
+#' @inherit ig_png_zscore2value references
 #' @source
 #' [INTERGROWTH-21<sup>st</sup> Postnatal Growth of Preterm
 #'   Infants](https://intergrowth21.tghn.org/postnatal-growth-preterm-infants/)
@@ -148,51 +139,7 @@ NULL
 #'
 #' [INTERGROWTH-21<sup>st</sup> Gestational Weight Gain
 #'   standard](https://intergrowth21.tghn.org/gestational-weight-gain/)
-#' @references
-#' Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A
-#' et al. **International standards for fetal growth based on serial ultrasound
-#' measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st
-#' Project.** *Lancet* 2014, **384(9946):869-79.**
-#' \doi{10.1016/S0140-6736(14)61490-2}
-#'
-#' Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al.
-#' **International Estimated Fetal Weight Standards of the INTERGROWTH-21st
-#' Project.** *Ultrasound Obstet Gynecol* 2016, **49:478-486**
-#' \doi{10.1002/uog.17347}
-#'
-#' Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA, Pang R et al.
-#' **International standards for symphysis-fundal height based on serial
-#' measurements from the Fetal Growth Longitudinal Study of the INTERGROWTH-21st
-#' Project: prospective cohort study in eight countries.** *BMJ* 2016,
-#' **355:i5662** \doi{10.1136/bmj.i5662}
-#'
-#' Papageorghiou AT, Kennedy SH, Salomon LJ, Ohuma EO, Cheikh Ismail L, Barros
-#' FC et al. **International standards for early fetal size and pregnancy dating
-#' based on ultrasound measurement of crown-rump length in the first trimester
[…]
