### 2024-05-30 18d3833
docs: Updated planned development/scope nots in CONTRIBUTING.md
