### 2024-05-30 0dd2ac6
refactor: Changed internal function names for ig_png_equations() + ig_vpns_equations() to ig_png_mu_sigma() + ig_vpns_mu_sigma().
