+++ README.Rmd
@@ -105,13 +105,13 @@ life6mo_newborns <- gigs::life6mo[life6mo$age_days == 0, ]
 
 # Use classify_growth() to get multiple growth indicators at once
 life6mo_classified <- classify_growth(
-  .data = life6mo_newborns,
-  weight_kg = weight_g / 1000,
-  lenht_cm = len_cm,
-  gest_days = gestage,
-  age_days = age_days,
-  sex = as.character(sex),
-  .analyses = c("svn", "stunting")
+        .data = life6mo_newborns,
+        gest_days = gestage,
+        age_days = age_days,
+        sex = as.character(sex),
+        weight_kg = wt_kg,
+        lenht_cm = len_cm,
+        .outcomes = c("svn", "stunting")
 )
 
 head(life6mo_classified, n = 4)
@@ -122,12 +122,12 @@ wanted to run were successful. In the example below, because `lenht_cm` is not
 specified, stunting indicators cannot be computed.
 ```{r examples_classification3}
 life6mo_classified <- classify_growth(
-  .data = life6mo_newborns,
-  weight_kg = weight_g / 1000,
-  gest_days = gestage,
-  age_days = age_days,
-  sex = as.character(sex),
-  .analyses = c("svn", "stunting")
+        .data = life6mo_newborns,
+        gest_days = gestage,
+        age_days = age_days,
+        sex = as.character(sex),
+        weight_kg = wt_kg,
+        .outcomes = c("svn", "stunting")
 )
 
 head(life6mo_classified, n = 4)
@@ -139,7 +139,7 @@ indicator you'd like to calculate, for example `classify_svn()`:
 # Small vulnerable newborns
 life6mo_svn <- classify_svn(
   .data = life6mo_newborns,
-  weight_kg = weight_g / 1000,
+  weight_kg = wt_kg,
   gest_days = gestage,
   sex = as.character(sex)
 )
@@ -184,29 +184,29 @@ international growth standards from the WHO and `r intergrowth21st` project:
 * `ig_fet` - `r intergrowth21st` Fetal standards
   <details><summary>Component standards</summary>
 
-  | Acronym  | Description                                                  | Unit | `x` range       |
-  |----------|--------------------------------------------------------------|------|-----------------|
-  | `hcfga`  | head circumference-for-GA                                    | mm   | 98 to 280 days  |
-  | `bpdfga` | biparietal diameter-for-GA                                   | mm   | 98 to 280 days  |
-  | `acfga`  | abdominal circumference-for-GA                               | mm   | 98 to 280 days  |
-  | `flfga`  | femur length-for-GA                                          | mm   | 98 to 280 days  |
-  | `ofdfga` | occipito-frontal diameter for-GA                             | mm   | 98 to 280 days  |
-  | `efwfga` | estimated fetal weight-for-GA                                | g    | 154 to 280 days |
-  | `sfhfga` | symphisis-fundal height-for-GA                               | mm   | 112 to 294 days |
-  | `crlfga` | crown-rump length-for-GA                                     | mm   | 58 to 105 days  |
-  | `gafcrl` | GA-for-crown-rump length                                     | days | 15 to 95 mm     |
-  | `gwgfga` | gestational weight gain-for-GA                               | kg   | 98 to 280 days  |
-  | `pifga`  | pulsatility index-for-GA                                     |      | 168 to 280 days |
-  | `rifga`  | resistance index-for-GA                                      |      | 168 to 280 days |
-  | `sdrfga` | systolic/diastolic ratio-for-GA                              |      | 168 to 280 days |
-  | `tcdfga` | transcerebellar diameter-for-GA                              | mm   | 98 to 280 days  |
-  | `tcdfga` | GA-for-transcerebellar diameter                              | mm   | 98 to 280 days  |
[…]
