+++ README.md
@@ -93,24 +93,24 @@ life6mo_newborns <- gigs::life6mo[life6mo$age_days == 0, ]
 
 # Use classify_growth() to get multiple growth indicators at once
 life6mo_classified <- classify_growth(
-  .data = life6mo_newborns,
-  weight_kg = weight_g / 1000,
-  lenht_cm = len_cm,
-  gest_days = gestage,
-  age_days = age_days,
-  sex = as.character(sex),
-  .analyses = c("svn", "stunting")
+        .data = life6mo_newborns,
+        gest_days = gestage,
+        age_days = age_days,
+        sex = as.character(sex),
+        weight_kg = wt_kg,
+        lenht_cm = len_cm,
+        .outcomes = c("svn", "stunting")
 )
 #> `gigs::classify_growth()`
 #> ✔ Small vulnerable newborns: Success
 #> ✔ Stunting: Success
 
 head(life6mo_classified, n = 4)
-#>    id gestage sex visitweek pma age_days weight_g   len_cm headcirc_cm  muac_cm
-#> 1   1     273   M         0 273        0     2300 42.06667    33.26667 9.433333
-#> 28  4     250   F         0 250        0     1500 42.03333    30.03333 8.066667
-#> 36  5     238   F         0 238        0     2390 43.46667    33.63333 9.166667
-#> 56  8     240   F         0 240        0     1800 41.73333    31.46667 8.033334
+#>    id gestage sex visitweek pma age_days wt_kg   len_cm headcirc_cm  muac_cm
+#> 1   1     273   M         0 273        0  2.30 42.06667    33.26667 9.433333
+#> 28  4     250   F         0 250        0  1.50 42.03333    30.03333 8.066667
+#> 36  5     238   F         0 238        0  2.39 43.46667    33.63333 9.166667
+#> 56  8     240   F         0 240        0  1.80 41.73333    31.46667 8.033334
 #>    birthweight_centile         svn       lhaz        stunting stunting_outliers
 #> 1          0.010765424    Term SGA -3.5406446 stunting_severe   stunting_severe
 #> 28         0.002833163 Preterm SGA -2.2854751        stunting          stunting
@@ -125,23 +125,23 @@ computed.
 
 ``` r
 life6mo_classified <- classify_growth(
-  .data = life6mo_newborns,
-  weight_kg = weight_g / 1000,
-  gest_days = gestage,
-  age_days = age_days,
-  sex = as.character(sex),
-  .analyses = c("svn", "stunting")
+        .data = life6mo_newborns,
+        gest_days = gestage,
+        age_days = age_days,
+        sex = as.character(sex),
+        weight_kg = wt_kg,
+        .outcomes = c("svn", "stunting")
 )
 #> `gigs::classify_growth()`
 #> ✔ Small vulnerable newborns: Success
-#> ✔ Stunting: Success
+#> ! Stunting: Not computed (`lenht_cm` not supplied)
 
 head(life6mo_classified, n = 4)
-#>    id gestage sex visitweek pma age_days weight_g   len_cm headcirc_cm  muac_cm
-#> 1   1     273   M         0 273        0     2300 42.06667    33.26667 9.433333
-#> 28  4     250   F         0 250        0     1500 42.03333    30.03333 8.066667
-#> 36  5     238   F         0 238        0     2390 43.46667    33.63333 9.166667
-#> 56  8     240   F         0 240        0     1800 41.73333    31.46667 8.033334
+#>    id gestage sex visitweek pma age_days wt_kg   len_cm headcirc_cm  muac_cm
+#> 1   1     273   M         0 273        0  2.30 42.06667    33.26667 9.433333
[…]
