### 2024-05-30 fb34c09
docs: Updated gigs main vignette to match SJ article/SAS vignette
