+++ vignettes/gigs.Rmd
@@ -22,44 +22,41 @@ intergrowth21st <- if (knitr::is_html_output()) {
 
 # Introduction
 
-gigs is designed to make working with outputs from either the
-`r intergrowth21st` project and World Health Organisation Child Growth
-Standards as easy as possible. It also provides functions for easy and
-reproducible classification of fetal, newborn, and infant growth.
+gigs is a package designed to make the implementation of the international
+growth standards from the `r intergrowth21st` project and World Health
+Organisation Child Growth Standards as easy as possible. It provides
+functions for easy and reproducible classification of fetal, newborn, and infant
+growth using international growth standards. The package allows conversions and
+classifications of growth measurements to centiles, z-scores, and common
+international metrics such as stunting.
 
-Which functions you want to use depend on whether you're **converting between
-values and z-scores/centiles**, or **classifying observations**. Let's start
-by loading the package. We're also turning off the warnings which GIGS prints
-if inputs are `NA`, `NaN`, or similarly invalid, as to avoid cluttering the
-console output.
+## Getting started
+
+Let's start by loading the package. We're also turning off the warnings which
+gigs prints if inputs are `NA`, `NaN`, or similarly invalid, as to avoid
+cluttering the console output.
 
 ```{r setup}
 library(gigs)
 gigs_options_set(new_value = "quiet")
 ```
 
-We will show off two use cases for gigs. The first is using gigs to generate
-continuous indices of growth (i.e. centiles or z-scores). This can be done in a
-single step by letting gigs do the work in the background, or you can
-
-[//]: # (TODO: MAKE PURDY)
-
-involves simpler
-'single-step' approach involves passing the hard work to gigs, and letting it
-pick the right growth standards to apply for each measurement in your data. The
-more complex 'step-by-step' approach involves you picking and choosing your
-growth standards throughout.
+This vignette has two examples for applying gigs. The first is using gigs to
+convert growth measurements into centiles or z-scores according to appropriate
+standards from either the `r intergrowth21st` or WHO Child Growth standards. The
+second is using gigs to convert growth measurements to categorical growth
+outcomes, in this case size-for-gestational age.
 
 ## Example data
-Before we get started, we need some data. This tutorial uses a sample of data
-from 300 infants enrolled in the Low birthweight Infant Feeding Exploration
-(LIFE) study [@Vesel2022LowSettings; @Vesel2023FeedingStudy]. Thse infants were
-assessed at birth then followed up at intervals after birth, with data available
-up to six months of age in this data extract.
+This tutorial uses a sample of data from 300 infants enrolled in the Low
+birthweight Infant Feeding Exploration (LIFE) study [@Vesel2022LowSettings;
+@Vesel2023FeedingStudy]. These infants were assessed at birth then followed up
+at intervals after birth, with data available up to six months of age in this
+data extract.
 
 Though `gigs::life6mo` has data on length in cm (`len_cm`), head circumference
 (`headcirc_cm`), and mid-upper arm circumference (`muac_cm`), we are going to
-subset this dataset to focus on weight in grams (`weight_g`). The dataset
[…]
