+++ vignettes/gigs_bib.bib
@@ -27,6 +27,131 @@
   pages = {e067316}
 }
 
+@article{Papageorghiou2014AInternationalPregnancy,
+	title = {International standards for fetal growth based on serial ultrasound measurements: the {Fetal} {Growth} {Longitudinal} {Study} of the {INTERGROWTH}-21st {Project}},
+	volume = {384},
+	issn = {0140-6736, 1474-547X},
+	shorttitle = {International standards for fetal growth based on serial ultrasound measurements},
+	doi = {10.1016/S0140-6736(14)61490-2},
+	language = {English},
+	number = {9946},
+	urldate = {2024-01-16},
+	journal = {The Lancet},
+	author = {Papageorghiou, Aris T. and Ohuma, Eric O. and Altman, Douglas G. and Todros, Tullia and Cheikh Ismail, Leila and Lambert, Ann and Jaffer, Yasmin A. and Bertino, Enrico and Gravett, Michael G. and Purwar, Manorama and Noble, J. Alison and Pang, Ruyan and Victora, Cesar G. and Barros, Fernando C. and Carvalho, Maria and Salomon, Laurent J. and Bhutta, Zulfiqar A. and Kennedy, Stephen H. and Villar, José},
+	month = sep,
+	year = {2014},
+	pmid = {25209488},
+	pages = {869--879},
+}
+
+@article{Stirnemann2017InternationalProject,
+	title = {International estimated fetal weight standards of the {INTERGROWTH}-21st {Project}},
+	volume = {49},
+	copyright = {© 2016 The Authors. Ultrasound in Obstetrics \& Gynecology published by John Wiley \& Sons Ltd on behalf of the International Society of Ultrasound in Obstetrics and Gynecology.},
+	issn = {1469-0705},
+	doi = {10.1002/uog.17347},
+	abstract = {Objective Estimated fetal weight (EFW) and fetal biometry are complementary measures used to screen for fetal growth disturbances. Our aim was to provide international EFW standards to complement the INTERGROWTH-21st Fetal Growth Standards that are available for use worldwide. Methods Women with an accurate gestational-age assessment, who were enrolled in the prospective, international, multicenter, population-based Fetal Growth Longitudinal Study (FGLS) and INTERBIO-21st Fetal Study (FS), two components of the INTERGROWTH-21st Project, had ultrasound scans every 5 weeks from 9–14 weeks' until 40 weeks' gestation. At each visit, measurements of fetal head circumference (HC), biparietal diameter, occipitofrontal diameter, abdominal circumference (AC) and femur length (FL) were obtained blindly by dedicated research sonographers using standardized methods and identical ultrasound machines. Birth weight was measured within 12 h of delivery by dedicated research anthropometrists using standardized methods and identical electronic scales. Live babies without any congenital abnormality, who were born within 14 days of the last ultrasound scan, were selected for inclusion. As most births occurred at around 40 weeks' gestation, we constructed a bootstrap model selection and estimation procedure based on resampling of the complete dataset under an approximately uniform distribution of birth weight, thus enriching the sample size at extremes of fetal sizes, to achieve consistent estimates across the full range of fetal weight. We constructed reference centiles using second-degree fractional polynomial models. Results Of the overall population, 2404 babies were born within 14 days of the last ultrasound scan. Mean time between the last scan and birth was 7.7 (range, 0–14) days and was uniformly distributed. Birth weight was best estimated as a function of AC and HC (without FL) as log(EFW) = 5.084820 − 54.06633 × (AC/100)3 − 95.80076 × (AC/100)3 × log(AC/100) + 3.136370 × (HC/100), where EFW is in g and AC and HC are in cm. All other measures, gestational age, symphysis–fundus height, amniotic fluid indices and interactions between biometric m
[…]
