+++ pkgdown.yml
@@ -1,11 +1,11 @@
-pandoc: 2.19.2
-pkgdown: 2.0.7
+pandoc: 3.1.11
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
   benchmarking: benchmarking.html
-  ig-nbs-body-composition: ig-nbs-body-composition.html
   gigs: gigs.html
-last_built: 2024-02-07T12:49Z
+  ig-nbs-body-composition: ig-nbs-body-composition.html
+last_built: 2024-07-01T13:17Z
 urls:
   reference: https://simpar1471.github.io/gigs/reference
   article: https://simpar1471.github.io/gigs/articles
