+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   benchmarking: benchmarking.html
   gigs: gigs.html
   ig-nbs-body-composition: ig-nbs-body-composition.html
-last_built: 2024-07-01T13:17Z
+last_built: 2024-07-01T13:19Z
 urls:
   reference: https://simpar1471.github.io/gigs/reference
   article: https://simpar1471.github.io/gigs/articles
