### 2024-08-30 #626 @chitrams
Hello @rkillick I'd like to help review this package, please! I did work on the original WHO Anthro package and would love to help with this implementation :)
