### 2024-09-05 0544fef
feat: added informative warnings for `classify_growth`
