+++ R/growth_classify.R
@@ -675,43 +675,105 @@ classify_growth <- function(
                                missing = c(weight_kg = missing_weight,
                                            lenht_cm = missing_lenht,
                                            headcirc_cm = missing_headcirc))
-    rlang::inform(msg)
+    rlang::inform(msg, class = "gigs_classify_growth_msg")
   }
   .data
 }
 
-# INTERNAL ---------------------------------------------------------------------
+# INTERNAL; checkers for all growth_classify.R functions -----------------------
 
+#' Throw an error if elements of vectors in `.new` are in `.data`
+#' @param new A character vector of length one or more with new column names to
+#'   check against `existing`.
+#' @param .data_colnames A character vector of length one or more with column
+#'   names to be checked against `new`.
+#' @return Returns `new` invisibly if there are no matches between `new` and
+#'   `existing`, else throws an error.
 #' @noRd
-check_.new_vector_lengths <- function(.new) {
-  .new_names <- names(.new)
-  for (idx in seq_along(.new)) {
-    curr_vec <- .new[[idx]]
-    curr_name <- .new_names[idx]
-    expected_len <- switch(curr_name, sfga = 3, svn = 2, wasting = 3,
-                           stunting = 3, wfa = 3, headsize = 2)
-    checkmate::assert_character(
-      curr_vec,
-      any.missing = FALSE,
-      len = expected_len,
-      .var.name = paste0(".new[[\"", curr_name, "\"]]")
-    )
+err_if_.new_in_.data <- function(new, .data_colnames) {
+  matches <- new %in% .data_colnames
+  any_matches <- any(matches)
+  if (any_matches) {
+    matched_names <- paste0("`", new[matches], "`") |>
+      setNames(nm = rep.int("!", length(new[matches])))
+    rlang::abort(
+      message =
+        c(paste0("Column names requested in `.new` already exist in `.data`. ",
+                 "These are:"),
+          matched_names
+        ),
+      call = rlang::env_parent(),
+      class = "gigs_classify_.new_in_.data")
   }
+  invisible(new)
 }
 
 check_all_.new_names_valid <- function(.new, all) {
   .new_names <- names(.new)
   lgl_invalid_.new_name <- !.new_names %in% all
   if (any(lgl_invalid_.new_name)) {
-    rlang::abort(c(
-      "Some elements in `names(.new)` are not valid.",
-      "!" = paste0("Bad names in `.new`: ",
-                   paste_sep_commas_quoted(.new_names[lgl_invalid_.new_name])),
-      "i" = paste0("Each name in `.new` must be one of: ",
-                   paste_sep_commas_quoted(all))
-    ), call = rlang::expr(gigs::classify_growth()))
+    rlang::abort(
+      c(
+        "Some elements in `names(.new)` are not valid.",
+        "!" = paste0("Bad names in `.new`: ", paste_sep_commas_quoted(
+          .new_names[lgl_invalid_.new_name])
+        ),
+        "i" = paste0("Each name in `.new` must be one of: ",
+                     paste_sep_commas_quoted(all), ".")
+      ),
+      call = rlang::expr(gigs::classify_growth()),
+      class = "gigs_classify_growth_invalid_.new_names"
+    )
   }
-  return(invisible(.new))
+  invisible(.new)
+}
+
+# INTERNAL; checkers for classify_growth() -------------------------------------
+
+#' Check dot
+#' @noRd
+check_all_.outcomes_in_.new <- function(.outcomes, .new) {
+  lgl_.analyses_in_.new <- .outcomes %in% names(.new)
+  if (!all(lgl_.analyses_in_.new)) {
+    rlang::abort(
[…]
