### 2024-09-09 b54553f
feat: Fixed implementation of GIGS algorithm using `'id` parameter in GIGS WAZ/LAZ etc. functions. Quite slow right now.
