+++ R/gigs_zscoring.R
@@ -7,8 +7,8 @@
 #'   The growth standard used for each observation differs based on the
 #'   gestational and post-menstrual age of the infant being analysed. The
 #'   different procedures are based on the advised use of each growth standard,
-#'   where a measurement is considered a 'birth measurement' if taken <1 day
-#'   after birth:
+#'   where a measurement is considered a 'birth measurement' if it is the first
+#'   measurement taken from an infant and was taken <3 days after birth:
 #'   * `gest_days` < 24 weeks (168 days):
 #'     - Birth: No standard available
 #'     - Postnatal: IG-21<sup>st</sup> Postnatal Growth standards from 27 to 64
@@ -35,16 +35,21 @@
 #' @param age_days A numeric vector of length one or more with ages in days.
 #' @param gest_days A numeric vector of length one or more with gestational ages
 #'   in days.
+#' @param id A factor variable with an ID for each observation. When not `NULL`,
+#'   this variable is used to ensure that only the earliest measurement for each
+#'   individual is used as a birth measure. Leave this argument as `NULL` if all
+#'   your data comes from the same individual. Default = `NULL`.
 #' @inheritParams shared_roxygen_params
 #' @note These functions expect vectors which are recyclable with
 #'   [vctrs::vec_recycle_common()].
 #' @rdname gigs_zscoring
-#' @noRd
-gigs_waz <- function(weight_kg, age_days, gest_days, sex) {
+#' @export
+gigs_waz <- function(weight_kg, age_days, gest_days, sex, id = NULL) {
   validate_waz_params(weight_kg = weight_kg,
                       age_days = age_days,
                       gest_days = gest_days,
-                      sex = sex) |>
+                      sex = sex,
+                      id = id) |>
     do.call(what = gigs_waz_internal)
 }
 
@@ -52,35 +57,38 @@ gigs_waz <- function(weight_kg, age_days, gest_days, sex) {
 #' @param lenht_cm A numeric vector of length one or more with length or height
 #'   in cm. Should be recumbent length when `age_days < 731`, and standing
 #'   height when `age_days >= 731`.
-#' @noRd
-gigs_lhaz <- function(lenht_cm, age_days, gest_days, sex) {
+#' @export
+gigs_lhaz <- function(lenht_cm, age_days, gest_days, sex, id = NULL) {
   validate_lhaz_params(lenht_cm = lenht_cm,
                        age_days = age_days,
                        gest_days = gest_days,
-                       sex = sex) |>
+                       sex = sex,
+                       id = id) |>
     do.call(what = gigs_lhaz_internal)
 }
 
 #' @rdname gigs_zscoring
-#' @noRd
-gigs_wlz <- function(weight_kg, lenht_cm, age_days, gest_days, sex) {
+#' @export
+gigs_wlz <- function(weight_kg, lenht_cm, age_days, gest_days, sex, id = NULL) {
   validate_wlz_params(weight_kg = weight_kg,
                       lenht_cm = lenht_cm,
                       age_days = age_days,
                       gest_days = gest_days,
-                      sex = sex) |>
+                      sex = sex,
+                      id = id) |>
     do.call(what = gigs_wlz_internal)
 }
 
 #' @rdname gigs_zscoring
 #' @param headcirc_cm Numeric vector of length one or more with head
[…]
