+++ NAMESPACE
@@ -20,9 +20,13 @@ export(compute_stunting)
 export(compute_svn)
 export(compute_wasting)
 export(compute_wfa)
+export(gigs_hcaz)
+export(gigs_lhaz)
 export(gigs_option_get)
 export(gigs_option_set)
 export(gigs_options_set)
+export(gigs_waz)
+export(gigs_wlz)
 export(ig_fet_acfga_centile2value)
 export(ig_fet_acfga_value2centile)
 export(ig_fet_acfga_value2zscore)
