+++ man/ig_fet_value2zscore.Rd
@@ -1,352 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ig_fet.R
-\name{ig_fet_value2zscore}
-\alias{ig_fet_value2zscore}
-\alias{ig_fet_hcfga_value2zscore}
-\alias{ig_fet_bpdfga_value2zscore}
-\alias{ig_fet_acfga_value2zscore}
-\alias{ig_fet_flfga_value2zscore}
-\alias{ig_fet_ofdfga_value2zscore}
-\alias{ig_fet_efwfga_value2zscore}
-\alias{ig_fet_sfhfga_value2zscore}
-\alias{ig_fet_crlfga_value2zscore}
-\alias{ig_fet_gafcrl_value2zscore}
-\alias{ig_fet_gwgfga_value2zscore}
-\alias{ig_fet_pifga_value2zscore}
-\alias{ig_fet_rifga_value2zscore}
-\alias{ig_fet_sdrfga_value2zscore}
-\alias{ig_fet_tcdfga_value2zscore}
-\alias{ig_fet_gaftcd_value2zscore}
-\alias{ig_fet_poffga_value2zscore}
-\alias{ig_fet_sffga_value2zscore}
-\alias{ig_fet_avfga_value2zscore}
-\alias{ig_fet_pvfga_value2zscore}
-\alias{ig_fet_cmfga_value2zscore}
-\alias{ig_fet_hefwfga_value2zscore}
-\alias{ig_fet_value2centile}
-\alias{ig_fet_hcfga_value2centile}
-\alias{ig_fet_bpdfga_value2centile}
-\alias{ig_fet_acfga_value2centile}
-\alias{ig_fet_flfga_value2centile}
-\alias{ig_fet_ofdfga_value2centile}
-\alias{ig_fet_efwfga_value2centile}
-\alias{ig_fet_sfhfga_value2centile}
-\alias{ig_fet_crlfga_value2centile}
-\alias{ig_fet_gafcrl_value2centile}
-\alias{ig_fet_gwgfga_value2centile}
-\alias{ig_fet_pifga_value2centile}
-\alias{ig_fet_rifga_value2centile}
-\alias{ig_fet_sdrfga_value2centile}
-\alias{ig_fet_tcdfga_value2centile}
-\alias{ig_fet_gaftcd_value2centile}
-\alias{ig_fet_poffga_value2centile}
-\alias{ig_fet_sffga_value2centile}
-\alias{ig_fet_avfga_value2centile}
-\alias{ig_fet_pvfga_value2centile}
-\alias{ig_fet_cmfga_value2centile}
-\alias{ig_fet_hefwfga_value2centile}
-\title{Convert values to z-scores/centiles in the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Fetal
-Growth standards}
-\usage{
-ig_fet_value2zscore(y, x, acronym)
-
-ig_fet_hcfga_value2zscore(headcirc_mm, gest_days)
-
-ig_fet_bpdfga_value2zscore(bpd_mm, gest_days)
-
-ig_fet_acfga_value2zscore(abdocirc_mm, gest_days)
-
-ig_fet_flfga_value2zscore(femurlen_mm, gest_days)
-
-ig_fet_ofdfga_value2zscore(ofd_mm, gest_days)
-
-ig_fet_efwfga_value2zscore(efw_g, gest_days)
-
-ig_fet_sfhfga_value2zscore(sfh_cm, gest_days)
-
-ig_fet_crlfga_value2zscore(crl_mm, gest_days)
-
[…]
