+++ man/classify_growth.Rd
@@ -1,179 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/growth_classify.R
-\name{classify_growth}
-\alias{classify_growth}
-\title{Classify multiple growth indicators at the same time using GIGS-recommended
-growth standards}
-\usage{
-classify_growth(
-  .data,
-  gest_days,
-  age_days,
-  sex,
-  weight_kg = NULL,
-  lenht_cm = NULL,
-  headcirc_cm = NULL,
-  .outcomes = c("sfga", "svn", "stunting", "wasting", "wfa", "headsize"),
-  .new = list(sfga = c("birthweight_centile", "sfga", "sfga_severe"), svn =
-    c("birthweight_centile", "svn"), stunting = c("lhaz", "stunting",
-    "stunting_outliers"), wasting = c("wlz", "wasting", "wasting_outliers"), wfa =
-    c("waz", "wfa", "wfa_outliers"), headsize = c("hcaz", "headsize")),
-  .verbose = TRUE
-)
-}
-\arguments{
-\item{.data}{A \code{data.frame}-like tabular object with one or more rows. Must
-be permissible by \code{\link[checkmate:checkDataFrame]{checkmate::assert_data_frame()}}, so you can also supply
-a \code{tibble}, \code{data.table}, or similar.}
-
-\item{gest_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a numeric vector of gestational age(s) at birth
-in days. This column, in conjunction with the column referred to by
-\code{age_days}, is used to select which growth standard to use for each
-observation.}
-
-\item{age_days}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a numeric vector of age values in days.}
-
-\item{sex}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a case-sensitive character vector of sexes,
-either \code{"M"} (male) or \code{"F"} (female). By default, gigs will warn you if
-any elements of \code{sex} are not \code{"M"} or \code{"F"}, or are missing (\code{NA}). You
-can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}
-
-\item{weight_kg}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a numeric vector of weight values in kg. When
-performing size-for-GA and small vulnerable newborn classifications,
-centiles and classifications will only be provided where measurements were
-taken <12hrs after birth, i.e. \code{age_days < 0.5}.}
-
-\item{lenht_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a numeric vector of length/height values in cm.}
-
-\item{headcirc_cm}{<\code{\link[rlang:args_data_masking]{data-masking}}> The name of a
-column in \code{.data} which is a numeric vector of head circumference values in
-cm.}
-
-\item{.outcomes}{A character vector of up to six elements in length
-describing which growth outcomes you want to assess. Use this when you
-supply data that can be used to generate multiple growth indicators, but
-only want to run a specific set. This argument is case-sensitive, and an
-error will be thrown if you supply a string in \code{.analyses} which is either
-\code{NA} or is not a member of this argument's default. Default = \code{c("sfga", "svn", "stunting", "wasting", "wfa", "headsize")}.}
-
-\item{.new}{A list with names corresponding to \code{.analyses}, which describes
-the names of new columns to be added to \code{.data}. If any elements of the
-vectors in \code{.new} are equal to existing elements of \code{colnames(.data)}, the
[…]
