+++ man/shared_roxygen_params.Rd
@@ -1,29 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\name{shared_roxygen_params}
-\alias{shared_roxygen_params}
-\title{Shared roxygen parameters}
-\arguments{
-\item{y}{Numeric vector of length one or more with anthropometric
-measurement(s) to convert to centiles/z-scores. Units depend on which
-\code{acronym}(s) are in use.}
-
-\item{z, p}{Numeric vector of length one or more with centiles/z-scores to
-convert to values. For \code{p}, gigs will warn you if elements of \code{p} are not
-between \code{0} and \code{1}. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}
-
-\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
-(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
-will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
-and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}
-}
-\description{
-Many of the parameters in \verb{[gigs]} are shared between different functions.
-This roxygen block makes sharing these parameters more simple.
-}
-\note{
-Input vectors other than \code{acronym} are recycled by
-\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must adhere to the \link{vctrs} recycling
-rules.
-}
-\keyword{internal}

[…]
