### 2024-09-19 278c5e8
fix: When sex is not "M"/"F", only one `i` will show (used to be two).
