### 2024-09-19 8158d2c
fix: Removed global assignment operators in `catch_and_throw_validate_issues()`
