### 2024-09-19 c8b892c
docs: Documented all functions in `check_params.R` with `roxygen2`
