+++ R/growth_classify.R
@@ -163,15 +163,29 @@ classify_svn <- function(.data,
 #'   * `stunting_outliers` - Factor of stunting categories with outlier flagging
 #' @inherit categorise_stunting details references
 #' @examples
-#' # The first observation uses the INTERGROWTH-21st Postnatal Growth standards;
-#' # the next two use the WHO Child Growth Standards.
+#' # This dummy dataset contains data from two people, from birth (<3 days) to
+#' # 500 days of age.
 #' data <- data.frame(
-#'   length_cm = c(52.2, 75.4, 63.1),
-#'   agedays = c(357, 375, 250),
-#'   gestage  = c(196, 287, 266),
-#'   sex = c("M", "M", "F")
+#'   child = factor(rep.int(c("A", "B"), c(3, 3))),
+#'   agedays = c(0, 100, 500, 2, 100, 500),
+#'   gestage  = c(rep(35 * 7, 3), rep(35 * 7, 3)),
+#'   sex = rep.int(c("M", "F"), c(3, 3)),
+#'   weight_kg = c(3, 6, 9, 3, 6, 9),
+#'   length_cm = rep.int(c(52.2, 60.4, 75), 2),
+#'   headcirc_cm = rep.int(c(40, 50, 60), 2)
 #' )
 #'
+#' # Use the `id` argument to ensure that `classify_stunting` uses the correct
+#' # standard for each observation
+#' data |>
+#'   classify_stunting(lenht_cm = length_cm,
+#'                     age_days = agedays,
+#'                     gest_days = gestage,
+#'                     sex = sex,
+#'                     id = child)
+#'
+#' # If you don't specify `id`, `classify_stunting` will assume data is from one
+#' # child only
 #' data |>
 #'   classify_stunting(lenht_cm = length_cm,
 #'                     age_days = agedays,
@@ -184,19 +198,26 @@ classify_stunting <- function(
     age_days,
     gest_days,
     sex,
+    id = NULL,
     .new = c("lhaz", "stunting", "stunting_outliers")) {
   checkmate::assert_data_frame(.data, min.rows = 1)
   checkmate::qassert(.new, rules = "S3")
   err_if_.new_in_.data(new = .new, .data_colnames = colnames(.data))
 
-  lhaz <- gigs_lhaz(
-    lenht_cm = eval_tidy(enquo(lenht_cm), .data),
-    age_days = eval_tidy(enquo(age_days), .data),
-    gest_days = eval_tidy(enquo(gest_days), .data),
-    sex = eval_tidy(enquo(sex), .data)
-  )
+  if (!rlang::quo_is_null(enquo(id))) {
+    id <- eval_tidy(enquo(id), data = .data)
+  }
 
   .new <- vctrs::vec_as_names(.new, repair = "universal")
+  lhaz <- validate_lhaz_params(
+    lenht_cm = eval_tidy(enquo(lenht_cm), data = .data),
+    age_days = eval_tidy(enquo(age_days), data = .data),
+    gest_days = eval_tidy(enquo(gest_days), data = .data),
+    sex = eval_tidy(enquo(sex), data = .data),
+    id = id
+  ) |>
+    do.call(what = gigs_lhaz_internal)
+
   .data[[.new[1]]] <- lhaz
   .data[[.new[2]]] <- categorise_stunting_internal(lhaz, outliers = FALSE)
   .data[[.new[3]]] <- categorise_stunting_internal(lhaz, outliers = TRUE)
@@ -224,19 +245,34 @@ classify_stunting <- function(
 #' @inherit categorise_wasting details
 #' @inherit categorise_stunting references
 #' @examples
-#' # Returns factor with stunting classifications
[…]
