### 2024-09-19 1fa0b99
refactor: Internal changes - internal z-scoring functions now accept cached GIGS z-scoring logicals as input, which speeds up classify_growth()
