+++ R/ig_fet_estimation.R
@@ -187,8 +187,8 @@ validate_ig_fet_estimate_fetal_weight <- function(abdocirc_mm, headcirc_mm) {
   validate_parameter_lengths(abdocirc_mm = abdocirc_mm,
                              headcirc_mm = headcirc_mm)
   catch_and_throw_validate_issues({
-    abdocirc_mm <- validate_ig_fet_estimate_fw_param(abdocirc_mm, "abdocirc_mm")
     headcirc_mm <- validate_ig_fet_estimate_fw_param(headcirc_mm, "headcirc_mm")
+    abdocirc_mm <- validate_ig_fet_estimate_fw_param(abdocirc_mm, "abdocirc_mm")
   }, call = rlang::caller_env())
   vctrs::vec_recycle_common(abdocirc_mm = abdocirc_mm,
                             headcirc_mm = headcirc_mm)
@@ -202,14 +202,14 @@ validate_ig_fet_estimate_ga <- function(crl_mm, headcirc_mm, femurlen_mm) {
                              headcirc_mm = headcirc_mm,
                              femurlen_mm = femurlen_mm)
   catch_and_throw_validate_issues({
-    crl_mm <- if (is.null(crl_mm)) NULL else {
-      validate_numeric(crl_mm, "crl_mm")
+    femurlen_mm <- if (is.null(femurlen_mm)) NULL else {
+      validate_numeric(femurlen_mm, "femurlen_mm")
     }
     headcirc_mm <- if (is.null(headcirc_mm)) NULL else {
       validate_numeric(headcirc_mm, "headcirc_mm")
     }
-    femurlen_mm <- if (is.null(femurlen_mm)) NULL else {
-      validate_numeric(femurlen_mm, "femurlen_mm")
+    crl_mm <- if (is.null(crl_mm)) NULL else {
+      validate_numeric(crl_mm, "crl_mm")
     }
   }, call = rlang::caller_env())
   vctrs::vec_recycle_common(crl_mm = crl_mm, headcirc_mm = headcirc_mm,
