### 2024-09-19 1f01270
feat: In growth_classify.R, `classify_svn()` and `classify_sfga()` now throw informative warnings/errors if their input data is malformed
