### 2024-09-19 a04459d
feat: In growth_classify.R, all functions give consistent warnings when names in `.new` need repairing
