### 2024-09-19 a1fd202
refactor+docs: In growth_classify.R, all functions are now documented with roxygen2 + provide consistent warnings
