+++ R/growth_classify.R
@@ -634,7 +634,7 @@ classify_growth <- function(
   all_analyses <- c("sfga", "svn", "stunting", "wasting", "wfa", "headsize")
   all_analyses_len <- length(all_analyses)
   # Which analyses have been run
-  outcomes_run <- rep_len(x = FALSE, length.out = all_analyses_len) |>
+  outcomes_run <- logical(length = all_analyses_len) |>
     setNames(all_analyses)
 
   # Variable checking - basic assertions on length/type
@@ -801,20 +801,21 @@ classify_growth <- function(
 
 # INTERNAL; checkers for all growth_classify.R functions -----------------------
 
-#' Throw an error if elements of vectors in `.new` are in `.data`
-#' @param new A character vector of length one or more with new column names to
-#'   check against `existing`.
+#' For `classify_*() functions`:Throw an error if elements of vectors in `.new`
+#' are in `.data`.
+#' @param .new A character vector of length one or more with new column names to
+#'   check against `.data`.
 #' @param .data_colnames A character vector of length one or more with column
-#'   names to be checked against `new`.
-#' @return Returns `new` invisibly if there are no matches between `new` and
+#'   names to be checked against `.new`.
+#' @return Returns `.new` invisibly if there are no matches between `new` and
 #'   `existing`, else throws an error.
 #' @noRd
-err_if_.new_in_.data <- function(new, .data_colnames) {
-  matches <- new %in% .data_colnames
+err_if_.new_in_.data <- function(.new, .data_colnames) {
+  matches <- .new %in% .data_colnames
   any_matches <- any(matches)
   if (any_matches) {
-    matched_names <- paste0("`", new[matches], "`") |>
-      setNames(nm = rep.int("!", length(new[matches])))
+    matched_names <- paste0("`", .new[matches], "`") |>
+      setNames(nm = rep.int("!", length(.new[matches])))
     rlang::abort(
       message =
         c(paste0("Column names requested in `.new` already exist in `.data`. ",
@@ -824,9 +825,18 @@ err_if_.new_in_.data <- function(new, .data_colnames) {
       call = rlang::env_parent(),
       class = "gigs_classify_.new_in_.data")
   }
-  invisible(new)
+  invisible(.new)
 }
 
+
+
+# INTERNAL; checkers for classify_growth() -------------------------------------
+
+#' For `classify_growth()`: Check that the names of vectors in `.new` map onto
+#' the range of available analyses (i.e. `all`).
+#' @inheritParams classify_growth
+#' @returns Invisibly returns `.new`, unchanged.
+#' @noRd
 check_all_.new_names_valid <- function(.new, all) {
   .new_names <- names(.new)
   lgl_invalid_.new_name <- !.new_names %in% all
@@ -847,9 +857,10 @@ check_all_.new_names_valid <- function(.new, all) {
   invisible(.new)
 }
 
-# INTERNAL; checkers for classify_growth() -------------------------------------
-
-#' Check dot
+#' For `classify_growth()`: check that all requested outcomes in `.outcomes`
+#' have named vectors with column names in `.new`.
+#' @inheritParams classify_growth
+#' @returns @returns Invisibly returns `.new`, unchanged.
 #' @noRd
 check_all_.outcomes_in_.new <- function(.outcomes, .new) {
   lgl_.analyses_in_.new <- .outcomes %in% names(.new)
@@ -862,8 +873,17 @@ check_all_.outcomes_in_.new <- function(.outcomes, .new) {
       class = "gigs_classify_growth_.outcomes_not_in_.new"
     )
   }
+  invisible(.new)
 }
 
+#' Check whether vectors in `.new` are the correct length, given the number of
+#' new columns to be generated
+#' @inheritParams classify_growth
+#' @returns Invisibly returns `.new`, unchanged.
[…]
