### 2024-09-19 bb91ac3
docs/refactor: Insignificant line removals/indentation change
