### 2024-09-19 57dc9f2
docs: Updated benchmarking article to include AGD; updated README.Rmd accordingly; added alt text to all articles
