### 2024-09-19 512f581
docs: Updated NEWS.md based on changes from today
