### 2024-09-20 8cb46ef
docs: Updated gigs.Rmd to have alt text + IDs
