### 2024-09-20 9330548
docs: Added files from re-running devtools::document()
