+++ tests/testthat/helper-gigs_zscoring.R
@@ -1,0 +1,260 @@
+#' Generate growth data for testing
+#' @param n Integer > 10; number of rows in the returned data frame. Must be
+#'   divisible by 5.
+#' @param seed Integer; seed used to generate GA, age, and z-score data.
+#' @param restrict Logical; Should the function automatically remove rows in the
+#'   output where one of `weight_kg`, `length_cm`, or `headcirc_cm` are `NA`?
+#'   Default = `TRUE`.
+#' @note This function produces data which is always in the GA range of the
+#'   INTERGROWTH-21st Newborn Size standards and age range of the WHO standards.
+#'   Whether observations can be used with the INTERGROWTH-21st PNG standards
+#'   depends on the random seed used.
+#' @returns A data frame with `n` rows and 11 columns containing:
+#'   * `gest_age`: Gestational age in days
+#'   * `age_days`: Chronological age in days
+#'   * `sex`: Sex (`"M"` = male; `"F"` = female)
+#'   * `waz`: Weight-for-age z-score
+#'   * `lhaz`: Length/height-for-age z-score
+#'   * `wlz`: Weight-for-length/height z-score
+#'   * `hcaz`: Head circumference-for-age z-score
+#'   * `weight_kg`: Expected weight, given the value of `waz` and age/sex
+#'   * `lenht_cm`: Expected length/height, given the value of `lhaz` and age/sex
+#'   * `weight_kg_from_wlz`: Expected weight, given the value of `wlz` and
+#'        length/height/sex
+#'   * `headcirc_cm`: Expected head circumference, given the value of `hcaz` and
+#'        age/sex
+#' @note This is **not** suitable for simulating growth data accurately. The
+#'   z-score sampling procedure does **not** account for autocorrelation in
+#' @noRd
+gigs_random_growth_dataset <- function(n = 10L,
+                                       seed = 472346,
+                                       restrict = TRUE) {
+  checkmate::assert_int(n, lower = 10)
+  checkmate::assert_integerish(n / 5, lower = 1)
+  checkmate::assert_int(seed)
+  who_lim <- 1856L
+
+  data <- withr::with_seed(seed = seed, {
+    data.frame(
+      id = cut(1:n, breaks = n / 5, labels = 1:(n/5))
+    ) |>
+      dplyr::mutate(
+        gest_age = 24 * 7 + base::sample(x = 0:32, size = 1, replace = TRUE),
+        age_days = base::sample.int(
+          size = 5,
+          n = who_lim,
+          replace = FALSE,
+          prob = c(0.04, 0.04, 0.04, 0.04, rep.int(x = 0.0005, times = who_lim - 4))
+        ),
+        sex = base::sample(x = c("M", "F"), size = 1, replace = FALSE),
+        waz_exp = stats::rnorm(n = 5),
+        lhaz_exp = stats::rnorm(n = 5),
+        wlz_exp = stats::rnorm(n = 5),
+        hcaz_exp = stats::rnorm(n = 5),
+        .by = id
+      ) |>
+    dplyr::arrange(id, age_days)
+  })
+  gigs_options_set("quiet", TRUE)
+
+  gigs_lgls <- with(data, gigs_zscoring_lgls(age_days, gest_age, id))
+
+  out <- data |>
+    dplyr::mutate(
+      id = droplevels(id),
+      weight_kg = weight_from_waz(waz_exp,
+                                  gest_days = gest_age,
+                                  age_days = age_days,
+                                  sex = sex,
+                                  id = id,
+                                  gigs_lgls = gigs_lgls),
+      lenht_cm = lenht_from_lhaz(lhaz_exp,
+                                 gest_days = gest_age,
[…]
