+++ man/ig_nbs_coeffs.Rd
@@ -1,0 +1,36 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{ig_nbs_coeffs}
+\alias{ig_nbs_coeffs}
+\title{INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}} Newborn Size Standards GAMLSS coefficients}
+\source{
+Mu/sigma/nu/tau values were provided by Dr Eric Ohuma.
+}
+\description{
+A set of nested lists containing mu, sigma, nu and tau values across
+gestational ages for either sex, for the INTERGROWTH-21\if{html}{\out{<sup>}}st\if{html}{\out{</sup>}}
+weight/length/head circumference-for-gestational age standards. The lists are
+ordered by acronym, then sex.
+}
+\examples{
+names(gigs::ig_nbs_coeffs)
+head(gigs::ig_nbs_coeffs$wfga$male)
+}
+\references{
+Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al.
+\strong{International standards for newborn weight, length, and head circumference
+by gestational age and sex: the Newborn Cross-Sectional Study of the
+INTERGROWTH-21st Project.} \emph{Lancet} 2014, \strong{384(9946):857-68.}
+\doi{10.1016/S0140-6736(14)60932-6}
+
+Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
+\strong{INTERGROWTH-21st very preterm size at birth reference charts.} \emph{Lancet}
+2016, \strong{387(10021):844-45.} \doi{10.1016/S0140-6736(16)00384-6}
+
+Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al.
+\strong{Body composition at birth and its relationship with neonatal anthropometric
+ratios: the newborn body composition study of the INTERGROWTH-21st project.}
+\emph{Pediatric Research} 2017, \strong{82:305-316.} \doi{10.1038/pr.2017.52}
+}
+\keyword{data}
