+++ man/compute_wfa.Rd
@@ -1,0 +1,79 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/growth_compute.R
+\name{compute_wfa}
+\alias{compute_wfa}
+\title{Get weight-for-age categories using multiple vectors and GIGS-recommended
+growth standards}
+\usage{
+compute_wfa(weight_kg, age_days, gest_days, sex, outliers = FALSE)
+}
+\arguments{
+\item{weight_kg}{Numeric vector of length one or more with weight
+measurement(s) in kg.}
+
+\item{age_days}{Numeric vector of length one or more with age(s) in days
+for each child. Should be between \code{0} to \code{1856} days. By default, gigs will
+replace out-of-bounds elements in \code{age_days} with \code{NA} and warn you. This
+behaviour can be customised using the functions in \link{gigs_options}.}
+
+\item{gest_days}{Numeric vector with gestational age(s) at birth in days.}
+
+\item{sex}{Character vector of length one or more with sex(es), either \code{"M"}
+(male) or \code{"F"} (female). This argument is case-sensitive. By default, gigs
+will replace elements of \code{sex} which are not \code{"M"} or \code{"F"} with \code{NA}
+and warn you. You can customise this behaviour using the \link[=gigs_options]{GIGS package-level options}.}
+
+\item{outliers}{A single \code{TRUE} or \code{FALSE} value specifying whether
+implausible z-score thresholds should be applied. Default = \code{FALSE}.}
+}
+\value{
+An object of class factor with the same length as \code{wlz}, containing
+wasting classifications. Its levels are \code{c("wasting_severe", "wasting", "not_wasting", "overweight")} if \code{outliers = FALSE} (the default), else
+\code{c("wasting_severe", "wasting", "not_wasting", "overweight", "outlier")}.
+}
+\description{
+Get weight-for-age categories using multiple vectors and GIGS-recommended
+growth standards
+}
+\details{
+Cut-offs for weight-for-age categories are:
+
+\tabular{lll}{
+\strong{Category}    \tab \strong{Factor level}  \tab
+\strong{Z-score bounds} \cr
+Severely underweight \tab \code{"underweight_severe"} \tab \code{waz} ≤ -3      \cr
+Underweight          \tab \code{"underweight"}        \tab -3 < \code{waz} ≤ -2 \cr
+Normal weight        \tab \code{"normal_weight"}      \tab \code{abs(waz)} < 2  \cr
+Overweight           \tab \code{"overweight"}         \tab \code{waz} ≥ 2       \cr
+Outlier              \tab \code{"outlier"}        \tab \code{waz} < -6 or \code{waz} > 5
+}
+}
+\examples{
+compute_wfa(
+  weight_kg = c(7.2, 4.5, 9.1, 24),
+  age_days = c(401, 185, 101, 607),
+  gest_days = 7 * c(27, 36, 40, 41),
+  sex = c("F", "M", "F", "M")
+)
+
+# And with outlier flagging:
+compute_wfa(
+  weight_kg = c(7.2, 4.5, 9.1, 24),
+  age_days = c(401, 185, 101, 607),
+  gest_days = 7 * c(27, 36, 40, 41),
+  sex = c("F", "M", "F", "M"),
+  outliers = TRUE
+)
+}
+\references{
+\strong{'Implausible z-score values'} \emph{in} World Health Organization (ed.)
+\emph{Recommendations for data collection, analysis and reporting on
+anthropometric indicators in children under 5 years old}. Geneva: World
+Health Organization and the United Nations Children's Fund UNICEF, (2019).
+pp. 64-65.
+
[…]

[…]
