### 2024-09-20 9eb3f2f
docs: Updated _pkgdown.yml to include GIGS z-scoring functions
