### 2024-09-20 8c7383c
docs: Updated CONTRIBUTING.md to reflect that we've made the GIGS z-scoring functions user-facing; updated DESCRIPTION with email now I've moved from LSHTM; updated CITATION and related files
