+++ DESCRIPTION
@@ -3,8 +3,8 @@ Package: gigs
 Title: Assess Fetal, Newborn, and Child Growth with International Standards
 Version: 0.4.1.9000
 Authors@R: c(
-    person("Simon", "Parker", , "simon.parker@lshtm.ac.uk", role = c("aut", "cre"),
-           comment = c(ORCID = "0009-0003-8214-4496")),
+    person("Simon", "Parker", , "simon.parker1471@outlook.com",
+           role = c("aut", "cre"), comment = c(ORCID = "0009-0003-8214-4496")),
     person("Linda", "Vesel", role = "dtc",
            comment = c(ORCID = "0000-0003-3753-4172")),
     person("Eric", "Ohuma", role = "dtc",
@@ -34,6 +34,7 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
+    dplyr,
     units,
     withr
 VignetteBuilder: 
