+++ CITATION.cff
@@ -1,7 +1,7 @@
-# -----------------------------------------------------------
-# CITATION file created with {cffr} R package, v0.5.0
+# --------------------------------------------
+# CITATION file created with {cffr} R package
 # See also: https://docs.ropensci.org/cffr/
-# -----------------------------------------------------------
+# --------------------------------------------
  
 cff-version: 1.2.0
 message: 'To cite package "gigs" in publications use:'
@@ -9,6 +9,9 @@ type: software
 license: GPL-3.0-or-later
 title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
 version: 0.4.1.9000
+identifiers:
+- type: url
+  value: https://lshtm-gigs.github.io/gigs/
 abstract: Convert between anthropometric measures and z-scores/centiles in multiple
   growth standards, and classify fetal, newborn, and child growth accordingly. With
   a simple interface to growth standards from the World Health Organisation and International
@@ -17,7 +20,7 @@ abstract: Convert between anthropometric measures and z-scores/centiles in multi
 authors:
 - family-names: Parker
   given-names: Simon
-  email: simon.parker@lshtm.ac.uk
+  email: simon.parker1471@outlook.com
   orcid: https://orcid.org/0009-0003-8214-4496
 preferred-citation:
   type: manual
@@ -25,9 +28,13 @@ preferred-citation:
   authors:
   - family-names: Parker
     given-names: Simon
-    email: simon.parker@lshtm.ac.uk
+    email: simon.parker1471@outlook.com
     orcid: https://orcid.org/0009-0003-8214-4496
-  year: '2023'
+  - family-names: Vesel
+    given-names: Linda
+  - family-names: Ohuma
+    given-names: Eric
+  year: '2024'
   notes: 0.4.1.9000
   url: https://github.com/lshtm-gigs/gigs/
 repository-code: https://github.com/lshtm-gigs/gigs
@@ -35,7 +42,7 @@ url: https://github.com/lshtm-gigs/gigs/
 contact:
 - family-names: Parker
   given-names: Simon
-  email: simon.parker@lshtm.ac.uk
+  email: simon.parker1471@outlook.com
   orcid: https://orcid.org/0009-0003-8214-4496
 keywords:
 - anthropometry
@@ -52,11 +59,10 @@ references:
   url: https://www.R-project.org/
   authors:
   - name: R Core Team
-  location:
-    name: Vienna, Austria
-  year: '2024'
   institution:
     name: R Foundation for Statistical Computing
+    address: Vienna, Austria
+  year: '2024'
   version: '>= 4.1.0'
 - type: software
   title: checkmate
@@ -70,6 +76,7 @@ references:
     email: michellang@gmail.com
     orcid: https://orcid.org/0000-0001-9754-0393
   year: '2024'
+  doi: 10.32614/CRAN.package.checkmate
 - type: software
   title: gamlss.dist
   abstract: 'gamlss.dist: Distributions for Generalized Additive Models for Location
@@ -87,17 +94,17 @@ references:
     email: r.rigby@gre.ac.uk
     orcid: https://orcid.org/0000-0003-3853-1707
   year: '2024'
+  doi: 10.32614/CRAN.package.gamlss.dist
 - type: software
   title: stats
   abstract: 'R: A Language and Environment for Statistical Computing'
   notes: Imports
   authors:
   - name: R Core Team
-  location:
-    name: Vienna, Austria
-  year: '2024'
   institution:
     name: R Foundation for Statistical Computing
+    address: Vienna, Austria
[…]
