### 2024-09-20 1e4ad55
tests: Added GIGS z-scoring test file to increase coverage there; updated `test-growth_classify.R` to reflect changes to growth_classify.R
