+++ tests/testthat/test-growth_classify.R
@@ -1,3 +1,1106 @@
-test_that("test growth_classify", {
-  expect_equal(2 * 2, 4)
+#' @srrstats {G5.1} Users can run this function to make their own data set for
+#'   test confirmation
+test_data <- gigs_random_growth_dataset(10000, seed = 11, restrict = TRUE)
+
+test_data_birth <- test_data[with(test_data, !is.na(bweight_centile_exp)),]
+test_data_postnatal <- test_data[with(test_data, age_days > 3), ]
+test_data_postnatal_wlz <- test_data_postnatal[
+    !is.na(with(test_data_postnatal, weight_kg_from_wlz)),
+]
+
+# Testing `classify_sfga()` ----------------------------------------------------
+
+#' @srrstats {G5.4, G5.4c} Tests to ensure that `classify_sfga()` reproduces
+#'   expected z-scores from randomly generated growth data, and the appropriate
+#'   categorical data.
+test_that(desc = "classify_sfga() reproduces expected z-scores", {
+  sfga <- classify_sfga(
+    .data = test_data_birth,
+    weight_kg = weight_kg,
+    gest_days = gest_age,
+    sex = sex,
+    .new = c("bweight_centile_test", "sfga", "sfga_severe")
+  )
+  sfga_temp <- dplyr::select(sfga, id, tidyselect::starts_with("bweight"), tidyselect::starts_with("sfga"))
+  testthat::expect_equal(
+    sfga$bweight_centile_test, sfga$bweight_centile_exp
+  )
+  testthat::expect_equal(
+    categorise_sfga(p = sfga$bweight_centile_test, severe = FALSE),
+    sfga$sfga
+  )
+  testthat::expect_equal(
+    categorise_sfga(p = sfga$bweight_centile_test, severe = TRUE),
+    sfga$sfga_severe
+  )
+})
+
+#' @srrstats {G5.2, G5.2a, G5.2b} Tests of message/warning/error behaviour
+#'   in `classify_sfga()`.
+test_that(desc = "classify_sfga() throws appropriate errors/warnings", {
+  testthat::expect_error(
+    classify_sfga(
+      .data = test_data_birth,
+      weight_kg = weight_kg,
+      gest_days = gest_age,
+      sex = sex,
+      .new = c("bweight_centile_exp", "waz_exp", "sfga_severe")
+    ),
+    class = "gigs_classify_.new_in_.data"
+  )
+
+  testthat::expect_warning(
+    classify_sfga(
+      .data = test_data_birth,
+      weight_kg = weight_kg,
+      gest_days = gest_age,
+      sex = sex,
+      .new = c(".3415789.", "sfga", "sfga")
+    ),
+    class = "gigs_repaired_names_in_.new"
+  )
+})
+
+#' @srrstats {G5.8, G5.8a, G5.8b} Testing bad data inputs to `classify_sfga()`.
+test_that(desc = "classify_sfga() errors on bad data types", {
+
+  # Not using a data frame for .data
+  testthat::expect_error(
+    classify_sfga(
+      .data = as.list(test_data_birth),
+      weight_kg = weight_kg,
+      gest_days = gest_age,
+      sex = sex,
+      .new = c("bweight_centile_exp", "waz_exp", "sfga_severe")
+    ),
+    regexp = regex_error_wrong_type(".data", "data.frame", "list")
+  )
+
+  # Using the wrong data types
+  testthat::expect_error(
+    classify_sfga(
+      .data = test_data_birth,
+      weight_kg = as.complex(weight_kg),
+      gest_days = gest_age,
+      sex = sex,
+      .new = c("bweight_centile_test", "sfga", "sfga_severe")
+    ),
+    regexp = regex_error_wrong_type("weight_kg", "numeric", "complex")
+  )
+  testthat::expect_error(
[…]
