### 2024-09-20 0613a23
tests: Added tests for a lot of the growth_compute.R functions --> still adding more; also uncommented many functions used to generate regexps in test code
