+++ tests/testthat/test-growth_classify.R
@@ -73,7 +73,7 @@ test_that(desc = "classify_sfga() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_exp", "waz_exp", "sfga_severe")
     ),
-    regexp = regex_error_wrong_type(".data", "data.frame", "list")
+    regexp = test_error_wrong_type(".data", "data.frame", "list")
   )
 
   # Using the wrong data types
@@ -85,7 +85,7 @@ test_that(desc = "classify_sfga() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_test", "sfga", "sfga_severe")
     ),
-    regexp = regex_error_wrong_type("weight_kg", "numeric", "complex")
+    regexp = test_error_wrong_type("weight_kg", "numeric", "complex")
   )
   testthat::expect_error(
     classify_sfga(
@@ -95,7 +95,7 @@ test_that(desc = "classify_sfga() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_test", "sfga", "sfga_severe")
     ),
-    regexp = regex_error_wrong_type("gest_days", "numeric", "character")
+    regexp = test_error_wrong_type("gest_days", "numeric", "character")
   )
   testthat::expect_error(
     classify_sfga(
@@ -105,7 +105,7 @@ test_that(desc = "classify_sfga() errors on bad data types", {
       sex = sex,
       .new = 1:3
     ),
-    regexp = regex_error_wrong_type(".new", "string", "integer")
+    regexp = test_error_wrong_type(".new", "string", "integer")
   )
   testthat::expect_error(
     classify_sfga(
@@ -182,7 +182,7 @@ test_that(desc = "classify_svn() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_test", "svn")
     ),
-    regexp = regex_error_wrong_type(".data", "data.frame", "list")
+    regexp = test_error_wrong_type(".data", "data.frame", "list")
   )
 
   # Using the wrong data types
@@ -194,7 +194,7 @@ test_that(desc = "classify_svn() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_test", "svn")
     ),
-    regexp = regex_error_wrong_type("weight_kg", "numeric", "complex")
+    regexp = test_error_wrong_type("weight_kg", "numeric", "complex")
   )
   testthat::expect_error(
     classify_svn(
@@ -204,7 +204,7 @@ test_that(desc = "classify_svn() errors on bad data types", {
       sex = sex,
       .new = c("bweight_centile_test", "svn")
     ),
-    regexp = regex_error_wrong_type("gest_days", "numeric", "character")
+    regexp = test_error_wrong_type("gest_days", "numeric", "character")
   )
   testthat::expect_error(
     classify_svn(
@@ -214,7 +214,7 @@ test_that(desc = "classify_svn() errors on bad data types", {
       sex = sex,
       .new = 5:6
     ),
-    regexp = regex_error_wrong_type(".new", "string", "integer")
+    regexp = test_error_wrong_type(".new", "string", "integer")
   )
   testthat::expect_error(
     classify_svn(
@@ -300,7 +300,7 @@ test_that(desc = "classify_stunting() errors on bad data", {
       sex = sex,
       .new = c("lhaz_test", "stunting", "stunting_outliers")
     ),
-    regexp = regex_error_wrong_type(".data", "data.frame", "list")
+    regexp = test_error_wrong_type(".data", "data.frame", "list")
   )
 
   # Using the wrong data types
@@ -313,7 +313,7 @@ test_that(desc = "classify_stunting() errors on bad data", {
       sex = sex,
[…]
