+++ tests/testthat/helper-messages.R
@@ -6,61 +6,70 @@
 #'   type.
 #' @param got Single-length character vector with name of received data
 #'   type.
+#' @returns A single-length character vector which describes an error message
+#'   that gigs should output when inputs are the wrong type.
 #' @description Used to test `{checkmate}` error messages. These errors are
 #'   given by checkmate::assert_*()-style functions in check_params.R.
-regex_error_wrong_type <- function(name, wanted, got) {
-  paste0("Assertion on '", name, "' failed(:|\\.) Must be of (class|type) '", wanted,
-         "', not '", got, "'\\.")
+#' @noRd
+test_error_wrong_type <- function(name, wanted, got) {
+  paste0("Assertion on '", name, "' failed(:|\\.) Must be of (class|type) '",
+         wanted, "', not '", got, "'\\.")
 }
 
-# #' Generate an expected zero-length error from [err_input_is_zero_length]
-# #' @param name Name of input vector.
-# test_error_zero_length <- function(names) {
-#   count <- length(names)
-#   var_str <- if(count > 1) "Variables" else "Variable"
-#   input_str <- if(count > 1) "Inputs" else "Input"
-#   varnames_str <- paste0(names, collapse = "', '")
-#   paste0(var_str, " '", varnames_str, "': ", input_str, " had length 0, ",
-#          "but must have length 1 or greater.")
-# }
-#
-# #' Generate an unrecyclable vectors error from [err_inputs_unrecyclable()].
-# #' @param name Name of input vector.
-# test_error_unrecyclable <- function(names) {
-#   count <- length(names)
-#   var_str <- if(count > 1) "Variables" else "Variable"
-#   input_str <- if(count > 1) "Inputs" else "Input"
-#   varname_str <- paste0(names, collapse = "', '")
-#   paste0(var_str, " '", varname_str, "': ", input_str, " cannot be ",
-#          "recycled with `vctrs\\:\\:vec_recycle_common\\(\\)`.")
-# }
-#
-# # Error/warning string functions for .gigs_options issues ----------------------
-#
-# #' Replicate error/warning messages from gigs for bad input data (TESTING
-# #' ONLY)
-# #' @param name Single-length character vector with a variable name.
-# #' @param length Expected length of the variable specified by `name`.
-# #' @param int_undefined Number of elements in `name` which were expected to
-# #'   be undefined by whatever test is calling this function.
-# #' @note Used for testing only. Mimics the output of [msg_undefined_data()].
-# test_msg_undefined <- function(name, length, int_undefined) {
-#   paste0("Variable '", name, "': ", int_undefined,
-#          " in ", length, " elements were undefined \\(`NaN`,",
-#          " `Inf`, or `-Inf`\\).")
-# }
-#
-# #' Replicate error/warning messages from gigs for bad input data (TESTING
-# #' ONLY)
-# #' @param name Single-length character vector with a variable name.
-# #' @param length Expected length of the variable specified by `name`.
-# #' @param int_undefined Number of elements in `name` which were expected to
-# #'   be missing in whatever test is calling this function.
-# #' @note Used for testing only. Mimics the output of [msg_missing_data()].
-# test_msg_missing <- function(name, length, int_undefined) {
-#   paste0("Variable '", name, "': ", int_undefined,
-#          " in ", length, " elements were missing \\(`NA`\\).")
-# }
+#' Generate an expected zero-length error from [err_input_is_zero_length]
+#' @param names A character vector with variable names that had length zero.
+#' @returns A single-length character vector which describes an error message
+#'   that gigs should output for zero-length inputs.
+#' @noRd
+test_error_zero_length <- function(names) {
[…]

[…]
