### 2024-09-20 212a601
refactor: Added logic to make errors occur if `id` inputs contain `NA` values
