### 2024-09-20 268e618
fix: When GIGS options are set to "error", these bubble up identically to when GIGS options are set to "warn" --> due to including `rlang::zap()` in try_fetch() code
