### 2024-09-20 ddebd75
fix:   (1) When GIGS options are set to "error", these bubble up identically to when GIGS options are set to "warn" --> due to including `rlang::zap()` in try_fetch() code   (2) Errors due to non-recyclability or zero-length inputs now include the bad function call
