+++ R/ig_png.R
@@ -504,7 +504,7 @@ validate_ig_png <- function(y = NULL,
                             x_name = NULL) {
   validate_parameter_lengths(y = y, z = z, p = p, x = x, sex = sex,
                              acronym = acronym, y_name = y_name,
-                             x_name = x_name)
+                             x_name = x_name, call = rlang::caller_env())
   catch_and_throw_validate_issues({
     standard <- "ig_png"
     acronym <- validate_acronym(acronym, names(gigs::ig_png), standard)
