### 2024-09-20 9c820be
feat: Added `id` parameter to appropriate compute_*() functions
