### 2024-09-20 b950fa0
tests: Added compute_wfa() tests
