+++ tests/testthat/test-growth_compute.R
@@ -357,7 +357,7 @@ test_that(desc = "compute_stunting() reproduces expected growth categories", {
 #'   in `compute_stunting()`.
 test_that(desc = "compute_stunting() throws appropriate errors/warnings", {
   expected_length <- nrow(test_data_postnatal)
-  withr::with_seed(seed = 367495, code = {
+  withr::with_seed(seed = 904813, code = {
     n_missing_1 <- sample(1:(expected_length - 5), size = 1)
     n_missing_2 <- sample(1:(expected_length - 5), size = 1)
     n_missing_3 <- sample(1:(expected_length - 5), size = 1)
@@ -713,10 +713,194 @@ test_that(desc = "compute_wasting() errors on bad data types", {
   )
 })
 
+# Testing `compute_wfa()` ------------------------------------------------------
 
+#' @srrstats {G5.4, G5.4c} Tests to ensure that `compute_wfa()` reproduces
+#'   expected appropriate categorical data.
+test_that(desc = "compute_wfa() reproduces expected growth categories", {
+  wfa_test <- with(test_data_postnatal, {
+    compute_wfa(
+      weight_kg = weight_kg,
+      age_days = age_days,
+      gest_days = gest_age,
+      sex = sex,
+      id = id
+    )
+  })
 
+  wfa_out_test <- with(test_data_postnatal, {
+    compute_wfa(
+      weight_kg = weight_kg,
+      age_days = age_days,
+      gest_days = gest_age,
+      sex = sex,
+      id = id,
+      outliers = TRUE
+    )
+  })
 
+  testthat::expect_equal(wfa_test, test_data_postnatal$wfa_exp)
+  testthat::expect_equal(wfa_out_test, test_data_postnatal$wfa_out_exp)
+})
 
+#' @srrstats {G5.2, G5.2a, G5.2b} Tests of message/warning/error behaviour
+#'   in `compute_wfa()`.
+test_that(desc = "compute_wfa() throws appropriate errors/warnings", {
+  expected_length <- nrow(test_data_postnatal)
+  withr::with_seed(seed = 346633, code = {
+    n_missing_1 <- sample(1:(expected_length - 5), size = 1)
+    n_missing_2 <- sample(1:(expected_length - 5), size = 1)
+    n_missing_3 <- sample(1:(expected_length - 5), size = 1)
+    n_missing_4 <- sample(1:(expected_length - 5), size = 1)
+    n_missing_5 <- sample(1:(expected_length - 5), size = 1)
+  })
+
+  # For missing data (i.e. `NA`)
+  testthat::expect_warning(
+    with(test_data_postnatal, {
+      compute_wfa(
+        weight_kg = replace(weight_kg, 1:n_missing_1, values = NA),
+        age_days = age_days,
+        gest_days = gest_age,
+        sex = sex,
+        id = id,
+        outliers = TRUE
+      )
+    }),
+    regexp = test_msg_missing("weight_kg", expected_length, n_missing_1)
+  )
+  testthat::expect_warning(
+    with(test_data_postnatal, {
+      compute_wfa(
+        weight_kg = weight_kg,
+        age_days = age_days,
+        gest_days = gest_age,
+        sex = replace(sex, 1:n_missing_3, values = NA),
+        id = id,
+        outliers = TRUE
+      )
+    }),
+    regexp = test_msg_missing("sex", expected_length, n_missing_3)
+  )
+
+  # For ID, including `NA` values causes an error.
+  testthat::expect_error(
+    with(test_data_postnatal, {
+      compute_wfa(
+        weight_kg = weight_kg,
+        age_days = age_days,
+        gest_days = gest_age,
+        sex = sex,
+        id = replace(id, 1:n_missing_2, values = NA),
[…]
