### 2024-09-20 13653b5
tests: Added compute_headsize() tests
