### 2024-09-20 7d291c2
tests: Added test to ensure 100% coverage of `repair_.new_names()`
