+++ tests/testthat/test-ig_fet_estimation.R
@@ -62,444 +62,430 @@ test_that(desc = "GA estimation function tolerates some noise", {
 })
 
 # Appropriate errors for zero-length data + data of wrong type -----------------
-#
-# #' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-# #' @srrstats {G5.8, G5.8a, G5.8b} Including errors for zero-length inputs and
-# #'   incorrect data types.
-# test_that("Bad GA estimation calls give good errors", {
-#   # Non-numeric input errors
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = character(1), headcirc_mm = 250, femurlen_mm = 50),
-#     test_error_wrong_type("crl_mm", "numeric", "character")
-#   )
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = 25, headcirc_mm = complex(1), femurlen_mm = 50),
-#     test_error_wrong_type("headcirc_mm", "numeric", "complex")
-#   )
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = 25, headcirc_mm = 250, femurlen_mm = logical(1)),
-#     test_error_wrong_type("femurlen_mm", "numeric", "logical")
-#   )
-#
-#   # Errors based on input length
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = integer(), headcirc_mm = 250, femurlen_mm = 50),
-#     test_error_zero_length("crl_mm")
-#   )
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = 25, headcirc_mm = double(), femurlen_mm = 50),
-#     test_error_zero_length("headcirc_mm")
-#   )
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = integer(), headcirc_mm = 50, femurlen_mm = numeric()),
-#     test_error_zero_length(c("crl_mm", "femurlen_mm"))
-#   )
-#
-#   # NULL for crl_mm + headcirc_mm inputs
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = NULL, headcirc_mm = NULL, femurlen_mm = 50),
-#     regexp = "At least one of `crl_mm` or `headcirc_mm` must not be `NULL`"
-#   )
-#
-#   # NULL for all inputs
-#   expect_error(
-#     ig_fet_estimate_ga(crl_mm = NULL, headcirc_mm = NULL, femurlen_mm = NULL),
-#     paste0("All inputs were `NULL`. Ensure some inputs are not `NULL`, then ",
-#            "try again.")
-#   )
-# })
-#
-# test_that(
-#   desc = "Bad fetal weight estimation calls give good errors", code = {
-#   # Non-numeric input errors
-#   expect_error(
-#     ig_fet_estimate_fetal_weight(abdocirc_mm = character(1), headcirc_mm = 250),
-#     test_error_wrong_type("abdocirc_mm", "numeric", "character")
-#   )
-#   expect_error(
-#     ig_fet_estimate_fetal_weight(abdocirc_mm = 250, headcirc_mm = logical(1)),
-#     test_error_wrong_type("headcirc_mm", "numeric", "logical")
-#   )
-#
-#   # Errors based on input length
-#   expect_error(
-#     ig_fet_estimate_fetal_weight(abdocirc_mm = numeric(), headcirc_mm = 250),
-#     test_error_zero_length("abdocirc_mm")
-#   )
-#   expect_error(
-#     ig_fet_estimate_fetal_weight(abdocirc_mm = 250, headcirc_mm = double()),
-#     test_error_zero_length("headcirc_mm")
-#   )
-#   expect_error(
-#     ig_fet_estimate_fetal_weight(abdocirc_mm = numeric(), headcirc_mm = double()),
[…]
