### 2024-09-20 7a6c72d
tests: Re-introduced tests of ig_fet conversion function errors/warnings etc. for ropensci
