+++ tests/testthat/helper-messages.R
@@ -25,8 +25,8 @@ test_error_zero_length <- function(names) {
   count <- length(names)
   var_str <- if (count > 1) "Variables" else "Variable"
   input_str <- if (count > 1) "Inputs" else "Input"
-  varnames_str <- paste0(names, collapse = "', '")
-  paste0(var_str, " '", varnames_str, "': ", input_str, " had length 0, ",
+  varnames_str <- paste0(names, collapse = "`, `")
+  paste0(var_str, " `", varnames_str, "`: ", input_str, " had length 0, ",
          "but must have length 1 or greater.")
 }
 
@@ -42,6 +42,16 @@ test_error_unrecyclable <- function(names) {
   varname_str <- paste0(names, collapse = "', '")
   paste0(var_str, " '", varname_str, "': ", input_str, " cannot be ",
          "recycled with `vctrs\\:\\:vec_recycle_common\\(\\)`.")
+}
+
+#' Generate a 'your acronym is too long!' error, like from [validate_acronym()]
+#' @param length The length of the non-scalar acronym data, used to generate
+#'   the error message
+#' @returns A single-length character vector which describes an error message
+#'   that gigs should output for non-scalar acronym inputs.
+test_error_acronym_not_scalar <- function(length) {
+  paste0("Assertion on 'acronym' failed. Must be of length == 1, but has ",
+         "length ", length, ".")
 }
 
 # Error/warning string functions for .gigs_options issues ----------------------
@@ -70,37 +80,37 @@ test_msg_missing <- function(varname, length, int_undefined) {
   paste0("Argument `", varname, "`: ", int_undefined,
          " in ", length, " elements were missing \\(`NA`\\).")
 }
-#
-# #' Replicate error/warning messages from gigs for bad input data (TESTING
-# #' ONLY)
-# #' @param name Single-length character vector with a variable name.
-# #' @param length Expected length of the variable specified by `name`.
-# #' @param int_undefined Number of elements in `p` which were expected to
-# #'   be outside the proper bounds (0 and 1) in whatever test is calling this
-# #'   function.
-# #' @note Used for testing only. Mimics the output of [msg_oob_centiles()].
-# test_msg_centile_oob <- function(length, int_undefined) {
-#   paste0("Variable 'p': ", int_undefined, " in ", length,
-#          " elements were not between 0 and 1.")
-# }
-#
-# #' Replicate error/warning messages from gigs for bad input data (TESTING
-# #' ONLY)
-# #' @param x_name Single-length character vector with the `x` variable name.
-# #' @param length Expected length of the variable specified by `name`.
-# #' @param int_undefined Number of elements in `x` which were expected to
-# #'   be outside the proper bounds (for each standard/acronym) in whatever test
-# #'   is calling this function.
-# #' @param standard Single-length character vector denoting the collection of
-# #'   growth standards in use. Should be one of `"ig_fet"`, `"ig_nbs"`,
-# #'   `"ig_png"`, or `"who_gs"`.
-# #' @note Used for testing only. Mimics the output of [msg_oob_xvar()].
-# test_msg_xvar_oob <- function(x_name, length, int_undefined, standard) {
-#   paste0("Variable '", x_name, "': ", int_undefined, " in ", length,
-#          " elements were out-of-bounds \\(see the '", standard, "' conversion ",
-#          "functions documentation\\).")
-# }
-#
+
+#' Replicate error/warning messages from gigs for bad input data (TESTING
+#' ONLY)
+#' @param name Single-length character vector with a variable name.
+#' @param length Expected length of the variable specified by `name`.
+#' @param int_undefined Number of elements in `p` which were expected to
[…]
