+++ tests/testthat/test-ig_fet.R
@@ -142,703 +142,697 @@ test_that(
 )
 
 # Appropriate errors for zero-length data + data of wrong type -----------------
-#
-# #' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-# #' @srrstats {G5.8, G5.8a, G5.8b} Including errors for zero-length inputs and
-# #'   incorrect data types.
-# test_that(
-#   desc = "Correct errors occur with zero-length/incorrectly typed data:",
-#   code = {
-#     # Error message testing: Iterate through ig_fet functions, demonstrating
-#     #   with passed `expect_error()` tests that all error messages are unique
-#     #   to the data passed in.
-#     ig_fet_acronyms <- names(ig_fet)
-#     acronyms_to_test <- withr::with_seed(150, sample(ig_fet_acronyms, 5))
-#     fn_suffixes <- c("value2zscore", "value2centile", "centile2value",
-#                      "zscore2value")
-#     z <- 0
-#     p <- 0.5
-#     y <- 4
-#     for (fn_suffix in fn_suffixes) {
-#       for (str_acronym in acronyms_to_test) {
-#         fn_name <- paste("ig_fet", str_acronym, fn_suffix, sep = "_")
-#         gigs_fn <- get(fn_name)
-#         arg_x_name <- gigs::ig_fet[[str_acronym]][["x"]]
-#         y_name <- gigs::ig_fet[[str_acronym]][["y"]]
-#
-#         arg_zyp <- switch(fn_suffix, centile2value = p, zscore2value = z, y)
-#         arg_zyp_name <- switch(fn_suffix, centile2value = "p",
-#                                zscore2value = "z", y_name)
-#         arg_x <- gigs::ig_fet[[str_acronym]][[1]][[1]]
-#
-#         # G5.8a --> zero-length data inputs cause errors
-#         expect_error(object = gigs_fn(double(), arg_x),
-#                      regexp = test_error_zero_length(arg_zyp_name))
-#         expect_error(object = gigs_fn(arg_zyp, double()),
-#                      regexp = test_error_zero_length(arg_x_name))
-#         expect_error(object = gigs_fn(double(), double()),
-#                      regexp = test_error_zero_length(c(arg_zyp_name, arg_x_name)))
-#
-#         # G5.8b --> incorrect data types
-#         expect_error(
-#           object = gigs_fn(as.character(arg_zyp), arg_x),
-#           regexp = test_error_wrong_type(arg_zyp_name, "numeric", "character")
-#         )
-#         expect_error(
-#           object = gigs_fn(arg_zyp, TRUE),
-#           regexp = test_error_wrong_type(arg_x_name, "numeric", "logical")
-#         )
-#
-#         # G5.2, --> errors if vector inputs cannot be recycled
-#         expect_error(
-#           object = gigs_fn(rep_len(arg_zyp, 2), arg_x),
-#           regexp = test_error_unrecyclable(names = arg_zyp_name)
-#         )
-#         expect_error(
-#           object = gigs_fn(rep_len(arg_zyp, 3), rep_len(arg_x, 2)),
-#           regexp = test_error_unrecyclable(names = "gest_days")
-#         )
-#       }
-#
-#       # Repeat for standard-agnostic conversion function
-#       fn_name <- paste("ig_fet", fn_suffix, sep = "_")
-#       gigs_fn <- get(fn_name)
-#
-#       arg_zyp_name <- switch(fn_suffix, centile2value = "p",
-#                              zscore2value = "z", "y")
-#       arg_acronym <- str_acronym
-#       # G5.8a --> zero-length data inputs cause errors
-#       expect_error(object = gigs_fn(double(), arg_x, arg_acronym),
-#                    regexp = test_error_zero_length(arg_zyp_name))
[…]
