### 2024-09-20 7be0eae
tests: Added back in much of the ig-nbs testing code, will amend across afternoon
