### 2024-09-20 c8d78b7
tests: Added back in much of the ig-nbs testing code, will add to this across the afternoon
