+++ tests/testthat/test-ig_nbs.R
@@ -719,238 +719,218 @@ test_that(
   }
 )
 
-# #' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-# #' @srrstats{G5.8, G5.8d} Show that gigs can handle data outside the
-# #'   scope of its algorithms by replacing bad inputs with `NA` and giving clear
-# #'   errors.
-# #' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-# #' @srrstats{G5.8, G5.8d} Show that gigs can handle data outside the
-# #'   scope of its algorithms by replacing bad inputs with `NA` and giving clear
-# #'   warnings.
-# test_that(
-#   desc = "Invalid values can be replaced with `NA` with errors",
-#   code = {
-#     # Make gigs stop() with bad data
-#     for (option in names(.gigs_options)) {
-#       gigs_option_set(option, new_value = "error", silent = TRUE)
-#     }
-#
-#     acronyms_to_test <- names(ig_nbs)[5]
-#     fn_suffixes <- c("value2zscore", "value2centile", "centile2value",
-#                      "zscore2value")
-#     z <- 0
-#     p <- 0.5
-#
-#     # Random set of indices to replace with bad data in each case
-#     num_to_replace1 <- 6
-#     withr::with_seed(200, code = {
-#       replace_ints_1 <- sample(1L:20L, size = num_to_replace1)
-#     })
-#     num_to_replace2 <- 5
-#     withr::with_seed(200, code = {
-#       replace_ints_2 <- sample(1L:20L, size = num_to_replace2)
-#     })
-#     replace_ints_3 <- union(replace_ints_1, replace_ints_2)
-#     num_to_replace3 <- length(union(replace_ints_1, replace_ints_2))
-#
-#     for (fn_suffix in fn_suffixes) {
-#       for (str_acronym in acronyms_to_test) {
-#         fn_name <- paste("ig_nbs", str_acronym, fn_suffix, sep = "_")
-#         gigs_fn <- get(fn_name)
-#         x <- gigs::ig_nbs[[str_acronym]][[1]][[1]][[1]]
-#         len_x <- length(x) * 4
-#         arg_x_name <- gigs::ig_nbs[[str_acronym]][["x"]]
-#         y_name <- gigs::ig_nbs[[str_acronym]][["y"]]
-#
-#         arg_zyp <- rep_len(switch(fn_suffix,
-#                                   centile2value = p,
-#                                   zscore2value = z,
-#                                   gigs::ig_nbs[[str_acronym]][[1]][[1]][[3]]),
-#                            len_x)
-#         arg_zyp_name <- switch(fn_suffix,
-#                                centile2value = "p",
-#                                zscore2value = "z",
-#                                y_name)
-#         arg_x <- rep_len(x, len_x)
-#         arg_sex <- withr::with_seed(25, {
-#           sample(c("M", "F"), replace = TRUE, size = len_x)
-#         })
-#
-#         # Bad input 1: Undefined data (NaN, Inf, -Inf) -------------------------
-#         for (undefined_val in c(NaN, Inf, -Inf)) {
-#           ## Replace z/y/p variable and x variable with undefined_val
-#           expect_error(
-#             gigs_fn(replace(arg_zyp, replace_ints_2, undefined_val),
-#                     replace(arg_x, replace_ints_3, undefined_val),
-#                     arg_sex),
-#             test_msg_undefined(arg_zyp_name, len_x, num_to_replace2)
-#           )
-#         }
-#
-#         # Bad input 2: Missing data (NA) ---------------------------------------
-#         ## Replace only x variable with NA
-#         expect_error(
-#           gigs_fn(arg_zyp,
[…]
