### 2024-09-20 f488238
tests: Added back all the commented-out ig_nbs testing code --> checks errors, warnings, and more
