### 2024-09-20 f7c24fc
tests: Added back all the commented-out ig_png and who_gs testing code with some fixes --> checks errors, warnings, and more
