+++ tests/testthat/test-ig_png.R
@@ -103,915 +103,856 @@ test_that(
 
 # Appropriate errors for zero-length data + data of wrong type -----------------
 
-# #' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
-# #' @srrstats {G5.8, G5.8a, G5.8b} Including errors for zero-length inputs and
-# #'   incorrect data types.
-# test_that(
-#   desc = "Correct errors occur when .gigs_options are set to 'quiet':",
-#   code = {
-#     # Error message testing: Iterate through ig_png functions, demonstrating
-#     #   with passed `expect_error()` tests that all error messages are unique
-#     #   to the data passed in.
-#     acronyms_to_test <- names(ig_png)
-#     fn_suffixes <- c("value2zscore", "value2centile", "centile2value",
-#                      "zscore2value")
-#     z <- 0
-#     p <- 0.5
-#     y <- 4
-#     for (fn_suffix in fn_suffixes) {
-#       for (str_acronym in acronyms_to_test) {
-#         fn_name <- paste("ig_png", str_acronym, fn_suffix, sep = "_")
-#         gigs_fn <- get(fn_name)
-#         arg_x_name <- gigs::ig_png[[str_acronym]][["x"]]
-#         y_name <- gigs::ig_png[[str_acronym]][["y"]]
-#
-#         arg_zyp <- switch(fn_suffix, centile2value = p, zscore2value = z, y)
-#         arg_zyp_name <- switch(fn_suffix, centile2value = "p",
-#                                zscore2value = "z", y_name)
-#         arg_x <- gigs::ig_png[[str_acronym]][[1]][[1]][[1]]
-#         arg_sex <- withr::with_seed(25, {
-#           sample(c("M", "F"), replace = TRUE, size = length(arg_x))
-#         })
-#
-#         # G5.8a --> zero-length data inputs cause errors
-#         expect_error(object = gigs_fn(double(), arg_x, arg_sex),
-#                      regexp = test_error_zero_length(arg_zyp_name))
-#         expect_error(object = gigs_fn(arg_zyp, double(), arg_sex),
-#                      regexp = test_error_zero_length(arg_x_name))
-#         expect_error(object = gigs_fn(double(), arg_x, character()),
-#                      regexp = test_error_zero_length(c(arg_zyp_name, "sex")))
-#
-#         # G5.8b --> incorrect data types
-#         expect_error(
-#           object = gigs_fn(as.character(arg_zyp), arg_x, arg_sex),
-#           regexp = test_error_wrong_type(arg_zyp_name, "numeric", "character")
-#         )
-#         expect_error(
-#           object = gigs_fn(arg_zyp, logical(1), arg_sex),
-#           regexp = test_error_wrong_type(arg_x_name, "numeric", "logical")
-#         )
-#         expect_error(
-#           object = gigs_fn(arg_zyp, arg_x, complex(1)),
-#           regexp = test_error_wrong_type("sex", "character", "complex")
-#         )
-#
-#         # G5.2, --> errors if vector inputs cannot be recycled
-#         expect_error(
-#           object = gigs_fn(rep_len(arg_zyp, 2),
-#                            arg_x,
-#                            arg_sex),
-#           regexp = test_error_unrecyclable(names = arg_zyp_name)
-#         )
-#         expect_error(
-#           object = gigs_fn(rep_len(arg_zyp, 3),
-#                            rep_len(arg_x, 2),
-#                            arg_sex),
-#           regexp = test_error_unrecyclable(names = c(arg_zyp_name, arg_x_name))
-#         )
-#         expect_error(
-#           object = gigs_fn(rep(arg_zyp, 3),
-#                            rep_len(arg_x, 3),
[…]
