### 2024-09-20 f10ac51
docs: Fixed up docs for R CMD check + updated CITATION back in
