+++ DESCRIPTION
@@ -35,6 +35,7 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0),
     dplyr,
+    tidyselect,
     units,
     withr
 VignetteBuilder: 

[…]
