### 2024-09-20 3405563
docs: Added minor stuff to README.Rmd
