### 2024-09-20 4a7d934
docs: Added content to SRR tags in gigs_options.R
