### 2024-09-25 d15bf97
tests: Added tests to satisfy rOpenSci standards for making new columns in EDA software
