+++ tests/testthat/test-growth_classify.R
@@ -8,20 +8,37 @@ test_data_postnatal_wlz <- test_data_postnatal[
     !is.na(with(test_data_postnatal, weight_kg_from_wlz)),
 ]
 
+#' @srrstats {EA6.0, EA6.0a, EA6.0b, EA6.0c, EA6.0d} This function checks the
+#' output type, output dimensions, column names, and the types of the new
+#' columns in tabular output.
+check_classify_fn_output <- function(old, new, new_names) {
+  checkmate::expect_data_frame(
+    new, nrows = nrow(old), ncols = ncol(old) + length(new_names)
+  )
+  testthat::expect_equal(colnames(new), c(colnames(old), new_names))
+
+  checkmate::expect_numeric(new[[new_names[1]]]) # First new col is continuous
+  for (name in new_names[2:length(new_names)]) {
+    checkmate::expect_factor(new[[name]]) # All other new cols are factors
+  }
+}
+
 # Testing `classify_sfga()` ----------------------------------------------------
 
 #' @srrstats {G5.4, G5.4c} Tests to ensure that `classify_sfga()` reproduces
 #'   expected z-scores from randomly generated growth data, and the appropriate
 #'   categorical data.
 test_that(desc = "classify_sfga() reproduces expected z-scores", {
+  new_names <- c("bweight_centile_test", "sfga", "sfga_severe")
   sfga <- classify_sfga(
     .data = test_data_birth,
     weight_kg = weight_kg,
     gest_days = gest_age,
     sex = sex,
-    .new = c("bweight_centile_test", "sfga", "sfga_severe")
+    .new = new_names
   )
-  sfga_temp <- dplyr::select(sfga, id, tidyselect::starts_with("bweight"), tidyselect::starts_with("sfga"))
+  check_classify_fn_output(test_data_birth, sfga, new_names)
+
   testthat::expect_equal(
     sfga$bweight_centile_test, sfga$bweight_centile_exp
   )
@@ -127,13 +144,16 @@ test_that(desc = "classify_sfga() errors on bad data types", {
 #'   expected z-scores from randomly generated growth data, and the appropriate
 #'   categorical data.
 test_that(desc = "classify_svn() reproduces expected z-scores", {
+  new_names <- c("bweight_centile_test", "svn")
   svn <- classify_svn(
     .data = test_data_birth,
     weight_kg = weight_kg,
     gest_days = gest_age,
     sex = sex,
-    .new = c("bweight_centile_test", "svn")
+    .new = new_names
   )
+  check_classify_fn_output(test_data_birth, svn, new_names)
+
   testthat::expect_equal(
     svn$bweight_centile_exp, svn$bweight_centile_test
   )
@@ -236,6 +256,7 @@ test_that(desc = "classify_svn() errors on bad data types", {
 #'   expected z-scores from randomly generated growth data, and the appropriate
 #'   categorical data.
 test_that(desc = "classify_stunting() reproduces expected z-scores", {
+  new_names <- c("lhaz_test", "stunting", "stunting_outliers")
   stunting <- classify_stunting(
     .data = test_data[with(test_data, !is.na(lenht_cm)), ],
     lenht_cm = lenht_cm,
@@ -243,8 +264,11 @@ test_that(desc = "classify_stunting() reproduces expected z-scores", {
     age_days = age_days,
     sex = sex,
     id = id,
-    .new = c("lhaz_test", "stunting", "stunting_outliers")
+    .new = new_names
   )
+  check_classify_fn_output(test_data[with(test_data, !is.na(lenht_cm)), ],
+                           stunting, new_names)
+
   testthat::expect_equal(
     stunting$lhaz_test, stunting$lhaz_exp
   )
[…]
