### 2024-09-25 7189c98
docs: Added `sitar` and `zscorer` packages to benchmarking/prior art sections
