+++ vignettes/articles/benchmarking.Rmd
@@ -12,7 +12,12 @@ knitr::opts_chunk$set(
 
 # Set ggplot theming for consistent plots
 ggplot2::theme_set(ggplot2::theme_bw())
-intergrowth21st <- knitr::asis_output(x = "INTERGROWTH-21<sup>st</sup>")
+
+intergrowth21st <- if (knitr::is_html_output()) {
+  knitr::asis_output(x = "INTERGROWTH-21<sup>st</sup>")
+} else {
+  knitr::asis_output(x = "INTERGROWTH-21\textsupercript{st}")
+}
 ```
 
 ```{r srr-tags, eval = FALSE, echo = FALSE}
@@ -26,18 +31,34 @@ load(file = file.path("benchmarking.rda"))
 
 # Rationale
 This short article describes the performance of gigs relative to a
-non-exhaustive group of R and non-R packages designed to implement child growth
-standards.
+non-exhaustive group of R and non-R packages which implement international
+growth standards which are also found in gigs.
 
-| Package name                                                                 | Language | On CRAN? |
-|------------------------------------------------------------------------------|:---------|:---------|
-| [anthro](https://cran.r-project.org/web/packages/anthro/index.html)          | R        | Yes      |
-| [AGD](https://cran.r-project.org/web/packages/AGD/index.html)                | R        | Yes      |
-| [childsds](https://cran.r-project.org/web/packages/childsds/index.html)      | R        | Yes      |
-| [ki-tools/growthstandards](https://www.github.com/ki-tools/growthstandards/) | R        | No       |
-| [zanthro](https://journals.sagepub.com/doi/epdf/10.1177/1536867X1301300211)  | Stata    | No       |
-| [gigs](https://www.github.com/lshtm-gigs/gigs/)                              | R        | No       |
-| [gigs](https://www.github.com/lshtm-gigs/gigs-stata/)                        | Stata    | No       |
+```{r pkgnames, echo = FALSE}
+anthro <- knitr::asis_output(x = "[anthro](https://cran.r-project.org/web/packages/anthro/index.html)")
+AGD <- knitr::asis_output(x = "[AGD](https://cran.r-project.org/web/packages/AGD/index.html)")
+childsds <- knitr::asis_output(x = "[childsds](https://cran.r-project.org/web/packages/childsds/index.html)")
+gigs_r <- knitr::asis_output(x = "[gigs](https://www.github.com/lshtm-gigs/gigs/)")
+growthstans <- knitr::asis_output(x = "[ki-tools/growthstandards](https://www.github.com/ki-tools/growthstandards/)")
+sitar <- knitr::asis_output(x = "[sitar](https://cran.r-project.org/web/packages/sitar/index.html)")
+zscorer <- knitr::asis_output(x = "[zscorer](https://cran.r-project.org/web/packages/zscorer/index.html)")
+gigs_stata <- knitr::asis_output(x = "[gigs](https://www.github.com/lshtm-gigs/gigs-stata/) (Stata)")
+zanthro <- knitr::asis_output(x = "[zanthro](https://journals.sagepub.com/doi/epdf/10.1177/1536867X1301300211) (Stata)")
+intergrowth <- knitr::asis_output(x = "[nutriverse/intergrowth](https://github.com/nutriverse/intergrowth/)")
+```
+
+| Package name    | Language | On CRAN? |
+|-----------------|:---------|:---------|
+| `r gigs_r`      | R        | Yes      |
+| `r anthro`      | R        | Yes      |
+| `r AGD`         | R        | Yes      |
+| `r childsds`    | R        | Yes      |
+| `r sitar`       | R        | Yes      |
+| `r zscorer`     | R        | Yes      |
+| `r growthstans` | R        | No       |
+| `r intergrowth` | R        | No       |
+| `r gigs_stata`  | Stata    | No       |
[…]
