### 2024-09-25 f557eba
docs: Added EA3.0/EA3.1 tag regarding `classify_growth()`
