### 2024-09-26 #626 @chitrams
-   Briefly describe any working relationship you may have (had) with the package authors (or otherwise remove this statement)

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

With the following point:

> The potential reviewer/editor has significantly contributed to a competitor project.

I would like to mention that I have worked on developing the WHO `{anthro}` package in 2019. However, I do not believe I have a conflict of interest as (1) I am not an active developer of `{anthro}`; and (2) I am excited to see the development of `{gigs}` for the field of childhood malnutrition generally.

### Compliance with Standards

-   [x] This package complies with a sufficient number of standards for a silver badge
-   [x] This grade of badge is the same as what the authors wanted to achieve
-   [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

Documentation and testing are excellent for this package, fulfilling many sub-categories: for example 6.1.1 Documentation, 6.1.5 Testing, 6.3.1 Documentation, 6.3.4 Return results/output data (results are consistent with input type), 6.3.6 Testing.

### General Review

-   [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING

Perhaps you can add a CONTRIBUTING document to specify how people could contribute? I know you've added a Code of Conduct at the bottom of your README, but this additional CONTRIBUTING document could be helpful.

-   [x] The documentation is sufficient to enable general use of the package beyond one specific use case

#### Algorithms

I think these algorithms are well-encoded. Conversions I tested were fine.

Tests seem good to me!

#### Package Design

Package design is good for its intended purpose. I have some suggestions that might make this package easier to use below. Feel free to implement these changes or not :)

-   I wonder if it would be annoying for users to type the standard before the conversion?
    -   As an example, there are three beginning with `ig_`. Perhaps it would be better to name the functions to start with `nbs_ig`, `png_ig`, and `fet_ig` instead?
-   In the package's current state, you have to type the whole standard out before the conversion. I think as a user, I would be thinking about the conversion first, for example: `value2zscore`, `value2centile`, `zscore2value`, and `centile2value`, then the standard name can be an option in the function call instead?
    -   Then as a user when I use Tab to auto-complete, this could be more convenient.
    -   The function names as they currently are also just leads to unwieldy function names–they are quite long.
    -   My suggestion would be a bit of work, as you would need to restructure the function organisation, but I believe would make it easier for the user to use these functions.
-   Similar to the point above. Might it be easier to have the function names shorter as `val2z`, `z2val`, and `val2centile` instead? This point might be a bit overkill though.

### Other Comments

A few other minor suggestions:

-   Potential issue within readme.md: exclamation mark emoji is white, may be hard to read on white background
-   Have you considered submitting your software to [Zenodo](https://zenodo.org/) after? It's an open data repository operated by CERN. Submitting your code to Zenodo means you'll [get a persistent digital object identifier (DOI)](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content), which means your code will be more easily cited in academic contexts. [Here](https://help.zenodo.org/docs/get-started/) are some instructions on getting started.

In summary: thank you @simpar1471 for submitting this package, it's fabulous! I can see that a lot of work has gone into this. Congratulations and well done for putting this together.
