### 2024-09-27 #626 @simpar1471
Thank you for your review @chitrams! With respect to your comments:

1. Package design
    - We like your idea of refactoring the functions to make them less unwieldy. This is something we've heard from several others who've looked at the package. I think it's probably time to rip the plaster off.
    - I'm reluctant to rename the INTERGROWTH-21<sup>st</sup> standard acronyms from `"ig_nbs"` to `"nbs_ig"` etc., if they're going to be parameters in overall `value2centile()`/`value2zscore()` etc. functions - mostly to save on extra dev time, as I've moved from the role which was employing me full-time to build the package out. Is that alright if we're implementing your refactoring suggestions? So the interface would look something like this:
     ```
     value2zscore(y = weight_kg , x = age_days, sex = sex, family = "ig_nbs", acronym = "wfga")
     ```
2. Other comments
    - We'll replace the exclamation mark emoji - for now either with the red exclamation mark (❗) or warning sign (⚠️).
    - We weren't considering Zenodo, but we will now that you've suggested it! We'll sort a DOI out after we've addressed your comments.
3. Other things we should note
    - In the [dev branch](https://github.com/lshtm-gigs/gigs/tree/dev) for gigs we have a range of new functions which apply specific growth standards for infants depending on their gestational + chronological ages. These new functions will be included alongside the refactors you've suggested. This branch also has a `CONTRIBUTING.md` file.
    - We've had some issues with `autotest` recently, which are detailed below. Can we resubmit whilst `autotest` is failing?
        - It recognises our "M"/"F" sex parameters as logical whenever we pass in an `"F"` value, I think because without quotes this maps onto `FALSE` in R. We added this an issue in the autotest repo (ropensci-review-tools/autotest/issues/86).
        - `autotest` also fails for some logical parameters which control whether messages are printed or not, given that inverting these from `TRUE` to `FALSE` changes the console output.

Thank you again for taking the time to review the package, we really appreciate you giving the time for this.
