### 2024-09-30 #626 @BroVic
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software have been confirmed.
-   ☒ **Performance:** Any performance claims of the software have been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 12

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

First, allow me to thank the author and other contributors/funders for this package, as it will be very useful to health and other researchers and policymakers. The following are my comments/suggestions for further improvement:

1.  There are some typographical errors in the documentation (READMe.md):
    i). At the beginning of para 2, it should read “**gigs** is to be of use…”.
    ii). Component standards for INTERGROWTH-21^st Newborn Size, first row of field **Description** should read “weight-for-gestational age”.
    iii). Section **Conversion functions &gt; Terminology** currently reads “*The package operates with on measured anthropometric values*”. Use one preposition instead.

2.  Accessibility to non-expert users: The README assumes that the user is intimately familiar with the subject. I suggest an introductory paragraph that would appeal to a non-expert (maybe even lay) audience. Consider it a gentle introduction :)

3.  The author has provided links to related functions within a help file, but I would suggest adding more. For example, the help file for `classify_sfga` might benefit from links to other `classify_*` functions. This will ease navigation for the user. **Alternatively**, where these functions share enough similarity in their signature, a combined help file would be easier for users, as was done for, say, the `ig_fet_*` functions. Key differences can then be highlighted in the *Details* section.

4.  The benchmarking of `{gigs}` with other packages/software is highly commendable. In the benchmarking for running time, please indicate which of the gigs is for R and Stata.

5.  It seems to me that `{gigs}` is very useful as an intermediary in the data analysis pipeline. So, I have some concerns about the issue of **unused factor levels** when it comes to the `classify_*` functions. The main purpose is to classify based on the available data (as can be seen with the use of the `severe` parameter), but what if a user wants to represent unused factor levels downstream? This is a design decision of the author, but even if the outputs remain unchanged, I would recommend a clear mention in the help file

6.  Looking through `NEWS` I noted that API changes were made via PATCH version. Following the [Semantic Versioning](https://semver.org) guideline, I would recommend that such changes be made, at minimum, a MINOR version update. Then, seeing that the author has made changes to parameter names before, I would suggest the deprecation of `lenht_cm` (for length/height) in favour of just `len_cm`, in the future. This could help naive users who are not that familiar with the correct terminology.

7.  In the file `data.R`, change the `@format` tag element for describing the fields of the dataset from `itemize` to `describe` (line 264); this will remove the `R CMD check` NOTE.

8.  Consider adding `{ggsci}` and `{microbenchmark}` to *Suggests* - required for building the *Benchmarking gigs against other software packages* website article; and `{cowplot}` and `{viridis}` for *NTERGROWTH-21st Body Composition Equations*.

9.  Fix the broken link pointing to <https://simpar1471.github.io/gigs/> as specified in `_pkgdown.yml`.

10. Provide alt-text for images the README and vignette generation in the R chunks via the `fig.alt` chunk option (see `vignette(pkgdown::accessibility)`).

11. `goodpractice:gp()` complains about some lines of code that were longer than 80 characters. Although none was longer than 82 characters, the maintainer might consider refactoring the code so as to adhere to the recommended limit.
