### 2024-10-02 #626 @chitrams
Hi @simpar1471, I will reply point-by-point:

>    * We like your idea of refactoring the functions to make them less unwieldy. This is something we've heard from several others who've looked at the package. I think it's probably time to rip the plaster off.
>    * I'm reluctant to rename the INTERGROWTH-21st standard acronyms from `"ig_nbs"` to `"nbs_ig"` etc., if they're going to be parameters in overall `value2centile()`/`value2zscore()` etc. functions - mostly to save on extra dev time, as I've moved from the role which was employing me full-time to build the package out. Is that alright if we're implementing your refactoring suggestions? So the interface would look something like this:
>
>    ```
>    value2zscore(y = weight_kg , x = age_days, sex = sex, family = "ig_nbs", acronym = "wfga")
>    ```

Yep, that is more than fair enough! If the standard acronyms are already a function parameter, I think that would be more than sufficient for usability :)

>    * We'll replace the exclamation mark emoji - for now either with the red exclamation mark (❗) or warning sign (⚠️).

Fantastic! I'd suggest the yellow warning sign as you already have an existing red emoji for the cross. Thank you for taking this into consideration!

>    * We weren't considering Zenodo, but we will now that you've suggested it! We'll sort a DOI out after we've addressed your comments.

Great! Feel free to ask me if you have any questions about this process.

>    * In the [dev branch](https://github.com/lshtm-gigs/gigs/tree/dev) for gigs we have a range of new functions which apply specific growth standards for infants depending on their gestational + chronological ages. These new functions will be included alongside the refactors you've suggested. This branch also has a `CONTRIBUTING.md` file.

Perfect! I had a quick look and this looks great to me. Quick question, I was wondering why the `CONTRIBUTING.md` file is in the `.github` folder?

Not a problem at all, I'm more than happy to help review `gigs`. Again, great work! :)
