### 2024-10-02 #626 @chitrams
Hi @rkillick, could you help with @simpar1471's question from his [comment above](https://github.com/ropensci/software-review/issues/626#issuecomment-2379046356_) please? Thank you! 😄

>    * We've had some issues with `autotest` recently, which are detailed below. Can we resubmit whilst `autotest` is failing?
>
>      * It recognises our "M"/"F" sex parameters as logical whenever we pass in an `"F"` value, I think because without quotes this maps onto `FALSE` in R. We added this an issue in the autotest repo ([`autotest` parses `"M"`/`"F"` in examples as logical, not character ropensci-review-tools/autotest#86](https://github.com/ropensci-review-tools/autotest/issues/86)).
>      * `autotest` also fails for some logical parameters which control whether messages are printed or not, given that inverting these from `TRUE` to `FALSE` changes the console output.
