### 2024-10-02 #626 @rkillick
@simpar1471 I'm not up to date with the `autotest` issues, maybe @ldecicco-USGS knows more as EiC?
My personal opinion is that if automated things are not working, this should not preclude resubmission.  As with CRAN checks, if there is a result that needs explaining then a comment on this should be sufficient.
