### 2024-10-03 #626 @mpadge
@chitrams @simpar1471  autotest is still only "recommended" and not "required", so you are free to ignore any issues there, and proceed regardless. Sorry that I haven't yet responded to your issue in the autotest repo :smile:
