### 2024-10-03 #626 @simpar1471
@BroVic Thanks you the comprehensive review, especially the pointers to issues in the docs + `DESCRIPTION`. These will be fixed in the next version. Do you have an ORCID we can insert in `DESCRIPTION` for your reviewer role?

@chitrams On the `CONTRIBUTING.md` front - I can't quite recall why it's buried there! I think I copied it from the [dplyr](https://github.com/tidyverse/dplyr). Would you recommend I put it somewhere else?

@mpadge No worries! I'll keep chugging on for now in that case.

I'm going to make the changes to the API as discussed above. @rkillick What should I do when I feel I have a version ready to be looked over?
