### 2024-10-04 #626 @chitrams
@simpar1471 in response to your comment:

> On the `CONTRIBUTING.md` front - I can't quite recall why it's buried there! I think I copied it from the [dplyr](https://github.com/tidyverse/dplyr). Would you recommend I put it somewhere else?

I would recommend having it in a `docs/` directory or in the root directory, although opinions vary and this is probably a matter of personal preference 😆 The rOpenSci [dev guide](https://devguide.ropensci.org/maintenance_collaboration.html#contributing-guide) recommends having it in `.github/` or `docs/`.

My reasoning for **not** having it in `.github/` is: I believe it would be more versatile in case a user does not use GitHub. In that instance, a reasonable place someone would look for this documentation is in a `docs/` dir or in the root dir.

Cheers,
Chitra
