### 2024-11-09 2e20ac3
feat(breaking): Removed family/acronym-specific GIGS conversion functions and replaced with `value2zscore()`, `zscore2value()`, `zscore2value()`, and `centile2value()`.
Also performed edits to input-checking functionality, so error/warning messages are formatted better and more consistently.
