+++ tests/testthat/test-ig_png.R
@@ -23,11 +23,11 @@ test_that(desc = "Conversion of z-scores/centiles to values works", {
                              zscore = -3:3)
         xvar <- ref_tbl[[1]]
         sexvar <- if (chr_sex != "female") "M" else "F"
-        fn <- get(paste0("ig_png_", acronym, "_", chr_z_or_p, "2value"))
+        conv_fn <- getFromNamespace(paste0(chr_z_or_p, "2value"), ns = "gigs")
         pkg_tbl <- lapply(X = dbl_z_or_p,
                           FUN = \(zp) {
-                            round(fn(zp, xvar, sexvar),
-                                  digits = png_roundto(acronym))
+                            conv_fn(zp, xvar, sexvar, "ig_png", acronym) |>
+                              round(digits = png_roundto(acronym))
                           }) |>
           do.call(what = cbind) |>
           as.data.frame() |>
@@ -57,12 +57,12 @@ test_that(desc = "Conversion of values to z-scores works", {
         })
         if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
-        fn_stem <- paste0("ig_png_", acronym)
-        fn_zp2val <- get(paste0(fn_stem, "_", chr_z_or_p, "2value"))
-        y_gigs <- fn_zp2val(dbl_z_or_p, xvar, sexvar)
+        fn_zp2val <- getFromNamespace(paste0(chr_z_or_p, "2value"), ns = "gigs")
+        y_gigs <- fn_zp2val(dbl_z_or_p, xvar, sexvar, "ig_png", acronym)
 
-        fn_val2zp <- get(paste0(fn_stem, "_value2", chr_z_or_p))
-        gigs_z_or_p <- fn_val2zp(y_gigs, xvar, sexvar)
+        fn_val2zp <- getFromNamespace(paste0("value2", chr_z_or_p),
+                                      ns = "gigs")
+        gigs_z_or_p <- fn_val2zp(y_gigs, xvar, sexvar, "ig_png", acronym)
 
         expect_equal(gigs_z_or_p, expected = dbl_z_or_p, tolerance = 10e-10)
       }
@@ -87,871 +87,16 @@ test_that(
         })
         if (chr_z_or_p == "centile") dbl_z_or_p <- pnorm(dbl_z_or_p)
 
-        fn_stem <- paste0("ig_png_", acronym)
-        fn_zp2val <- get(paste0(fn_stem, "_", chr_z_or_p, "2value"))
-        y_gigs <- fn_zp2val(dbl_z_or_p, xvar, sexvar)
+        fn_zp2val <- getFromNamespace(paste0(chr_z_or_p, "2value"), ns = "gigs")
+        y_gigs <- fn_zp2val(dbl_z_or_p, xvar, sexvar, "ig_png", acronym)
 
-        fn_val2zp <- get(paste0(fn_stem, "_value2", chr_z_or_p))
-        gigs_z_or_p <- fn_val2zp(y_gigs + .Machine$double.eps, xvar, sexvar)
+        fn_val2zp <- getFromNamespace(paste0("value2", chr_z_or_p),
+                                      ns = "gigs")
+        gigs_z_or_p <- fn_val2zp(y_gigs + .Machine$double.eps,
+                                 xvar, sexvar, "ig_png", acronym)
 
         expect_equal(gigs_z_or_p, expected = dbl_z_or_p,
                      tolerance = sqrt(.Machine$double.eps))
-      }
-    }
-  }
-)
-
-# Appropriate errors for zero-length data + data of wrong type -----------------
-
-#' @srrstats {G5.2, G5.2a, G5.2b} Explicit tests of error and warning behaviour.
[…]
