### 2024-11-09 57e0670
feat(breaking): The classify_growth() functions and others which use `gigs_zcoring_lgls()` now issue helpful warnings based on the supplied data; information on these warnings is in the docs.
