+++ R/gigs_zscoring.R
@@ -41,6 +41,60 @@
 #'   each individual is used as a birth measure. Leave this argument as `NULL`
 #'   if all your data comes from the same individual. Default = `NULL`.
 #' @inheritParams shared_roxygen_params
+#' @examples
+#' gest_days <- c(rep(35 * 7, 3), rep(35 * 7, 3))
+#' age_days <- c(0, 100, 500, 2, 100, 500)
+#' sex <- rep.int(c("M", "F"), c(3, 3))
+#' wt_kg <- c(3, 6, 9, 3, 6, 9)
+#' len_cm <- rep.int(c(52.2, 60.4, 75), 2)
+#' head_cm <- rep.int(c(30, 40, 49), 2)
+#' ids <- factor(rep.int(c("A", "B"), c(3, 3)))
+#'
+#' # Weight-for-age z-score (WAZ)
+#' waz <- gigs_waz(weight_kg = wt_kg,
+#'                 age_days = age_days,
+#'                 gest_days = gest_days,
+#'                 sex = sex,
+#'                 id = ids)
+#' print(waz)
+#'
+#' # Note - if you don't specify 'id' you'll get different results!
+#' waz_no_id <- gigs_waz(weight_kg = wt_kg,
+#'                       age_days = age_days,
+#'                       gest_days = gest_days,
+#'                       sex = sex)
+#' print(waz == waz_no_id)
+#'
+#' # Length/height-for-age z-score (LHAZ)
+#' lhaz <- gigs_lhaz(lenht_cm = len_cm,
+#'                   age_days = age_days,
+#'                   gest_days = gest_days,
+#'                   sex = sex,
+#'                   id = ids)
+#' print(lhaz)
+#'
+#' # Weight-for-length/height z-score (WLZ)
+#' #   Note: There's no at-birth standards for weight-for-length, so elements 1
+#' #   and 4 have `NA` values
+#' wlz <- gigs_wlz(weight_kg = wt_kg,
+#'                 lenht_cm = len_cm,
+#'                 age_days = age_days,
+#'                 gest_days = gest_days,
+#'                 sex = sex,
+#'                 id = ids)
+#' print(wlz)
+#'
+#' # Head circumference-for-age z-score (HCAZ)
+#' hcaz <- gigs_hcaz(headcirc_cm = head_cm,
+#'                   age_days = age_days,
+#'                   gest_days = gest_days,
+#'                   sex = sex,
+#'                   id = ids)
+#' print(hcaz)
+#' @returns A numeric vector of z-scores, derived using the appropriate growth
+#'   standard for each element-wise combination of `gest_days` and `age_days`.
+#'   For `gigs_wlz()`, all birth WLZs will be missing (`NA`), as there is no
+#'   INTERGROWTH-21<sup>st<sup> Newborn standard for weight-for-length.
 #' @note These functions expect vectors which are recyclable with
 #'   [vctrs::vec_recycle_common()].
 #' @rdname gigs_zscoring
@@ -282,7 +336,7 @@ gigs_hcaz_internal <- function(headcirc_cm,
 #'   from that named list should be used.
 #' @rdname gigs_xaz
 #' @noRd
-gigs_zscoring_lgls <- function(age_days, gest_days, id) {
+gigs_zscoring_lgls <- function(age_days, gest_days, id = NULL) {
   len_age_days <- length(age_days)
   len_gest_days <- length(gest_days)
   if (len_age_days != len_gest_days) {
@@ -291,12 +345,22 @@ gigs_zscoring_lgls <- function(age_days, gest_days, id) {
                   "!" = paste0("`age_days` had length ", len_age_days,
                                "; `gest_days` had length ", len_gest_days,
                                ".")),
[…]
