### 2024-11-09 34c214f
tests: Updated test suite to reflect package refactoring.
