### 2024-11-09 cfccb88
docs: Updates to benchmarking and body-comp articles. In benchmarking article, there's now data on more packages recommended by @stefvanbuuren during ISGA 2024.
