+++ vignettes/articles/benchmarking.Rmd
@@ -84,7 +84,7 @@ no <- if (knitr::is_html_output()) {
   knitr::asis_output(x = "✘")
 }
 half <- if (knitr::is_html_output()) {
-  knitr::asis_output(x = "❕")
+  knitr::asis_output(x = "⚠️")
 } else {
   knitr::asis_output(x = "⚠")
 }
@@ -100,8 +100,8 @@ Growth standards, `r intergrowth21st` Newborn Size standards (including Very
 Preterm), and the `r intergrowth21st` Postnatal Growth standards for preterm
 infants. A tick (`r yes`) indicates that a package contains functions for all
 growth standards within a set, a cross (`r no`) indicates that a package has no
-functions for a set of growth standards, and an exclamation mark/warning sign
-(`r half`) indicates that some standards in a set are implemented, but not others.
+functions for a set of growth standards, and a warning sign (`r half`) indicates
+that some standards in a set are implemented, but not others.
 
 ```{r arrows, echo = FALSE}
 ig21st <- if (knitr::is_html_output()) {
@@ -130,7 +130,7 @@ harr <- knitr::asis_output(x = "&harr;")
 
 # Shared benchmark code
 
-```{r load_gigs, eval = FALSE}
+```{r load_gigs}
 library(gigs)
 ```
 
@@ -150,9 +150,9 @@ expected measurements for each observation. This dataset can then be used to
 compare speeds for each package. These datasets are generated with the following
 function:
 ```{r benchmark_dataset, echo = TRUE}
-benchmark_dataset <- function(n, standard, acronym) {
-  data <- rlang::eval_bare(rlang::parse_expr(paste0("gigs::", standard)))
-  if (standard == "ig_fet") {
+benchmark_dataset <- function(n, family, acronym) {
+  data <- get(family, envir = asNamespace("gigs"))
+  if (family == "ig_fet") {
     xvars <- data[[acronym]][[1]][[1]]
   } else {
     xvars <- data[[acronym]][[1]][[1]][[1]]
@@ -172,19 +172,16 @@ benchmark_dataset <- function(n, standard, acronym) {
       sex <- sample(c("M", "F"), size = n, replace = TRUE)
     }
   )
-  gigs_fn <- rlang::eval_bare(
-    rlang::parse_expr(paste0("gigs::", paste0(standard, "_zscore2value")))
-  )
-  out <- if (standard != "ig_fet") {
-    if (standard == "ig_nbs") {
-      list(z = z, gest_days = x, sex = sex, acronym = acronym)
-    } else {
-      list(z = z, x = x, sex = sex, acronym = acronym)
-    }
+  gigs_fn <- get("zscore2value", envir = asNamespace("gigs"))
+  out <- if (family != "ig_fet") {
+    list(z = z, x = x, sex = sex, family = family, acronym = acronym)
   } else {
-    list(z = z, x = x, acronym = acronym)
+    list(z = z, x = x, family = family, acronym = acronym)
   }
   out$y <- do.call(what = gigs_fn, args = out)
+  if (family == "ig_fet") {
+    out$sex <- NA
+  }
   as.data.frame(out, stringsAsFactors = FALSE)
 }
 ```
@@ -252,7 +249,7 @@ in a variety of other packages.
 # Generate 100,000-row dataset
 ac_who_gs <- "wfa"
 ds_who_gs <- benchmark_dataset(n = 100000,
-                               standard = "who_gs",
+                               family = "who_gs",
                                acronym = ac_who_gs)
 ```
 
@@ -275,21 +272,18 @@ temp_bp_who_gs_most <- bench::press(
     sex <- ds_who_gs$sex[selector]
[…]
