### 2024-11-09 36793fd
docs: Adding @returns tag for gigs_options.R
