### 2024-11-09 67bacf3
docs: Update introductory article based on GIGS package refactor
