### 2024-11-09 5c0dbb3
docs: Update NEWS, DESCRIPTION and README based on package refactoring
