+++ README.Rmd
@@ -163,15 +163,15 @@ international growth standards from the WHO and `r intergrowth21st` project:
 * `ig_nbs` - `r intergrowth21st` Newborn Size standards (including very preterm)
   <details><summary>Component standards</summary>
 
-  | Acronym  | Description                   | Unit  | `gest_days` range |
-  |----------|-------------------------------|-------|-------------------|
-  | `wfga`   | weight-for-GA                 | kg    | 168 to 300 days   |
-  | `lfga`   | length-for-GA                 | cm    | 168 to 300 days   |
-  | `hcfga`  | head circumference-for-GA     | cm    | 168 to 300 days   |
-  | `wlrfga` | weight-to-length ratio-for-GA | kg/cm | 168 to 300 days   |
-  | `ffmfga` | fat-free mass-for-GA          | kg    | 266 to 294 days   |
-  | `bfpfga` | body fat percentage-for-GA    | %     | 266 to 294 days   |
-  | `fmfga`  | fat mass-for-GA               | kg    | 266 to 294 days   |
+  | Acronym  | Description                   | Unit  | `x` range       |
+  |----------|-------------------------------|-------|-----------------|
+  | `wfga`   | weight-for-GA                 | kg    | 168 to 300 days |
+  | `lfga`   | length-for-GA                 | cm    | 168 to 300 days |
+  | `hcfga`  | head circumference-for-GA     | cm    | 168 to 300 days |
+  | `wlrfga` | weight-to-length ratio-for-GA | kg/cm | 168 to 300 days |
+  | `ffmfga` | fat-free mass-for-GA          | kg    | 266 to 294 days |
+  | `bfpfga` | body fat percentage-for-GA    | %     | 266 to 294 days |
+  | `fmfga`  | fat mass-for-GA               | kg    | 266 to 294 days |
 
   </details>
 * `ig_png` - `r intergrowth21st` Postnatal Growth of Preterm Infants standards
@@ -232,29 +232,36 @@ international growth standards from the WHO and `r intergrowth21st` project:
 
 ### Usage
 
-Conversion functions are named according to the set of standards in use, the
-component standard from that set, then the type of conversion. For example, to
-convert *values to z-scores* in the *weight-for-GA* standard from the
-*`r intergrowth21st` Newborn Size Standards* would be:
-  `ig_nbs`/`_wfga`/`_value2zscore()`
+Conversion functions are named according to the conversion they perform. Either
+they convert measured values to z-scores/centiles
+(`value2zscore()`/`value2centile()`), or they generate expected values for given
+z-scores/centiles (`zscore2value()`/`centile2value()`).
+
+You tell gigs which international growth standard to use with the `family` and
+`acronym` parameters. The `family` parameter tells gigs which set of standards
+you want to use - e.g. `"ig_nbs"` for the `r intergrowth21st` Newborn Size
+standards (including very preterm). The `acronym` parameter describes which
+exact growth standard you want out of all the growth standards in your 'family'
+of standards.
+
+For example, to convert *values to z-scores* in the *weight-for-GA* standard
+from the *`r intergrowth21st` Newborn Size standards*, you would run
+`value2zscore(..., family = "ig_nbs", acronym = "wfga")`.
 
 Similarly, the conversion of length-for-age values to centiles in term and
-preterm infants could be performed with the WHO Child Growth standards and
-`r intergrowth21st` Postnatal Growth of Preterm Infants standards,
+preterm infants could be performed with the *WHO Child Growth standards* and
+*`r intergrowth21st` Postnatal Growth of Preterm Infants standards*,
 respectively:
 
-* Term infants:    `who_gs`/`_lhfa`/`_value2zscore()`
-* Preterm infants: `ig_png`/`_lfa`/`_value2centile()`
[…]
