+++ README.md
@@ -102,7 +102,8 @@ life6mo_classified <- classify_growth(
         id = as.factor(id),
         .outcomes = c("svn", "stunting")
 )
-#> `gigs::classify_growth()`
+#> 
+#> ── `gigs::classify_growth()` ───────────────────────────────────────────────────
 #> ✔ Small vulnerable newborns: Success
 #> ✔ Stunting: Success
 
@@ -134,7 +135,8 @@ life6mo_classified <- classify_growth(
         id = as.factor(id),
         .outcomes = c("svn", "stunting")
 )
-#> `gigs::classify_growth()`
+#> 
+#> ── `gigs::classify_growth()` ───────────────────────────────────────────────────
 #> ✔ Small vulnerable newborns: Success
 #> ! Stunting: Not computed (`lenht_cm` not supplied)
 
@@ -196,15 +198,15 @@ INTERGROWTH-21<sup>st</sup> project:
   Component standards
   </summary>
 
-  | Acronym  | Description                   | Unit  | `gest_days` range |
-  |----------|-------------------------------|-------|-------------------|
-  | `wfga`   | weight-for-GA                 | kg    | 168 to 300 days   |
-  | `lfga`   | length-for-GA                 | cm    | 168 to 300 days   |
-  | `hcfga`  | head circumference-for-GA     | cm    | 168 to 300 days   |
-  | `wlrfga` | weight-to-length ratio-for-GA | kg/cm | 168 to 300 days   |
-  | `ffmfga` | fat-free mass-for-GA          | kg    | 266 to 294 days   |
-  | `bfpfga` | body fat percentage-for-GA    | %     | 266 to 294 days   |
-  | `fmfga`  | fat mass-for-GA               | kg    | 266 to 294 days   |
+  | Acronym  | Description                   | Unit  | `x` range       |
+  |----------|-------------------------------|-------|-----------------|
+  | `wfga`   | weight-for-GA                 | kg    | 168 to 300 days |
+  | `lfga`   | length-for-GA                 | cm    | 168 to 300 days |
+  | `hcfga`  | head circumference-for-GA     | cm    | 168 to 300 days |
+  | `wlrfga` | weight-to-length ratio-for-GA | kg/cm | 168 to 300 days |
+  | `ffmfga` | fat-free mass-for-GA          | kg    | 266 to 294 days |
+  | `bfpfga` | body fat percentage-for-GA    | %     | 266 to 294 days |
+  | `fmfga`  | fat mass-for-GA               | kg    | 266 to 294 days |
 
   </details>
 - `ig_png` - INTERGROWTH-21<sup>st</sup> Postnatal Growth of Preterm
@@ -275,33 +277,42 @@ INTERGROWTH-21<sup>st</sup> project:
 
 ### Usage
 
-Conversion functions are named according to the set of standards in use,
-the component standard from that set, then the type of conversion. For
-example, to convert *values to z-scores* in the *weight-for-GA* standard
-from the *INTERGROWTH-21<sup>st</sup> Newborn Size Standards* would be:
-`ig_nbs`/`_wfga`/`_value2zscore()`
+Conversion functions are named according to the conversion they perform.
+Either they convert measured values to z-scores/centiles
+(`value2zscore()`/`value2centile()`), or they generate expected values
+for given z-scores/centiles (`zscore2value()`/`centile2value()`).
+
+You tell gigs which international growth standard to use with the
+`family` and `acronym` parameters. The `family` parameter tells gigs
+which set of standards you want to use - e.g. `"ig_nbs"` for the
+INTERGROWTH-21<sup>st</sup> Newborn Size standards (including very
+preterm). The `acronym` parameter describes which exact growth standard
[…]
