### 2024-11-09 ebd55ed
docs: Re-generate all roxygen docs
