### 2024-11-09 d7bd17c
chore: Update test-coverage to include install for autotest package
