+++ .github/workflows/test-coverage.yaml
@@ -21,6 +21,11 @@ jobs:
         with:
           use-public-rspm: true
 
+      - name: Install gyro from Github
+        run: |
+          options(repos = c(ropenscireviewtools = "https://ropensci-review-tools.r-universe.dev", CRAN = "https://cloud.r-project.org"))
+          install.packages ("autotest")
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::covr
