### 2024-11-09 1fe37c3
chore: Update test-coverage to include install for autotest package
