### 2024-11-09 91a3d73
chore: Update test-coverage to include install for autotest package
