### 2024-11-09 993c542
chore: Add github package spec for autotest to CI YAML
