+++ .github/workflows/test-coverage.yaml
@@ -21,13 +21,11 @@ jobs:
         with:
           use-public-rspm: true
 
-      - name: Install autotest from Github
-        run: |
-          Rscript -e "pak::pkg_install('ropensci-review-tools/autotest')"
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::covr
+          extra-packages: |
+            github::ropensci-review-tools/autotest
+            any::covr
           needs: coverage
 
       - name: Test coverage
