### 2024-11-09 5670dcb
chore: Add growth analysis packages to pkgdown CI so that packageVersion() doesn't bork out when making plots
