### 2024-11-09 f647a55
chore: Add autotest download step to pkgcheck + cffr CI
