### 2024-11-09 108c079
chore: Drop autotest from suggests, due to issues with CI
