+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/benchmarking: benchmarking.html
   gigs: gigs.html
   articles/ig-nbs-body-composition: ig-nbs-body-composition.html
-last_built: 2024-11-09T03:05Z
+last_built: 2024-11-09T03:12Z
 urls:
   reference: https://lshtm-gigs.github.io/gigs/reference
   article: https://lshtm-gigs.github.io/gigs/articles
