### 2024-11-09 12c88db
chore: Drop autotest installation from pkgcheck CI
