### 2024-11-09 825c877
docs: Replace leq/geq symbols with '=<' and '>=' to prevent LaTeX errors
