### 2024-11-09 c451a25
chore: Add extra packages to 'suggests' given they're used in articles
