### 2024-11-09 15fa1c1
chore: Add some extra context to README.md
