### 2024-11-09 3ec893e
docs: Add notes about unused factor levels and some extra `@seealso` linkages
