### 2024-11-09 #626 @simpar1471
Hi folks! It's been a bit of while since the changes you asked for required some pretty hefty refactoring of the source code + tests, but I've updated the package to reflect your comments, and further updated it to have some nicer formatting in warnings/errors with `cli`.

# General response
## Package design
The package has been refactored considerably, with growth standard agnostic `value2score()`, `value2centile()`, `centile2value()` and `zscore2value()` functions from which users pick a growth standard with `family` and `acronym` parameters. For example, the INTERGROWTH-21<sup>st</sup> newborn size standard for weight-for-gestational age is now accessed with `value2zscore(y = ..., x = ...,  sex = ..., family = "ig_nbs", acronym = "wfga")`. This was updated in commit https://github.com/lshtm-gigs/gigs/commit/2e20ac39a57dbf1618965ca23bf4689f9312611b.

# Response to @chitrams
## Documentation
- [x]  Community guidelines including contribution guidelines in the README or CONTRIBUTING
- **Response:** Added in https://github.com/lshtm-gigs/gigs/commit/c02b7f14a1132bc8d928d62780c90f7ea8a1aac7
## Other comments
- Potential issue within README.md: exclamation mark emoji (❕) is white, this may be hard to read on a white background.
- **Response:** Changed to a warning symbol (⚠) in https://github.com/lshtm-gigs/gigs/commit/5c0dbb303bf23c98bf4fa27f962bfa8788e6a6b1

# Response to @BroVic
##  Review comments
1. README issues
**Response:** The README has been updated to not have the misakes you pointed out.

2. Accessibility to non-expert users
**Response:** I've added some more context in commit https://github.com/lshtm-gigs/gigs/commit/15fa1c1e9756fd55e471dd35dad0a727a481c5b5 - is this gentle enough, or should I change the tone further?

3. Help file linking
**Response:** Added some more links in https://github.com/lshtm-gigs/gigs/commit/3ec893ea1bde54544b7985c04fd33b96d9d8129c. Are more still needed?

4. Benchmarking
**Response:** The benchmarking file has been updated generally; is the new format okay? This is mostly in commit https://github.com/lshtm-gigs/gigs/commit/752c21f9738422d5e5e7b58403bc19e3d206f242, but you can see the results best on the [website](https://lshtm-gigs.github.io/gigs/articles/benchmarking.html).

5. **Unused factor levels** in `classify_()`/`compute_()`/`categorise_*()` outputs.
**Response:** I have added `@notes` specifying that unused factor levels will be present in `classify_()`/`compute_()`/`categorise_*()` outputs (commit ) in https://github.com/lshtm-gigs/gigs/commit/3ec893ea1bde54544b7985c04fd33b96d9d8129c. Are these sufficient?

7. Semantic versioning considerations
**Response:** Acknowledged + I'll make sure these are at least a MINOR version bump in the future. As for the `lenht_cm` parameter is specifically used because length and height in anthropometry are measured differently, and we want users to know they can use height values here too.

8. Change `\itemize` to `\describe`
**Response:** Done in commit https://github.com/lshtm-gigs/gigs/commit/c02b7f14a1132bc8d928d62780c90f7ea8a1aac7

9. Adding items to _Suggest_ in `DESCRIPTION`
**Response:** Done in commit https://github.com/lshtm-gigs/gigs/commit/c451a250a79daf17c25407604d5a1e5b7a3262fe

10. Fix the broken link pointing to https://simpar1471.github.io/gigs/ as specified in _pkgdown.yml.
**Response:** Done recently, contained the PR merge at https://github.com/lshtm-gigs/gigs/commit/752c21f9738422d5e5e7b58403bc19e3d206f242

11. Provide alt-text for images the README and vignette generation...
**Response:** Done recently, contained the PR merge at https://github.com/lshtm-gigs/gigs/commit/752c21f9738422d5e5e7b58403bc19e3d206f242

12. Edit all lines of code down to <80 characters based on `goodpractice` output
**Response:** I've not done this given the breadth of other changes required to get the package ready to submit. Some lines of the roxygen documentation are also >80 characters because I'm writing out Markdown tables, which I don't think I can make any narrower.
