### 2024-11-09 0afe6e2
Update README.Rmd
Add rOpenSci badge to README.Rmd
