### 2024-11-09 #626 @chitrams
@simpar1471 fantastic work refactoring the functions to be standard-agnostic! I'll have a go testing these in the next week or so, I'm looking forward to trying out the changes :)

Well done on implementing the changes and for the excellent documentation in responding to our suggestions and comments! I really appreciate the organisation.
