### 2024-11-10 #626 @BroVic
@simpar1471

Thanks for your detailed report on the changes made! And I think the change made for would-be non-expert users is adequate, as long as they RTFM 😊. A suggestion for future versions just came to mind - detect unused factor levels and signal a warning.

Cheers!
