### 2024-11-10 #626 @simpar1471
@chitrams Thanks for the quick feedback!

@BroVic I'll add in something for warning users about unused factor levels - should have it implemented soon.
