### 2024-11-11 35c606d
feat: Added .gigs_options, which controls how missing levels are dealt with by gigs following growth outcome categorisation
