### 2024-11-11 18762f0
chore: Update CITATION.cff and codemeta.json
