+++ CITATION.cff
@@ -8,7 +8,7 @@ message: 'To cite package "gigs" in publications use:'
 type: software
 license: GPL-3.0-or-later
 title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
-version: 0.4.1.9001
+version: 0.4.1.9002
 identifiers:
 - type: url
   value: https://lshtm-gigs.github.io/gigs/
