### 2024-11-11 96beb45
fix: Fixed unused factor level handling (for SVN categorisation, specifically); added some unit tests too
