+++ tests/testthat/test-growth_categorise.R
@@ -17,124 +17,162 @@ test_that("Size-for-GA categorisation works as expected", {
     levels = c("SGA(<3)", sfga_lvls)
   )
 
-    expect_equal(sfga, sfga_expected)
-    expect_equal(sfga_severe, sfga_severe_expected)
+  expect_equal(sfga, sfga_expected)
+  expect_equal(sfga_severe, sfga_severe_expected)
 })
 
-test_that(
-  desc = "SVN categorisation works as expected",
-  code = {
-    centile <- c(0.01, 0.029, 0.031, 0.09, 0.10, 0.11, 0.89, 0.91, 0.999)
-    gest_days <- seq(250, 290, 5)
-    svn <- categorise_svn(centile, gest_days = gest_days)
-    
-    svn_lvls <- paste(c(rep_len("Preterm", 3), rep_len("Term", 3)),
-                      rep.int(c("SGA", "AGA", "LGA"), times = 2))
-    svn_chr <- paste0(
-      c(rep.int("Preterm ", 2), rep.int("Term ", 7)),
-      c("SGA", "SGA", "SGA", "SGA", "AGA", "AGA", "AGA", "LGA", "LGA")
-    )
-    expect_equal(svn, expected = factor(svn_chr, levels = svn_lvls))
+test_that("SVN categorisation works as expected", {
+  centile <- c(0.01, 0.029, 0.031, 0.09, 0.10, 0.11, 0.89, 0.91, 0.999)
+  gest_days <- seq(250, 290, 5)
+  svn <- categorise_svn(centile, gest_days = gest_days)
+
+  svn_lvls <- paste(c(rep_len("Preterm", 3), rep_len("Term", 3)),
+                    rep.int(c("SGA", "AGA", "LGA"), times = 2))
+  svn_chr <- paste0(
+    c(rep.int("Preterm ", 2), rep.int("Term ", 7)),
+    c("SGA", "SGA", "SGA", "SGA", "AGA", "AGA", "AGA", "LGA", "LGA")
+  )
+  expect_equal(svn, expected = factor(svn_chr, levels = svn_lvls))
+})
+
+test_that("Stunting categorisation works as expected", {
+  lhaz <- c(-6.5, -6, -3.1, -3, -2.9, -2, -1.9, -0.5, -.5, 3, 4.999, 6, 6.5)
+  stunting_lvls <- c("stunting_severe", "stunting", "not_stunting")
+  stunting_outlier_lvls <- c(stunting_lvls, "outlier")
+
+  stunting_expected <- factor(
+    c(rep(stunting_lvls[1], 4),
+      rep(stunting_lvls[2], 2),
+      rep(stunting_lvls[3], 7)), levels = stunting_lvls)
+  stunting_outliers_expected <- factor(
+    c(rep(stunting_outlier_lvls[4], 1),
+      rep(stunting_outlier_lvls[1], 3),
+      rep(stunting_outlier_lvls[2], 2),
+      rep(stunting_outlier_lvls[3], 6),
+      rep(stunting_outlier_lvls[4], 1)), levels = stunting_outlier_lvls)
+
+
+  stunting <- categorise_stunting(lhaz, outliers = FALSE)
+  stunting_outliers <- categorise_stunting(lhaz, outliers = TRUE)
+  expect_equal(stunting, stunting_expected)
+  expect_equal(stunting_outliers, stunting_outliers_expected)
+})
+
+test_that("Wasting categorisation works as expected", {
+  wlz <- c(-5.5, -5, -3.1, -3, -2.9, -2, -1.9, -0.5, -1.9, 2, 4.999, 5, 5.5)
[…]
