### 2024-11-11 #626 @simpar1471
@BroVic I've added some package-level options for handling unused factor levels. By default users will be warned about unused factor levels, but they'll remain the in the outputs. They can change this using `gigs_option_set()` to make it so that the package stops warning them, or starts dropping the unused levels. This is all in commits https://github.com/lshtm-gigs/gigs/commit/35c606d2a3ef8978b7fd36e6e43a34d701c800b0 to https://github.com/lshtm-gigs/gigs/commit/96beb4598fe89630040fb60ae1a95fae1165c0df; with corresponding updates to test suite/docs/website included. How are these?
