### 2024-11-15 45de6fe
Add notes about unused factor levels before pull
