### 2024-11-15 8446089
Updated documentation about unused factor levels for all growth computation/categorisation/classification docs
