+++ man/classify_growth.Rd
@@ -108,7 +108,10 @@ considered to be the observation per level of \code{id} where \code{age_days} is
 smallest, provided that \code{age_days} is between \verb{<3}.
 
 Categorical (factor) columns produced here may contain unused factor
-levels; use \code{\link[=droplevels]{droplevels()}} or similar to remove these.
+levels. By default, gigs will inform you if these columns have unused
+factor levels. You can change this behaviour using the
+\link[=gigs_options]{GIGS package-level option}
+\code{.gigs_options$handle_unused_levels}.
 }
 \examples{
 # This dummy dataset contains data from two people, from birth (<3 days) to
