+++ man/classify_wfa.Rd
@@ -83,7 +83,10 @@ Outlier              \tab \code{"outlier"}        \tab \code{waz} < -6 or \code{
 }
 \note{
 Categorical (factor) columns produced here may contain unused factor
-levels; use \code{\link[=droplevels]{droplevels()}} or similar to remove these.
+levels. By default, gigs will inform you if these columns have unused
+factor levels. You can change this behaviour using the
+\link[=gigs_options]{GIGS package-level option}
+\code{.gigs_options$handle_unused_levels}.
 }
 \examples{
 # This dummy dataset contains data from two people, from birth (<3 days) to
