### 2024-11-15 0241631
Fix documentation for GIGS package-level options
