### 2024-11-15 8b16a2a
fix type in README + update CITATION.cff/codemeta.json
