+++ CITATION.cff
@@ -35,7 +35,7 @@ preferred-citation:
   - family-names: Ohuma
     given-names: Eric
   year: '2024'
-  notes: 0.4.1.9001
+  notes: 0.4.1.9002
   url: https://github.com/lshtm-gigs/gigs/
 repository-code: https://github.com/lshtm-gigs/gigs
 url: https://github.com/lshtm-gigs/gigs/
