### 2024-11-18 69caa7f
feat: Added extended INTERGROWTH-21st Newborn Size standards, which have a wider GA range. Also added relevant article describing the process for extending these standards out.
