+++ vignettes/articles/benchmarking.Rmd
@@ -30,9 +30,9 @@ load(file = file.path("benchmarking.rda"))
 ```
 
 # Rationale
-This short article describes the performance of gigs relative to a
+This short article describes the performance of **gigs** relative to a
 non-exhaustive group of R and non-R packages which implement international
-growth standards which are also found in gigs.
+growth standards which are also found in **gigs**.
 
 ```{r pkgnames, echo = FALSE}
 anthro <- knitr::asis_output(x = "[anthro](https://cran.r-project.org/web/packages/anthro/index.html)")
@@ -91,17 +91,21 @@ half <- if (knitr::is_html_output()) {
 ```
 
 Other R packages can be used to analyse growth data with international
-standards, but have limitations which are not present in gigs. There are also
-software packages external to R which implement these standards. The table
+standards, but have limitations which are not present in **gigs**. There are
+also software packages external to R which implement these standards. The table
 below describes these packages, and to what extent they have implemented
 functions that let users convert anthropometric measurements to
-z-scores/centiles in each set of standards implemented in gigs - the WHO Child
-Growth standards, `r intergrowth21st` Newborn Size standards (including Very
-Preterm), and the `r intergrowth21st` Postnatal Growth standards for preterm
-infants. A tick (`r yes`) indicates that a package contains functions for all
-growth standards within a set, a cross (`r no`) indicates that a package has no
-functions for a set of growth standards, and a warning sign (`r half`) indicates
-that some standards in a set are implemented, but not others.
+z-scores/centiles in each set of standards implemented in **gigs** - the WHO
+Child Growth standards, `r intergrowth21st` Newborn Size standards (including
+Very Preterm), and the `r intergrowth21st` Postnatal Growth standards for
+preterm infants. No other packages (except **gigs** for Stata) includes the
+extended versions of the `r intergrowth21st` Newborn Size standards (incl. Very
+Preterm).
+
+A tick (`r yes`) indicates that all possible standards are
+included in a package, a red cross (`r no`) indicates that these standards are
+completely missing, and a warning sign (`r half`) indicates that some of these
+standards are implemented but not others.
 
 ```{r arrows, echo = FALSE}
 ig21st <- if (knitr::is_html_output()) {
@@ -135,7 +139,8 @@ library(gigs)
 ```
 
 In this vignette, we perform benchmark performance in each a growth standard
-from each set of growth standards implemented in gigs. These are the:
+from each set of growth standards implemented in **gigs**, excluding the
+extended `r intergrowth21st` Newborn Size standards. These are the:
 
 * WHO child growth standards (0-5 years)
 * `r intergrowth21st` Newborn Size standards
@@ -144,8 +149,8 @@ from each set of growth standards implemented in gigs. These are the:
 
 The benchmarks utilise pre-generated datasets for growth standard. The z-scores,
 `x` variable and sexes for each row are generated (pseudo)randomly. Because the
-`gigs` test suite validates its functions against published growth curve data
-from the WHO and the `r intergrowth21st` Project, we use `gigs` to generate
+**gigs** test suite validates its functions against published growth curve data
+from the WHO and the `r intergrowth21st` Project, we use **gigs** to generate
 expected measurements for each observation. This dataset can then be used to
 compare speeds for each package. These datasets are generated with the following
[…]
