+++ README.md
@@ -229,6 +229,20 @@ and INTERGROWTH-21<sup>st</sup> project:
   | `fmfga`  | fat mass-for-GA               | kg    | 266 to 294 days |
 
   </details>
+- `ig_nbs_ext` - INTERGROWTH-21<sup>st</sup> Extended Newborn Size
+  standards (including very preterm)
+  <details>
+  <summary>
+  Component standards
+  </summary>
+
+  | Acronym | Description               | Unit | `x` range       |
+  |---------|---------------------------|------|-----------------|
+  | `wfga`  | weight-for-GA             | kg   | 154 to 314 days |
+  | `lfga`  | length-for-GA             | cm   | 154 to 314 days |
+  | `hcfga` | head circumference-for-GA | cm   | 154 to 314 days |
+
+  </details>
 - `ig_png` - INTERGROWTH-21<sup>st</sup> Postnatal Growth of Preterm
   Infants standards
   <details>
@@ -250,29 +264,29 @@ and INTERGROWTH-21<sup>st</sup> project:
   Component standards
   </summary>
 
-  | Acronym | Description | Unit | `x` range |
-  |----|----|----|----|
-  | `hcfga` | head circumference-for-GA | mm | 98 to 280 days |
-  | `bpdfga` | biparietal diameter-for-GA | mm | 98 to 280 days |
-  | `acfga` | abdominal circumference-for-GA | mm | 98 to 280 days |
-  | `flfga` | femur length-for-GA | mm | 98 to 280 days |
-  | `ofdfga` | occipito-frontal diameter for-GA | mm | 98 to 280 days |
-  | `efwfga` | estimated fetal weight-for-GA | g | 154 to 280 days |
-  | `sfhfga` | symphisis-fundal height-for-GA | mm | 112 to 294 days |
-  | `crlfga` | crown-rump length-for-GA | mm | 58 to 105 days |
-  | `gafcrl` | GA-for-crown-rump length | days | 15 to 95 mm |
-  | `gwgfga` | gestational weight gain-for-GA | kg | 98 to 280 days |
-  | `pifga` | pulsatility index-for-GA |  | 168 to 280 days |
-  | `rifga` | resistance index-for-GA |  | 168 to 280 days |
-  | `sdrfga` | systolic/diastolic ratio-for-GA |  | 168 to 280 days |
-  | `tcdfga` | transcerebellar diameter-for-GA | mm | 98 to 280 days |
-  | `tcdfga` | GA-for-transcerebellar diameter | mm | 98 to 280 days |
-  | `poffga` | parietal-occipital fissure-for-GA | mm | 105 to 252 days |
-  | `sffga` | Sylvian fissue-for-GA | mm | 105 to 252 days |
-  | `avfga` | anterior horn of the lateral ventricle-for-GA | mm | 105 to 252 days |
-  | `pvfga` | atrium of the posterior horn of the lateral ventricle-for-GA | mm | 105 to 252 days |
-  | `cmfga` | cisterna magna-for-GA | mm | 105 to 252 days |
-  | `hefwfga` | Hadlock estimated fetal weight-for-GA | g | 126 to 287 days |
+  | Acronym   | Description                                                  | Unit | `x` range       |
+  |-----------|--------------------------------------------------------------|------|-----------------|
+  | `hcfga`   | head circumference-for-GA                                    | mm   | 98 to 280 days  |
+  | `bpdfga`  | biparietal diameter-for-GA                                   | mm   | 98 to 280 days  |
+  | `acfga`   | abdominal circumference-for-GA                               | mm   | 98 to 280 days  |
+  | `flfga`   | femur length-for-GA                                          | mm   | 98 to 280 days  |
[…]
