### 2024-11-18 9f68713
docs: minor changes to align package vignette + articles
