+++ pkgdown.yml
@@ -3,9 +3,10 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   articles/benchmarking: benchmarking.html
+  articles/extending-ig-nbs: extending-ig-nbs.html
   gigs: gigs.html
   articles/ig-nbs-body-composition: ig-nbs-body-composition.html
-last_built: 2024-11-15T22:39Z
+last_built: 2024-11-18T10:55Z
 urls:
   reference: https://lshtm-gigs.github.io/gigs/reference
   article: https://lshtm-gigs.github.io/gigs/articles
