### 2024-11-19 #626 @chitrams
Hi all,

I've reviewed the changes and think they look great!

@simpar1471 the package documentation is excellent! I liked that in each help file, the limits for each combination for `x` values are provided. I also really like the `report_units` function.

Well done on refactoring the code to use `zscore2value`, `value2zscore`, `value2centile` etc functions instead of the growth standard name. I appreciate this would have taken a lot of work, but I do believe that this would really help make the package user-friendly.

@rkillick I would like to recommend this software to be awarded a gold badge. This package:

- Complies with a good number (more than one half) of standards beyond those identified as minimally necessary
- Demonstrates excellence in compliance with multiple standards. Generally, I think this package is excellently documented and has a fantastic test suite. Specifically, these sub-categories were fulfilled: G1.1, G1.3, G1.4, G2.7, G5.2a, G5.2b, G5.3; EA1.0-1.3, EA4.0, EA6.0
    - @simpar1471 if you cite the growth standards (WHO and INTERGROWTH) you will be able to fulfil an additional standard, [G1.0: Statistical Software should list at least one primary reference from published academic literature.](https://stats-devguide.ropensci.org/standards.html#documentation)
- Excellent design choices for the internal aspects of package structure and design. I especially appreciate the refactoring for generalisability of the conversions (using the conversions instead of the growth standards as the function names).

@BroVic let me know your thoughts on the points above!

Thank you,
Chitra
