### 2024-11-19 d8a606c
chore: update files in ./data-raw for saving out ig_nbs_ext_coeffs as .dta files
