### 2024-11-19 #626 @simpar1471
Hello again folks!

@chitrams Thanks for the thorough look-over and gold(!) badge recommendation. In terms of references to primary literature, we have listed relevant articles for the data objects that have previously published z-score/centile data, e.g. for `gigs::who_gs` ([pkgdown page](https://lshtm-gigs.github.io/gigs/reference/who_gs.html)).

I assume you'd want references in `value2zscore()` and friends, so I was thinking adding something which users can use to click through to the set of references they want:

```r
#' ...
#' @references
#' * For the INTERGROWTH-21<sup>st</sup> Fetal standards see the
#'   [`gigs::ig_fet`][ig_fet] documentation
#' * For the INTERGROWTH-21<sup>st</sup> Newborn Size standards see the
#'   [`gigs::ig_nbs`][ig_nbs] documentation
#' * For the INTERGROWTH-21<sup>st</sup> Postnatal Growth standards see the
#'   [`gigs::ig_png`][ig_png] documentation
#' * For the WHO Growth standards see the [`gigs::who_gs`][who_gs] documentation
#' ...
```

I think this would be less unwieldy than listing >10 references on the `value2zscore()` and `zscore2value()` pages, but I'm happy to prioritise accessibility of the information if you think that's best. Also, do you want the references added elsewhere?

Also @chitrams and @BroVic, I've added another feature in the last week. It's an extended version of the INTERGROWTH-21<sup>st</sup> Newborn Size standards, which can take gestational age (GA) values in a wider range than the normal Newborn Size Standards. Eric, the co-PI for this project, has done these extrapolations before for other research groups but has needed to re-compute them at times where the underlying code/data has gone missing. Our hope is to situate them in gigs as an easy place to direct folks to. Users can apply the extended standards by specifying `family = "ig_nbs_ext"` in any of the conversion functions, and using an `acronym` out of `c("wfga", "lfga", "hcfga")`. This all in extra `./R` code, unit tests, and an article describing the extrapolation process, all added in https://github.com/lshtm-gigs/gigs/commit/69caa7f4deb244891a519cc76753a1d35794a42d and passing CI checks.
