### 2024-11-19 #626 @chitrams
Oh, awesome! I was only thinking at the bottom of the README under a "References" heading, but that (the former) sounds good too :) Whichever is easiest for you should be fine @simpar1471.

Ooh thank you for the update. I'll have a look at that new feature.

Small aside: I realise I've actually met Eric briefly last year, at a talk he gave at the RSS Conference in Harrogate. I vaguely remember him mentioning working on these child growth standards :) So it's super cool that it's come back full circle and I'm reviewing gigs!
