### 2024-11-19 5615503
docs: Added reference list to README + better reference links for conversion functions
