+++ README.Rmd
@@ -475,7 +475,7 @@ Health Centre, London School of Hygiene & Tropical Medicine
 
 ### Citation
 
-Parker S, Vesel L, Ohuma EO (2024). *gigs: Assess Fetal, Newborn, and
+Parker SR, Vesel L, Ohuma EO (2024). *gigs: Assess Fetal, Newborn, and
 Child Growth*. <https://github.com/lshtm-gigs/gigs/>.
 
 ## Code of Conduct
@@ -483,3 +483,103 @@ Child Growth*. <https://github.com/lshtm-gigs/gigs/>.
 Please note that the **gigs** package is released with a [Contributor Code of
 Conduct](https://ropensci.org/code-of-conduct/). By contributing to this project
 you agree to abide by its terms.
+
+## References
+
+```{r doi_fn, echo = FALSE}
+rmd_doi <- function(doi) {
+  paste0("doi: [", doi, "](https://dx.doi.org/", doi, ")") |>
+    knitr::asis_output()
+}
+```
+
+* INTERGROWTH-21<sup>st</sup> Newborn Size standards (including very preterm)
+  * Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG,
+et al. **International standards for newborn weight, length, and head
+circumference by gestational age and sex: the Newborn Cross-Sectional
+Study of the INTERGROWTH-21st Project.** *Lancet* 2014,
+**384(9946):857-68.** doi:
+[10.1016/S0140-6736(14)60932-6](https://dx.doi.org/10.1016/S0140-6736(14)60932-6)
+  * Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al.
+**INTERGROWTH-21st very preterm size at birth reference charts.**
+*Lancet* 2016, **387(10021):844-45.** doi:
+[10.1016/S0140-6736(16)00384-6](https://dx.doi.org/10.1016/S0140-6736(16)00384-6)
+  * Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F,
+et al. **Body composition at birth and its relationship with neonatal
+anthropometric ratios: the newborn body composition study of the
+INTERGROWTH-21st project.** *Pediatric Research* 2017, **82:305-316.**
+doi: [10.1038/pr.2017.52](https://dx.doi.org/10.1038/pr.2017.52)
+
+
+* INTERGROWTH-21<sup>st</sup> Postnatal Growth of Preterm Infants standards
+  * Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al.
+**Postnatal growth standards for preterm infants: the Preterm Postnatal
+Follow-up Study of the INTERGROWTH-21st Project.** *Lancet Glob Health*
+2015, *3(11):e681-e691.* doi:
+[10.1016/S2214-109X(15)00163-1](https://dx.doi.org/10.1016/S2214-109X(15)00163-1)
+
+
+* INTERGROWTH-21<sup>st</sup> Fetal standards
+  * Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L,
+Lambert A et al. **International standards for fetal growth based on
+serial ultrasound measurements: the Fetal Growth Longitudinal Study of
+the INTERGROWTH-21st Project.** *Lancet* 2014, **384(9946):869-79.**
+doi:
+[10.1016/S0140-6736(14)61490-2](https://dx.doi.org/10.1016/S0140-6736(14)61490-2)
+  * Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et
+al. **International Estimated Fetal Weight Standards of the
[…]
