### 2024-11-19 #626 @simpar1471
@chitrams I've added both sets of extra references in https://github.com/lshtm-gigs/gigs/commit/5615503faea9fe59243790214d504b7613a41cb7. As for you having met Eric IRL, looks like it's a small world after all!
