### 2024-11-21 6b642f4
docs: Added GIGS for SAS to README and a mention in the benchmarking articl.
