### 2024-11-21 8c39f93
chore: Updated r2sas.R, which produces data used in the SAS package. Added code to produce extended IG-21st NBS coeff table and code to produce tables to compare behaviour of gigs::classify_growth (R) with gigs_classify_growth (SAS)
