+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   articles/extending-ig-nbs: extending-ig-nbs.html
   gigs: gigs.html
   articles/ig-nbs-body-composition: ig-nbs-body-composition.html
-last_built: 2024-11-19T13:20Z
+last_built: 2024-11-21T12:39Z
 urls:
   reference: https://lshtm-gigs.github.io/gigs/reference
   article: https://lshtm-gigs.github.io/gigs/articles
