### 2024-11-26 f60de94
docs: Updated benchmarking article to include data from SAS implementation of GIGS
