+++ vignettes/articles/benchmarking.Rmd
@@ -70,9 +70,11 @@ z-scores in different growth standards.
 We performed these benchmarks on R version
 `r session_info[[1]][["major"]]`.`r session_info[[1]][["minor"]]`, using a
 Windows 10 system with a Ryzen 7 3700X processor and 16GB of DDR4 RAM. The Stata
-benchmarks were run in Stata 18.0 (revision 15 May 2023) on the same system,
+benchmarks were run in Stata 18.0 (revision 16 Oct 2024) on the same system,
 using the [benchmark](https://github.com/mcaceresb/stata-benchmark) package for
-Stata.
+Stata. The SAS benchmarks were also performed on this system, using SAS 9.4
+(9.04.01M6P111518), in a custom benchmarking process detailed in a code block
+at the bottom of this article.
 
 ```{r yes_no_half, echo = FALSE}
 yes <- if (knitr::is_html_output()) {
@@ -145,7 +147,7 @@ In this vignette, we perform benchmark performance in each a growth standard
 from each set of growth standards implemented in **gigs**, excluding the
 extended `r intergrowth21st` Newborn Size standards. These are the:
 
-* WHO child growth standards (0-5 years)
+* WHO Child Growth Standards (0-5 years)
 * `r intergrowth21st` Newborn Size standards
 * `r intergrowth21st` Postnatal Growth standards
 * `r intergrowth21st` Fetal standards
@@ -201,19 +203,21 @@ max_median <- function(bp, name) {
 
 pkg_str <- function(pkgname) paste("R:", pkgname, packageVersion(pkg = pkgname))
 
-ver_stata_gigs <- "0.4.0"
-ver_stata_zanthro <- "1.0.2"
-plot_bp <- function(bp_r, bp_stata = NULL) {
+ver_zanthro_stata <- "1.0.2"
+ver_gigs_sas <- "0.1.8"
+ver_gigs_stata <- "1.1.1"
+plot_bp <- function(bp) {
   key_pkg_colour <- c("gs" = "#925E9FFF",
                       "gigs" = "#00468BFF",
                       "anthro" = "#AD002AFF",
-                      "AGD" = "cornflowerblue",
+                      "AGD" = "#FFA3FF",
                       "intergrowth" = "#FDAF91FF",
                       "childsds" = "#663300FF",
                       "sitar" = "#005500FF",
                       "zscorer" = "#009900",
-                      "stata_gigs" = "skyblue",
-                      "stata_zanthro" = "darkolivegreen4")
+                      "gigs_stata" = "skyblue",
+                      "zanthro_stata" = "darkolivegreen4",
+                      "gigs_sas" = "#8585e0")
   key_pkg_label <- c("gs" = pkg_str("growthstandards"),
                      "gigs" = pkg_str("gigs"),
                      "anthro" = pkg_str("anthro"),
@@ -222,14 +226,15 @@ plot_bp <- function(bp_r, bp_stata = NULL) {
                      "childsds" = pkg_str("childsds"),
                      "sitar" = pkg_str("sitar"),
                      "zscorer" = pkg_str("zscorer"),
-                     "stata_gigs" = paste("Stata: gigs", ver_stata_gigs),
-                     "stata_zanthro" = paste("Stata: zanthro",
-                                             ver_stata_zanthro))
-  bp <- bp_r  |>
-    dplyr::select(desc, input_len, median)
-  if (!is.null(bp_stata)) {
[…]
