+++ README.Rmd
@@ -436,7 +436,7 @@ inputs. The code used to generate these timings can be seen online in the
 [article](https://lshtm-gigs.github.io/gigs/articles/benchmarking.html).
 
 | Software        | Platform | WHO (0-5 years) (ms)                      | `r ig21st` NBS (ms)                    | `r ig21st` PNG (ms)                    | `r ig21st` Fetal (ms)                   |
-|:----------------|----------|-------------------------------------------|----------------------------------------|----------------------------------------|-----------------------------------------|
+|-----------------|----------|-------------------------------------------|----------------------------------------|----------------------------------------|-----------------------------------------|
 | `r gigs_r`      | R        | `r max_len_bench[[1]][["gigs"]]`          | `r max_len_bench[[2]][["gigs"]]`       | `r max_len_bench[[3]][["gigs"]]`       | `r max_len_bench[[4]][["gigs"]]`        |
 | `r anthro`      | R        | `r max_len_bench[[1]][["anthro"]]`        | `r no`                                 | `r no`                                 | `r no`                                  |
 | `r AGD`         | R        | `r max_len_bench[[1]][["AGD"]]`           | `r no`                                 | `r no`                                 | `r no`                                  |
@@ -445,8 +445,9 @@ inputs. The code used to generate these timings can be seen online in the
 | `r intergrowth` | R        | `r no`                                    | `r no`                                 | `r no`                                 | `r max_len_bench[[4]][["intergrowth"]]` |
 | `r sitar`       | R        | `r max_len_bench[[1]][["sitar"]]`         | `r no`                                 | `r no`                                 | `r no`                                  |
 | `r zscorer`     | R        | NA                                        | `r no`                                 | `r no`                                 | `r no`                                  |
-| `r gigs_stata`  | Stata    | `r max_len_bench[[1]][["stata_gigs"]]`    | `r max_len_bench[[2]][["stata_gigs"]]` | `r max_len_bench[[3]][["stata_gigs"]]` | `r max_len_bench[[4]][["stata_gigs"]]`  |
-| `r zanthro`     | Stata    | `r max_len_bench[[1]][["stata_zanthro"]]` | `r no`                                 | `r no`                                 | `r no`                                  |
+| `r gigs_stata`  | Stata    | `r max_len_bench[[1]][["gigs_stata"]]`    | `r max_len_bench[[2]][["gigs_stata"]]` | `r max_len_bench[[3]][["gigs_stata"]]` | `r max_len_bench[[4]][["gigs_stata"]]`  |
+| `r zanthro`     | Stata    | `r max_len_bench[[1]][["zanthro_stata"]]` | `r no`                                 | `r no`                                 | `r no`                                  |
+| `r gigs_sas`    | SAS      | `r max_len_bench[[1]][["gigs_sas"]]`      | `r max_len_bench[[2]][["gigs_sas"]]`   | `r max_len_bench[[3]][["gigs_sas"]]`   | `r max_len_bench[[4]][["gigs_sas"]]`    |
 
 Note: `zscorer` is NA because we couldn't time it for 100,000 inputs (it takes
 too long).
