### 2024-11-29 2ba7ac1
chore: amend DESCRIPTION to enforce vctrs v0.4.0 or above
