+++ CITATION.cff
@@ -136,6 +136,7 @@ references:
     email: davis@posit.co
   year: '2024'
   doi: 10.32614/CRAN.package.vctrs
+  version: '>= 0.4.0'
 - type: software
   title: cli
   abstract: 'cli: Helpers for Developing Command Line Interfaces'
