### 2024-11-29 f860175
docs: get consistent with how we describe the Extended IG-21st Newborn Size standards
