+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   articles/extending-ig-nbs: extending-ig-nbs.html
   gigs: gigs.html
   articles/ig-nbs-body-composition: ig-nbs-body-composition.html
-last_built: 2024-11-26T16:48Z
+last_built: 2024-11-29T17:06Z
 urls:
   reference: https://lshtm-gigs.github.io/gigs/reference
   article: https://lshtm-gigs.github.io/gigs/articles
