### 2024-11-29 978b178
docs: build README.md to include **pak** installation instructions
