### 2024-12-10 #626 @simpar1471
Hiya @rkillick! Given Victor's agreement would it be possible to progress to another pkgcheck run here, or perform whatever the other next steps are?
