### 2024-12-10 #626 @ropensci-review-bot
Approved! Thanks @simpar1471 for submitting and @chitrams, @brovic for your reviews! :grin:
