### 2024-12-16 #626 @rkillick
@simpar1471  hmmmmm, it should be instant, it was for me.  I checked without being logged in and I can see my check marks but no more.  I'm not sure why it isn't registering any from you, if you want you can let me know and I'll tick them off.
