### 2024-12-18 #626 @simpar1471
Thanks @rkillick. I've done all of the items on the checklist, including setting up the old GH pages website redirect (try here: https://lshtm-gigs.github.io/gigs/).

Do you know much about the process of submitting to [JOSS](https://joss.theoj.org/) afer rOpenSci review?
