### 2024-12-18 #626 @rkillick
Great, I've checked everything off.  I haven't submitted to JOSS myself but would assume that you would link to this thread in the covering letter.  @mpadge do you know any specifics around this?
