### 2024-12-23 #626 @mpadge
@simpar1471 @rkillick JOSS submission should proceed fairlly automatically from this point on. The rOpenSci review is generally considered sufficient for JOSS, and only main decision made on that side is scope and fit. A recent example of the process in action is https://github.com/openjournals/joss-reviews/issues/6163. In short: @simpar1471 please go ahead and submit directly to JOSS
